Hacker News latest ask stories

Retrieves latest ask stories using Hacker News API

get_latest_ask_stories(max_items = NULL)

Arguments

max_items
Maximum number of items to retrieve. If max_items = NULL, returns all available

Value

list of latest ask HN stories ids

Details

Parallel api requests can be enabled by running future::plan(future::multiprocess)

Examples

# \donttest{
# get the lastest ask story on Hacker News
latest_ask_story <- get_latest_ask_stories(max_items = 1)
latest_ask_story

#> [[1]]
#> List of 9
#>  $ by         : chr "yolo42"
#>  $ descendants: int 154
#>  $ id         : int 21786111
#>  $ kids       : int [1:43] 21787359 21789293 21786329 21788992 21789209 21788907 21789189 21787465 21787014 21789157 ...
#>  $ score      : int 130
#>  $ text       : chr "Hello everyone,<p>I created product on the side as a side project in my own time and it has become a huge hit i"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 23:16:17"
#>  $ title      : chr "Ask HN: My company wants my side project. What can I do?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get 10 latest ask stories on Hacker News
latest_10_ask_stories <- get_latest_ask_stories(max_items = 10)
latest_10_ask_stories

#> [[1]]
#> List of 9
#>  $ by         : chr "yolo42"
#>  $ descendants: int 154
#>  $ id         : int 21786111
#>  $ kids       : int [1:43] 21787359 21789293 21786329 21788992 21789209 21788907 21789189 21787465 21787014 21789157 ...
#>  $ score      : int 130
#>  $ text       : chr "Hello everyone,<p>I created product on the side as a side project in my own time and it has become a huge hit i"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 23:16:17"
#>  $ title      : chr "Ask HN: My company wants my side project. What can I do?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 9
#>  $ by         : chr "gkop"
#>  $ descendants: int 192
#>  $ id         : int 21783303
#>  $ kids       : int [1:28] 21784601 21784548 21784269 21784310 21784084 21784197 21784059 21784399 21784699 21785272 ...
#>  $ score      : int 321
#>  $ text       : chr "HN threads from 2016-2018 complain that JWT gives us too much rope with which to hang ourselves, and some advoc"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 18:05:41"
#>  $ title      : chr "Ask HN: What's the current sentiment on JWT for stateless auth tokens?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 9
#>  $ by         : chr "marz0"
#>  $ descendants: int 88
#>  $ id         : int 21783323
#>  $ kids       : int [1:43] 21783848 21783820 21784660 21784165 21784689 21783858 21784076 21784656 21783532 21783775 ...
#>  $ score      : int 149
#>  $ text       : chr "What are some of the best ways to volunteer as someone working in tech? Many of us have valuable skills that we"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 18:08:21"
#>  $ title      : chr "Ask HN: Best ways to volunteer?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 9
#>  $ by         : chr "throwawayOpts"
#>  $ descendants: int 5
#>  $ id         : int 21785655
#>  $ kids       : int [1:4] 21789197 21786675 21786988 21787471
#>  $ score      : int 17
#>  $ text       : chr "Hey HN,<p>I was a software engineer at a US startup now considered a unicorn and recently in the news.  It&#x27"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 22:24:25"
#>  $ title      : chr "Ask HN: Company quietly did stock split, implications for options?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 8
#>  $ by         : chr "rahuldottech"
#>  $ descendants: int 0
#>  $ id         : int 21788709
#>  $ score      : int 2
#>  $ text       : chr "https:&#x2F;&#x2F;www.facebook.com&#x2F;facebookcorewwwi<p>Apparently will be down for the next few weeks becau"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-14 11:01:46"
#>  $ title      : chr "Tell HN: Facebook over Tor Is Down"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 9
#>  $ by         : chr "codesternews"
#>  $ descendants: int 6
#>  $ id         : int 21786687
#>  $ kids       : int [1:5] 21786809 21787248 21786779 21788898 21788067
#>  $ score      : int 7
#>  $ text       : chr "Most people have the Side Projects and most often employers force the employee to take ownership if it is succe"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-14 00:43:29"
#>  $ title      : chr "Ask HN: How to save your side projects from your Employers?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 9
#>  $ by         : chr "stw_fxpl"
#>  $ descendants: int 1
#>  $ id         : int 21779397
#>  $ kids       : int 21781351
#>  $ score      : int 46
#>  $ text       : chr "It’s our pleasure to announce the initial public release of<p>Structure and Interpretation of Computer Programs"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 06:50:26"
#>  $ title      : chr "SICP JavaScript Going Public"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "sapporosnow"
#>  $ descendants: int 4
#>  $ id         : int 21781574
#>  $ kids       : int [1:3] 21785818 21785512 21787644
#>  $ score      : int 4
#>  $ text       : chr "I have an app with a target market.  I can launch in my local city, which isn&#x27;t really known for startups,"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 15:06:10"
#>  $ title      : chr "Ask HN: Should I launch in my local city or abroad?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 9
#>  $ by         : chr "chrisshroba"
#>  $ descendants: int 45
#>  $ id         : int 21765528
#>  $ kids       : int [1:29] 21773575 21765787 21770777 21765634 21768609 21766872 21766866 21772632 21767165 21765591 ...
#>  $ score      : int 128
#>  $ text       : chr "I&#x27;m currently reading Seven Databases in Seven Weeks [1] and finding it to be a really fun and interesting"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-11 19:51:53"
#>  $ title      : chr "Ask HN: Fun Tech Book Recommendations?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 9
#>  $ by         : chr "yori"
#>  $ descendants: int 111
#>  $ id         : int 21774286
#>  $ kids       : int [1:49] 21774965 21775201 21774806 21774888 21774871 21775156 21788711 21782517 21780309 21775464 ...
#>  $ score      : int 44
#>  $ text       : chr "Do you still write code in BASIC? If yes, which BASIC compiler or interpreter do you use?"
#>  $ time       : POSIXct[1:1], format: "2019-12-12 18:52:47"
#>  $ title      : chr "Ask HN: Do you still code in BASIC?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get all latest ask stories on Hacker News
latest_ask_stories <- get_latest_ask_stories()
latest_ask_stories

#> [[1]]
#> List of 9
#>  $ by         : chr "yolo42"
#>  $ descendants: int 154
#>  $ id         : int 21786111
#>  $ kids       : int [1:43] 21787359 21789293 21786329 21788992 21789209 21788907 21789189 21787465 21787014 21789157 ...
#>  $ score      : int 130
#>  $ text       : chr "Hello everyone,<p>I created product on the side as a side project in my own time and it has become a huge hit i"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 23:16:17"
#>  $ title      : chr "Ask HN: My company wants my side project. What can I do?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 9
#>  $ by         : chr "gkop"
#>  $ descendants: int 192
#>  $ id         : int 21783303
#>  $ kids       : int [1:28] 21784601 21784548 21784269 21784310 21784084 21784197 21784059 21784399 21784699 21785272 ...
#>  $ score      : int 321
#>  $ text       : chr "HN threads from 2016-2018 complain that JWT gives us too much rope with which to hang ourselves, and some advoc"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 18:05:41"
#>  $ title      : chr "Ask HN: What's the current sentiment on JWT for stateless auth tokens?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 9
#>  $ by         : chr "marz0"
#>  $ descendants: int 88
#>  $ id         : int 21783323
#>  $ kids       : int [1:43] 21783848 21783820 21784660 21784165 21784689 21783858 21784076 21784656 21783532 21783775 ...
#>  $ score      : int 149
#>  $ text       : chr "What are some of the best ways to volunteer as someone working in tech? Many of us have valuable skills that we"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 18:08:21"
#>  $ title      : chr "Ask HN: Best ways to volunteer?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 9
#>  $ by         : chr "throwawayOpts"
#>  $ descendants: int 5
#>  $ id         : int 21785655
#>  $ kids       : int [1:4] 21789197 21786675 21786988 21787471
#>  $ score      : int 17
#>  $ text       : chr "Hey HN,<p>I was a software engineer at a US startup now considered a unicorn and recently in the news.  It&#x27"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 22:24:25"
#>  $ title      : chr "Ask HN: Company quietly did stock split, implications for options?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 8
#>  $ by         : chr "rahuldottech"
#>  $ descendants: int 0
#>  $ id         : int 21788709
#>  $ score      : int 2
#>  $ text       : chr "https:&#x2F;&#x2F;www.facebook.com&#x2F;facebookcorewwwi<p>Apparently will be down for the next few weeks becau"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-14 11:01:46"
#>  $ title      : chr "Tell HN: Facebook over Tor Is Down"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 9
#>  $ by         : chr "codesternews"
#>  $ descendants: int 6
#>  $ id         : int 21786687
#>  $ kids       : int [1:5] 21786809 21787248 21786779 21788898 21788067
#>  $ score      : int 7
#>  $ text       : chr "Most people have the Side Projects and most often employers force the employee to take ownership if it is succe"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-14 00:43:29"
#>  $ title      : chr "Ask HN: How to save your side projects from your Employers?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 9
#>  $ by         : chr "stw_fxpl"
#>  $ descendants: int 1
#>  $ id         : int 21779397
#>  $ kids       : int 21781351
#>  $ score      : int 46
#>  $ text       : chr "It’s our pleasure to announce the initial public release of<p>Structure and Interpretation of Computer Programs"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 06:50:26"
#>  $ title      : chr "SICP JavaScript Going Public"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "sapporosnow"
#>  $ descendants: int 4
#>  $ id         : int 21781574
#>  $ kids       : int [1:3] 21785818 21785512 21787644
#>  $ score      : int 4
#>  $ text       : chr "I have an app with a target market.  I can launch in my local city, which isn&#x27;t really known for startups,"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 15:06:10"
#>  $ title      : chr "Ask HN: Should I launch in my local city or abroad?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 9
#>  $ by         : chr "chrisshroba"
#>  $ descendants: int 45
#>  $ id         : int 21765528
#>  $ kids       : int [1:29] 21773575 21765787 21770777 21765634 21768609 21766872 21766866 21772632 21767165 21765591 ...
#>  $ score      : int 128
#>  $ text       : chr "I&#x27;m currently reading Seven Databases in Seven Weeks [1] and finding it to be a really fun and interesting"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-11 19:51:53"
#>  $ title      : chr "Ask HN: Fun Tech Book Recommendations?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 9
#>  $ by         : chr "yori"
#>  $ descendants: int 111
#>  $ id         : int 21774286
#>  $ kids       : int [1:49] 21774965 21775201 21774806 21774888 21774871 21775156 21788711 21782517 21780309 21775464 ...
#>  $ score      : int 44
#>  $ text       : chr "Do you still write code in BASIC? If yes, which BASIC compiler or interpreter do you use?"
#>  $ time       : POSIXct[1:1], format: "2019-12-12 18:52:47"
#>  $ title      : chr "Ask HN: Do you still code in BASIC?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[11]]
#> List of 9
#>  $ by         : chr "obilgic"
#>  $ descendants: int 12
#>  $ id         : int 21777109
#>  $ kids       : int [1:10] 21777372 21783861 21782523 21779533 21787676 21780591 21778762 21778493 21777600 21784282
#>  $ score      : int 25
#>  $ text       : chr "Hi HN, It hasn&#x27;t happened for a while but, It was a common theme to have a thread to recommend books for h"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 23:11:00"
#>  $ title      : chr "Ask HN: 2019 Winter Reading List?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[12]]
#> List of 9
#>  $ by         : chr "TurkishPoptart"
#>  $ descendants: int 5
#>  $ id         : int 21784715
#>  $ kids       : int [1:3] 21789072 21784943 21784879
#>  $ score      : int 8
#>  $ text       : chr "I&#x27;m reading The Click Moment, where the author writes, &quot;Ten years ago, if you wanted to guarantee you"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 20:46:15"
#>  $ title      : chr "Ask HN: Is there a bubble in software work? What's the 5/10 year forecast?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[13]]
#> List of 9
#>  $ by         : chr "burnedfaceless"
#>  $ descendants: int 4
#>  $ id         : int 21786702
#>  $ kids       : int [1:4] 21788348 21786983 21786763 21786783
#>  $ score      : int 2
#>  $ text       : chr "I&#x27;m a 32 year old computer science student. I&#x27;ve been on Adderall since I was 14.<p>I&#x27;m on 40mg "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-14 00:45:50"
#>  $ title      : chr "Adderall Tolerance"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[14]]
#> List of 9
#>  $ by         : chr "snejad123"
#>  $ descendants: int 7
#>  $ id         : int 21775705
#>  $ kids       : int [1:7] 21777999 21777617 21776435 21776558 21776481 21777510 21777215
#>  $ score      : int 13
#>  $ text       : chr "Hello, so I&#x27;m facing a conundrum right now.<p>I have the option of either picking something fun to build w"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 20:58:54"
#>  $ title      : chr "Should I build something fun or build something that solves a problem?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[15]]
#> List of 9
#>  $ by         : chr "jonnismash"
#>  $ descendants: int 21
#>  $ id         : int 21772202
#>  $ kids       : int [1:6] 21772244 21775677 21776852 21775333 21777447 21787732
#>  $ score      : int 28
#>  $ text       : chr "Please HN, please. This is one, if not, the first site that I visit after making my coffee. The glaring white b"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 15:46:39"
#>  $ title      : chr "Suggest to HN: D A R K M O D E"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[16]]
#> List of 9
#>  $ by         : chr "federiconitidi"
#>  $ descendants: int 7
#>  $ id         : int 21776054
#>  $ kids       : int [1:3] 21779838 21778752 21778184
#>  $ score      : int 17
#>  $ text       : chr "As per the title, Today I just introduced a couple of paid options for my Saas and I just recorded the first pa"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 21:27:21"
#>  $ title      : chr "Just introduced paid plans in my SaaS and got the first paying customer"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[17]]
#> List of 9
#>  $ by         : chr "founderling"
#>  $ descendants: int 1
#>  $ id         : int 21773510
#>  $ kids       : int 21777239
#>  $ score      : int 16
#>  $ text       : chr "For example, are Samsung and Huawei using the original Android technology to store the fingerprint? Or are they"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 17:50:47"
#>  $ title      : chr "Ask HN: Is stock Android more secure then a skin from Samsung, Huawei etc.?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[18]]
#> List of 9
#>  $ by         : chr "Anon84"
#>  $ descendants: int 22
#>  $ id         : int 21773158
#>  $ kids       : int [1:13] 21781284 21783175 21775646 21774565 21775205 21786467 21775641 21775056 21777073 21776111 ...
#>  $ score      : int 23
#>  $ text       : chr "Hi everyone,<p>With the holiday break coming up, many of us take advantage of the time to catch up on our profe"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 17:16:59"
#>  $ title      : chr "Ask HN: What are you looking forward to learn in 2020?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[19]]
#> List of 9
#>  $ by         : chr "abiro"
#>  $ descendants: int 1
#>  $ id         : int 21783817
#>  $ kids       : int 21783844
#>  $ score      : int 5
#>  $ text       : chr "A passage in the preface to SICP [1] on the real meaning of computer science has captured my imagination ever s"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 19:04:00"
#>  $ title      : chr "Ask HN: Which of our ideas about computation could survive a thousand years?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[20]]
#> List of 9
#>  $ by         : chr "switzerland"
#>  $ descendants: int 1
#>  $ id         : int 21775093
#>  $ kids       : int 21782525
#>  $ score      : int 7
#>  $ text       : chr "Dear hackers and tech entrepreneurs,<p>I want to start my first software company. Unfortunately I don&#x27;t kn"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 20:05:14"
#>  $ title      : chr "Accounting for Tech Entrepreneurs"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[21]]
#> List of 9
#>  $ by         : chr "googlycooly"
#>  $ descendants: int 23
#>  $ id         : int 21765296
#>  $ kids       : int [1:13] 21772785 21765771 21768760 21768115 21768718 21767741 21767114 21769253 21776350 21767679 ...
#>  $ score      : int 25
#>  $ text       : chr "I&#x27;m shy and I feel weird talking to people, looking into their eyes. Talking to a group of people will mak"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-11 19:29:03"
#>  $ title      : chr "Ask HN: Any tips for a shy person who wants to do sales?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[22]]
#> List of 9
#>  $ by         : chr "yori"
#>  $ descendants: int 6
#>  $ id         : int 21771994
#>  $ kids       : int [1:7] 21777309 21781613 21772726 21772034 21776243 21772035 21779068
#>  $ score      : int 6
#>  $ text       : chr "I am looking for some new blogs to follow. Let me explain the three criteria in the title of this self post.<p>"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 15:16:37"
#>  $ title      : chr "Ask HN: What are some simple, clean, off-beat, and self-hosted blogs you follow?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[23]]
#> List of 8
#>  $ by         : chr "geolgau"
#>  $ descendants: int 0
#>  $ id         : int 21779861
#>  $ score      : int 17
#>  $ text       : chr "According to [1] and other sources it looks like Igor Sysoev is accused of stealing the software by Russian Ram"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 08:49:20"
#>  $ title      : chr "Nginx Developer Arrested in Russia"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[24]]
#> List of 9
#>  $ by         : chr "zer0sand0nes"
#>  $ descendants: int 2
#>  $ id         : int 21776338
#>  $ kids       : int [1:2] 21776805 21776775
#>  $ score      : int 3
#>  $ text       : chr "Business Acquisitions, New Market Entries, New Waves of Technological Innovations, etc."
#>  $ time       : POSIXct[1:1], format: "2019-12-12 21:54:28"
#>  $ title      : chr "Ask HN: What Are Your Go-To Websites for Industry News and Business Moves?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[25]]
#> List of 9
#>  $ by         : chr "macco"
#>  $ descendants: int 18
#>  $ id         : int 21764779
#>  $ kids       : int [1:10] 21789337 21786820 21786931 21765094 21765193 21765945 21764976 21764868 21765522 21766187
#>  $ score      : int 24
#>  $ text       : chr "I am imaging many startups here use GSuite. I was wondering how you do knowledge management.<p>We tried Keep, b"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-11 18:38:37"
#>  $ title      : chr "Ask HN: GSuite users, how do you do Knowledge Management?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[26]]
#> List of 9
#>  $ by         : chr "hvenev"
#>  $ descendants: int 1
#>  $ id         : int 21770495
#>  $ kids       : int 21771744
#>  $ score      : int 12
#>  $ text       : chr "A few days ago I started getting 403 errors when accessing some comment pages on Hacker News. Yesterday I sent "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 10:19:33"
#>  $ title      : chr "Tell HN: AdNauseam got my IP address blocked from HN for being a bot"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[27]]
#> List of 7
#>  $ by         : chr "adawg_4"
#>  $ descendants: int 0
#>  $ id         : int 21786746
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-12-14 00:54:07"
#>  $ title      : chr "Ask HN: Best way to contribute to stack-overflow?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[28]]
#> List of 9
#>  $ by         : chr "ahmaman"
#>  $ descendants: int 19
#>  $ id         : int 21769606
#>  $ kids       : int [1:12] 21769872 21771299 21770453 21770644 21769692 21775391 21779244 21778418 21769658 21769817 ...
#>  $ score      : int 12
#>  $ text       : chr "Wonder if they really work and what are the risks with sharing my DNA with a private company.<p>Any recommendat"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 07:00:02"
#>  $ title      : chr "Ask HN: Would you trust an ancestry DNA test?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[29]]
#> List of 8
#>  $ by         : chr "NaOH"
#>  $ descendants: int 1
#>  $ id         : int 21785960
#>  $ kids       : int [1:2] 21786058 21785992
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-12-13 22:59:33"
#>  $ title      : chr "Tony Brooker Has Died"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[30]]
#> List of 9
#>  $ by         : chr "neural_thing"
#>  $ descendants: int 4
#>  $ id         : int 21769448
#>  $ kids       : int [1:3] 21770103 21786799 21769719
#>  $ score      : int 7
#>  $ text       : chr "A troubling thing happened to my wife today. She got an ad based on a very rare word that only appeared on her "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 06:15:54"
#>  $ title      : chr "Ask HN: Google having access to Signal messages?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[31]]
#> List of 9
#>  $ by         : chr "s_tech"
#>  $ descendants: int 4
#>  $ id         : int 21775811
#>  $ kids       : int [1:4] 21783818 21779817 21775963 21787658
#>  $ score      : int 4
#>  $ text       : chr "Hey HN,<p>I&#x27;m asking on behalf of a friend who is seeking to hire someone for their project (i&#x27;m a de"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 21:08:06"
#>  $ title      : chr "Ask HN: Where to hire freelance web/app developers without a annoying middleman?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[32]]
#> List of 9
#>  $ by         : chr "davidwparker"
#>  $ descendants: int 3
#>  $ id         : int 21773353
#>  $ kids       : int [1:3] 21774122 21779189 21780930
#>  $ score      : int 4
#>  $ text       : chr "I&#x27;m getting a new computer for the first time in ~6 years. I&#x27;ve been on MBP&#x2F;OSX for the last 8 y"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 17:36:18"
#>  $ title      : chr "Ask HN: What do you put on a new computer?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[33]]
#> List of 9
#>  $ by         : chr "aaronharnly"
#>  $ descendants: int 8
#>  $ id         : int 21768007
#>  $ kids       : int [1:5] 21769709 21768035 21782942 21777556 21777246
#>  $ score      : int 12
#>  $ text       : chr "I&#x27;ve now three or four times run into the situation where I want to securely and privately find the inters"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-12 00:47:16"
#>  $ title      : chr "Ask HN: Securely finding the intersection of two lists?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[34]]
#> List of 8
#>  $ by         : chr "rpiguy"
#>  $ descendants: int 0
#>  $ id         : int 21765980
#>  $ score      : int 2
#>  $ text       : chr "I would like to see more examples of successfully funded business plans.<p>I understand that financials and IP "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-11 20:41:57"
#>  $ title      : chr "Is there a repository of successfully funded business plans"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[35]]
#> List of 8
#>  $ by         : chr "it"
#>  $ descendants: int 3
#>  $ id         : int 21778029
#>  $ kids       : int [1:3] 21782981 21778647 21781978
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-12-13 01:28:02"
#>  $ title      : chr "Ask HN: Do you recommend splitting your app into an API and an FE that uses it?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[36]]
#> List of 8
#>  $ by         : chr "gtirloni"
#>  $ descendants: int 2
#>  $ id         : int 21777977
#>  $ kids       : int [1:2] 21778026 21782097
#>  $ score      : int 4
#>  $ time       : POSIXct[1:1], format: "2019-12-13 01:20:14"
#>  $ title      : chr "Ask HN: How do you intercept server-side HTTPS traffic for debugging?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[37]]
#> List of 9
#>  $ by         : chr "egze"
#>  $ descendants: int 61
#>  $ id         : int 21780294
#>  $ kids       : int [1:39] 21780464 21780323 21780390 21780479 21780356 21780380 21780322 21780415 21780446 21780340 ...
#>  $ score      : int 35
#>  $ text       : chr "Hi,<p>We think that a lot of developers want to contribute to open source, but don&#x27;t do it for various rea"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 10:34:35"
#>  $ title      : chr "Ask HN: Do you contribute to open source? Why? Why not?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[38]]
#> List of 8
#>  $ by         : chr "stirrat"
#>  $ descendants: int 1
#>  $ id         : int 21777342
#>  $ kids       : int 21777415
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-12-12 23:37:36"
#>  $ title      : chr "Ask HN: Is solving a problem always the foundation for a startup?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[39]]
#> List of 8
#>  $ by         : chr "iamjk"
#>  $ descendants: int 6
#>  $ id         : int 21769241
#>  $ kids       : int [1:6] 21774771 21769303 21776935 21771846 21770153 21769599
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-12-12 05:22:02"
#>  $ title      : chr "Ask HN: What is the best way to get involved in (US) politics?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[40]]
#> List of 8
#>  $ by         : chr "Antoninus"
#>  $ descendants: int 8
#>  $ id         : int 21769507
#>  $ kids       : int [1:7] 21786984 21770850 21769690 21772673 21769613 21777267 21780079
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-12-12 06:33:22"
#>  $ title      : chr "Ask HN: What is your favorite software that you adopted in 2019?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[41]]
#> List of 8
#>  $ by         : chr "saadalem"
#>  $ descendants: int 0
#>  $ id         : int 21786262
#>  $ score      : int 1
#>  $ text       : chr "Can we crawl image graphs from online pdfs ? like https:&#x2F;&#x2F;hai.stanford.edu&#x2F;sites&#x2F;g&#x2F;fil"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 23:37:26"
#>  $ title      : chr "Ask HN: Crawling from the Web Pdfs"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[42]]
#> List of 9
#>  $ by         : chr "TheAsprngHacker"
#>  $ descendants: int 6
#>  $ id         : int 21786100
#>  $ kids       : int [1:6] 21786266 21786369 21786267 21786177 21786399 21786201
#>  $ score      : int 1
#>  $ text       : chr "I just got rejected from Cornell.<p>Throughout my teenage years, I have been in love with PL theory&#x2F;type t"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 23:14:12"
#>  $ title      : chr "Ask HN: I got rejected from Cornell. I want to study PL theory; what to do?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[43]]
#> List of 9
#>  $ by         : chr "thisisxavier"
#>  $ descendants: int 1
#>  $ id         : int 21785570
#>  $ kids       : int 21785841
#>  $ score      : int 1
#>  $ text       : chr "YC teaches that startups should find a few customers in real life that really love your product. My ideal custo"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 22:15:34"
#>  $ title      : chr "Ask HN: How do I find a few customers for validating a niche product?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[44]]
#> List of 8
#>  $ by         : chr "it"
#>  $ descendants: int 1
#>  $ id         : int 21785274
#>  $ kids       : int [1:2] 21785600 21785601
#>  $ score      : int 1
#>  $ time       : POSIXct[1:1], format: "2019-12-13 21:45:07"
#>  $ title      : chr "Ask HN: Do you know of a good tool to order issues by importance on GitHub?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[45]]
#> List of 8
#>  $ by         : chr "ali_rehan"
#>  $ descendants: int 0
#>  $ id         : int 21784945
#>  $ score      : int 1
#>  $ text       : chr "Product &amp; support documentation takes a lot of time and it&#x27;s hard to keep it up to date. This, results"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 21:11:27"
#>  $ title      : chr "Best platform to create and update internal/product documentation?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[46]]
#> List of 7
#>  $ by         : chr "ramijames"
#>  $ descendants: int 0
#>  $ id         : int 21784553
#>  $ score      : int 1
#>  $ time       : POSIXct[1:1], format: "2019-12-13 20:29:05"
#>  $ title      : chr "HTTPS: //medium.com/getscatter/the-design-process-at-scatter-7b5f3bc7ef0d"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[47]]
#> List of 9
#>  $ by         : chr "learnfromerror"
#>  $ descendants: int 1
#>  $ id         : int 21784501
#>  $ kids       : int 21784613
#>  $ score      : int 1
#>  $ text       : chr "If the Exec Director of the ASA says &quot;don&#x27;t say &#x27;significance&#x27; must everyone follow? (even "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 20:22:54"
#>  $ title      : chr "Les stats, c'est moi: So I make the statistics laws for everyone"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[48]]
#> List of 9
#>  $ by         : chr "akjetma"
#>  $ descendants: int 1
#>  $ id         : int 21784494
#>  $ kids       : int 21786017
#>  $ score      : int 2
#>  $ text       : chr "I wanted try going back to using an old 3G &#x27;dumb phone&#x27; for awhile and in the process I learned that "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 20:22:30"
#>  $ title      : chr "Ask HN: What's going to happen to 3G infrastructure as it gets phased out?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[49]]
#> List of 9
#>  $ by         : chr "ryeguy_24"
#>  $ descendants: int 2
#>  $ id         : int 21784493
#>  $ kids       : int [1:2] 21785487 21784811
#>  $ score      : int 2
#>  $ text       : chr "As we get older, making friends takes more work.  Seems like most people are so busy with life that it&#x27;s h"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 20:22:28"
#>  $ title      : chr "Ask HN: How do you make new friends?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[50]]
#> List of 8
#>  $ by         : chr "parris"
#>  $ descendants: int 0
#>  $ id         : int 21784432
#>  $ score      : int 1
#>  $ text       : chr "I&#x27;m thinking about creating a network that lets founders test products on one another. As a network of fou"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 20:14:29"
#>  $ title      : chr "Ask HN: Thoughts on a network for founders/creators to iterate on products?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[51]]
#> List of 8
#>  $ by         : chr "aschismatic"
#>  $ descendants: int 0
#>  $ id         : int 21784427
#>  $ score      : int 1
#>  $ text       : chr "I often read on HN that paying for a custom domain is recommended for email, one of the reasons being that you "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 20:14:01"
#>  $ title      : chr "Ask HN: Registrar, email provider, and recovery email contact info best practice"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[52]]
#> List of 8
#>  $ by         : chr "throwailearner"
#>  $ descendants: int 0
#>  $ id         : int 21784394
#>  $ score      : int 1
#>  $ text       : chr "In my case my expertise is in JavaScript&#x2F;Typescript &amp; Java mainly, experienced with infrastructure, cl"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 20:09:20"
#>  $ title      : chr "Ask HN: How can a senior software developer transition into AI?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[53]]
#> List of 7
#>  $ by         : chr "estelaycacho"
#>  $ descendants: int 0
#>  $ id         : int 21784230
#>  $ score      : int 1
#>  $ time       : POSIXct[1:1], format: "2019-12-13 19:50:08"
#>  $ title      : chr "Videos YouTube en mi cuenta"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[54]]
#> List of 8
#>  $ by         : chr "it"
#>  $ descendants: int 6
#>  $ id         : int 21784028
#>  $ kids       : int [1:6] 21784168 21784080 21784182 21784346 21784500 21784047
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-12-13 19:27:30"
#>  $ title      : chr "Ask HN: Do you find daily stand-ups useful?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[55]]
#> List of 8
#>  $ by         : chr "it"
#>  $ descendants: int 3
#>  $ id         : int 21784017
#>  $ kids       : int [1:3] 21787302 21787284 21784245
#>  $ score      : int 4
#>  $ time       : POSIXct[1:1], format: "2019-12-13 19:25:53"
#>  $ title      : chr "Ask HN: How hard is it to hire Rust developers?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[56]]
#> List of 9
#>  $ by         : chr "codesternews"
#>  $ descendants: int 2
#>  $ id         : int 21783408
#>  $ kids       : int [1:2] 21788162 21783948
#>  $ score      : int 2
#>  $ text       : chr "How to design for your sideprojects. Is there any resources out there for color palette etc.<p>What is the proc"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 18:16:36"
#>  $ title      : chr "Ask HN: How to design for developers? Process and resources"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[57]]
#> List of 7
#>  $ by         : chr "ga-vu"
#>  $ descendants: int 0
#>  $ id         : int 21783316
#>  $ score      : int 1
#>  $ time       : POSIXct[1:1], format: "2019-12-13 18:07:20"
#>  $ title      : chr "#988 will be the new 911 for suicide prevention–by sometime in 2021"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[58]]
#> List of 9
#>  $ by         : chr "managementthrow"
#>  $ descendants: int 1
#>  $ id         : int 21783279
#>  $ kids       : int 21788860
#>  $ score      : int 1
#>  $ text       : chr "I&#x27;ve been in an engineering management role for 6 months at a large non-tech company in the financial sect"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 18:01:54"
#>  $ title      : chr "Ask HN: Is Engineering Management a good career choice?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[59]]
#> List of 9
#>  $ by         : chr "apacheCamel"
#>  $ descendants: int 2
#>  $ id         : int 21783100
#>  $ kids       : int [1:2] 21783180 21785124
#>  $ score      : int 3
#>  $ text       : chr "A couple threads on HN have repeated criticism towards Oracle (almost always negative) but never really fully e"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-12-13 17:43:05"
#>  $ title      : chr "Ask HN: Why do so many people have a problem with Oracle?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 

# }