Hacker News latest show stories

Retrieves latest show stories using Hacker News API

get_latest_show_stories(max_items = NULL)

Arguments

max_items
Maximum number of items to retrieve. If max_items = NULL, returns all available

Value

list of latest show HN stories

Details

Parallel api requests can be enabled by running future::plan(future::multiprocess)

Examples

# \donttest{
# get the latest show story on Hacker News
latest_show_story <- get_latest_show_stories(max_items = 1)
latest_show_story

#> [[1]]
#> List of 9
#>  $ by         : chr "lvscar"
#>  $ descendants: int 19
#>  $ id         : int 21780332
#>  $ kids       : int [1:12] 21787255 21789192 21788176 21785215 21787704 21788491 21786851 21786118 21785416 21784952 ...
#>  $ score      : int 105
#>  $ time       : POSIXct[1:1], format: "2019-12-13 10:43:43"
#>  $ title      : chr "Show HN: LastingPaper – A Desktop App for Your E-Ink Kindle"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.lastingpaper.com"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get 10 latest show stories on Hacker News
latest_10_show_stories <- get_latest_show_stories(max_items = 10)
latest_10_show_stories

#> [[1]]
#> List of 9
#>  $ by         : chr "lvscar"
#>  $ descendants: int 19
#>  $ id         : int 21780332
#>  $ kids       : int [1:12] 21787255 21789192 21788176 21785215 21787704 21788491 21786851 21786118 21785416 21784952 ...
#>  $ score      : int 105
#>  $ time       : POSIXct[1:1], format: "2019-12-13 10:43:43"
#>  $ title      : chr "Show HN: LastingPaper – A Desktop App for Your E-Ink Kindle"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.lastingpaper.com"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 8
#>  $ by         : chr "ekoutanov"
#>  $ descendants: int 0
#>  $ id         : int 21788863
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-12-14 11:56:50"
#>  $ title      : chr "Show HN: SpringBoot web app for browsing Kafka topics"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/obsidiandynamics/kafdrop"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 9
#>  $ by         : chr "phacks"
#>  $ descendants: int 12
#>  $ id         : int 21782714
#>  $ kids       : int [1:4] 21784986 21784359 21783307 21783149
#>  $ score      : int 94
#>  $ time       : POSIXct[1:1], format: "2019-12-13 17:02:30"
#>  $ title      : chr "Show HN: Falco – Open-Source Web Performance Monitoring"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/theodo/falco"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 9
#>  $ by         : chr "jeremysizle"
#>  $ descendants: int 48
#>  $ id         : int 21788359
#>  $ kids       : int [1:12] 21789132 21789086 21788897 21788435 21788596 21788452 21788986 21788702 21788558 21789163 ...
#>  $ score      : int 41
#>  $ time       : POSIXct[1:1], format: "2019-12-14 08:49:34"
#>  $ title      : chr "Show HN: Browser Based PowerPoint Alternative"
#>  $ type       : chr "story"
#>  $ url        : chr "https://sizle.io/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 9
#>  $ by         : chr "hachibu"
#>  $ descendants: int 12
#>  $ id         : int 21783003
#>  $ kids       : int [1:5] 21784077 21786282 21784083 21785996 21787690
#>  $ score      : int 12
#>  $ time       : POSIXct[1:1], format: "2019-12-13 17:31:28"
#>  $ title      : chr "Show HN: Be Your Own Password Generator"
#>  $ type       : chr "story"
#>  $ url        : chr "https://hachibu.github.io/be-your-own-password-generator/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 9
#>  $ by         : chr "l1am0"
#>  $ descendants: int 33
#>  $ id         : int 21777722
#>  $ kids       : int [1:8] 21778710 21778312 21780084 21779096 21778281 21778169 21778145 21778095
#>  $ score      : int 51
#>  $ time       : POSIXct[1:1], format: "2019-12-13 00:34:34"
#>  $ title      : chr "Show HN: I built a FOSS service to protect against link shorteners tracking us"
#>  $ type       : chr "story"
#>  $ url        : chr "https://unshort.link/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 8
#>  $ by         : chr "byeokim"
#>  $ descendants: int 0
#>  $ id         : int 21786790
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-12-14 01:00:29"
#>  $ title      : chr "Show HN: Gmailpush – Gmail API Push Notification Handler for Node"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/byeokim/gmailpush"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "kibitan"
#>  $ descendants: int 2
#>  $ id         : int 21780949
#>  $ kids       : int 21786855
#>  $ score      : int 10
#>  $ time       : POSIXct[1:1], format: "2019-12-13 13:23:15"
#>  $ title      : chr "Show HN: Masking – Generate anonymized database"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/kibitan/masking"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 8
#>  $ by         : chr "alexellisuk"
#>  $ descendants: int 0
#>  $ id         : int 21781042
#>  $ score      : int 9
#>  $ time       : POSIXct[1:1], format: "2019-12-13 13:42:37"
#>  $ title      : chr "Show HN: OpenFaaS Classic Templates"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/openfaas/templates"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 9
#>  $ by         : chr "dohliam"
#>  $ descendants: int 4
#>  $ id         : int 21782310
#>  $ kids       : int [1:2] 21782560 21789031
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-12-13 16:20:50"
#>  $ title      : chr "Show HN: Drop-In Minimal CSS – Preview Minimal CSS Frameworks on any HTML Page"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/dohliam/dropin-minimal-css"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get all latest show stories on Hacker News
latest_show_stories <- get_latest_show_stories()
latest_show_stories

#> [[1]]
#> List of 9
#>  $ by         : chr "lvscar"
#>  $ descendants: int 19
#>  $ id         : int 21780332
#>  $ kids       : int [1:12] 21787255 21789192 21788176 21785215 21787704 21788491 21786851 21786118 21785416 21784952 ...
#>  $ score      : int 105
#>  $ time       : POSIXct[1:1], format: "2019-12-13 10:43:43"
#>  $ title      : chr "Show HN: LastingPaper – A Desktop App for Your E-Ink Kindle"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.lastingpaper.com"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 8
#>  $ by         : chr "ekoutanov"
#>  $ descendants: int 0
#>  $ id         : int 21788863
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-12-14 11:56:50"
#>  $ title      : chr "Show HN: SpringBoot web app for browsing Kafka topics"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/obsidiandynamics/kafdrop"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 9
#>  $ by         : chr "phacks"
#>  $ descendants: int 12
#>  $ id         : int 21782714
#>  $ kids       : int [1:4] 21784986 21784359 21783307 21783149
#>  $ score      : int 94
#>  $ time       : POSIXct[1:1], format: "2019-12-13 17:02:30"
#>  $ title      : chr "Show HN: Falco – Open-Source Web Performance Monitoring"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/theodo/falco"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 9
#>  $ by         : chr "jeremysizle"
#>  $ descendants: int 48
#>  $ id         : int 21788359
#>  $ kids       : int [1:12] 21789132 21789086 21788897 21788435 21788596 21788452 21788986 21788702 21788558 21789163 ...
#>  $ score      : int 41
#>  $ time       : POSIXct[1:1], format: "2019-12-14 08:49:34"
#>  $ title      : chr "Show HN: Browser Based PowerPoint Alternative"
#>  $ type       : chr "story"
#>  $ url        : chr "https://sizle.io/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 9
#>  $ by         : chr "hachibu"
#>  $ descendants: int 12
#>  $ id         : int 21783003
#>  $ kids       : int [1:5] 21784077 21786282 21784083 21785996 21787690
#>  $ score      : int 12
#>  $ time       : POSIXct[1:1], format: "2019-12-13 17:31:28"
#>  $ title      : chr "Show HN: Be Your Own Password Generator"
#>  $ type       : chr "story"
#>  $ url        : chr "https://hachibu.github.io/be-your-own-password-generator/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 9
#>  $ by         : chr "l1am0"
#>  $ descendants: int 33
#>  $ id         : int 21777722
#>  $ kids       : int [1:8] 21778710 21778312 21780084 21779096 21778281 21778169 21778145 21778095
#>  $ score      : int 51
#>  $ time       : POSIXct[1:1], format: "2019-12-13 00:34:34"
#>  $ title      : chr "Show HN: I built a FOSS service to protect against link shorteners tracking us"
#>  $ type       : chr "story"
#>  $ url        : chr "https://unshort.link/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 8
#>  $ by         : chr "byeokim"
#>  $ descendants: int 0
#>  $ id         : int 21786790
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-12-14 01:00:29"
#>  $ title      : chr "Show HN: Gmailpush – Gmail API Push Notification Handler for Node"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/byeokim/gmailpush"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "kibitan"
#>  $ descendants: int 2
#>  $ id         : int 21780949
#>  $ kids       : int 21786855
#>  $ score      : int 10
#>  $ time       : POSIXct[1:1], format: "2019-12-13 13:23:15"
#>  $ title      : chr "Show HN: Masking – Generate anonymized database"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/kibitan/masking"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 8
#>  $ by         : chr "alexellisuk"
#>  $ descendants: int 0
#>  $ id         : int 21781042
#>  $ score      : int 9
#>  $ time       : POSIXct[1:1], format: "2019-12-13 13:42:37"
#>  $ title      : chr "Show HN: OpenFaaS Classic Templates"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/openfaas/templates"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 9
#>  $ by         : chr "dohliam"
#>  $ descendants: int 4
#>  $ id         : int 21782310
#>  $ kids       : int [1:2] 21782560 21789031
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-12-13 16:20:50"
#>  $ title      : chr "Show HN: Drop-In Minimal CSS – Preview Minimal CSS Frameworks on any HTML Page"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/dohliam/dropin-minimal-css"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[11]]
#> List of 8
#>  $ by         : chr "Lord_Zero"
#>  $ descendants: int 0
#>  $ id         : int 21782123
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-12-13 15:59:28"
#>  $ title      : chr "Show HN: Patchme.link – a GUI for services like patchbay.pub"
#>  $ type       : chr "story"
#>  $ url        : chr "https://patchme.link"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[12]]
#> List of 9
#>  $ by         : chr "vrathee"
#>  $ descendants: int 3
#>  $ id         : int 21781869
#>  $ kids       : int [1:2] 21781897 21782145
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-12-13 15:33:30"
#>  $ title      : chr "Show HN: Website change detection agent to track websites and get change alert"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.agenty.com/products/change-detection-agent/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[13]]
#> List of 9
#>  $ by         : chr "losteden1"
#>  $ descendants: int 2
#>  $ id         : int 21771940
#>  $ kids       : int [1:2] 21777367 21774724
#>  $ score      : int 29
#>  $ time       : POSIXct[1:1], format: "2019-12-12 15:10:02"
#>  $ title      : chr "Show HN: Octopus.do, a lightning-fast visual sitemap builder 2.0"
#>  $ type       : chr "story"
#>  $ url        : chr "https://octopus.do"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[14]]
#> List of 9
#>  $ by         : chr "applechen777"
#>  $ descendants: int 12
#>  $ id         : int 21766572
#>  $ kids       : int [1:4] 21769588 21770332 21770643 21766754
#>  $ score      : int 32
#>  $ time       : POSIXct[1:1], format: "2019-12-11 21:40:05"
#>  $ title      : chr "Show HN: Learn by Doing Projects for Engineers – Real Dev"
#>  $ type       : chr "story"
#>  $ url        : chr "https://real.dev/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[15]]
#> List of 9
#>  $ by         : chr "fearenales"
#>  $ descendants: int 1
#>  $ id         : int 21772694
#>  $ kids       : int 21788306
#>  $ score      : int 11
#>  $ time       : POSIXct[1:1], format: "2019-12-12 16:39:30"
#>  $ title      : chr "Show HN: Event-driven Shell – a way to control flows in multiple shell windows"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/fearenales/event-driven-shell"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[16]]
#> List of 9
#>  $ by         : chr "conroy"
#>  $ descendants: int 13
#>  $ id         : int 21765689
#>  $ kids       : int [1:6] 21765738 21784355 21765714 21768623 21767729 21783106
#>  $ score      : int 42
#>  $ time       : POSIXct[1:1], format: "2019-12-11 20:08:13"
#>  $ title      : chr "Show HN: sqlc – Compile SQL Queries to Type-Safe Go"
#>  $ type       : chr "story"
#>  $ url        : chr "https://conroy.org/introducing-sqlc"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[17]]
#> List of 9
#>  $ by         : chr "siruva07"
#>  $ descendants: int 4
#>  $ id         : int 21775255
#>  $ kids       : int [1:2] 21778972 21781092
#>  $ score      : int 12
#>  $ time       : POSIXct[1:1], format: "2019-12-12 20:18:10"
#>  $ title      : chr "Show HN: Tap water you can trust"
#>  $ type       : chr "story"
#>  $ url        : chr "https://findtap.com/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[18]]
#> List of 9
#>  $ by         : chr "aloukissas"
#>  $ descendants: int 3
#>  $ id         : int 21782067
#>  $ kids       : int 21782103
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-12-13 15:53:43"
#>  $ title      : chr "Show HN: A free tool to find correlated assets for tax harvesting"
#>  $ type       : chr "story"
#>  $ url        : chr "https://tools.agentrisk.com/tax-harvesting"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[19]]
#> List of 8
#>  $ by         : chr "jstayton"
#>  $ descendants: int 0
#>  $ id         : int 21775326
#>  $ score      : int 9
#>  $ time       : POSIXct[1:1], format: "2019-12-12 20:24:24"
#>  $ title      : chr "Show HN: QueryQL – Filtering, Sorting, and Pagination for Your Node.js REST API"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/TRUEPIC/queryql"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[20]]
#> List of 8
#>  $ by         : chr "deepthought42"
#>  $ descendants: int 0
#>  $ id         : int 21773634
#>  $ score      : int 8
#>  $ time       : POSIXct[1:1], format: "2019-12-12 18:00:49"
#>  $ title      : chr "Show HN: Automate your UI testing with a click of a button"
#>  $ type       : chr "story"
#>  $ url        : chr "http://qanairy.com"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[21]]
#> List of 8
#>  $ by         : chr "salsowelim"
#>  $ descendants: int 0
#>  $ id         : int 21780570
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-12-13 11:40:49"
#>  $ title      : chr "Show HN: Nadyy App:Identify Quran recitation being played near you"
#>  $ type       : chr "story"
#>  $ url        : chr "https://nadyy.app"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[22]]
#> List of 9
#>  $ by         : chr "mabridgland"
#>  $ descendants: int 1
#>  $ id         : int 21766318
#>  $ kids       : int 21787445
#>  $ score      : int 8
#>  $ time       : POSIXct[1:1], format: "2019-12-11 21:14:19"
#>  $ title      : chr "Show HN: Babysploit, the beginner's pentesting framework updated to v1.5"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/M4cs/Babysploit"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[23]]
#> List of 8
#>  $ by         : chr "umuplus"
#>  $ descendants: int 0
#>  $ id         : int 21770487
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-12-12 10:18:05"
#>  $ title      : chr "Show HN: Bee-queue based microservices for Node.js"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/umuplus/beems"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[24]]
#> List of 8
#>  $ by         : chr "tltjr"
#>  $ descendants: int 0
#>  $ id         : int 21781856
#>  $ score      : int 4
#>  $ time       : POSIXct[1:1], format: "2019-12-13 15:32:04"
#>  $ title      : chr "Show HN: GQLHub – Fully hosted, configuration-free GraphQL endpoints in seconds"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.kickstarter.com/projects/thomasthornton/gqlhub-effortless-apis"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[25]]
#> List of 9
#>  $ by         : chr "ryzalyusoff"
#>  $ descendants: int 1
#>  $ id         : int 21777755
#>  $ kids       : int 21777756
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-12-13 00:40:08"
#>  $ title      : chr "Show HN: Reader Mode – Remove clutter and ads from any article. Instantly"
#>  $ type       : chr "story"
#>  $ url        : chr "https://readermode.io"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[26]]
#> List of 9
#>  $ by         : chr "blader_johny"
#>  $ descendants: int 3
#>  $ id         : int 21770046
#>  $ kids       : int [1:2] 21771966 21770142
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-12-12 08:42:26"
#>  $ title      : chr "Show HN: Minimal Help Center Chatbot for Startups"
#>  $ type       : chr "story"
#>  $ url        : chr "https://sparrw.launchaco.com/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[27]]
#> List of 8
#>  $ by         : chr "KloudTrader"
#>  $ descendants: int 0
#>  $ id         : int 21766485
#>  $ score      : int 15
#>  $ time       : POSIXct[1:1], format: "2019-12-11 21:31:16"
#>  $ title      : chr "Show HN: Convert Fanfiction to Audiobooks"
#>  $ type       : chr "story"
#>  $ url        : chr "https://auditus.cc"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[28]]
#> List of 9
#>  $ by         : chr "himanshuy"
#>  $ descendants: int 3
#>  $ id         : int 21764174
#>  $ kids       : int [1:2] 21765743 21764219
#>  $ score      : int 8
#>  $ time       : POSIXct[1:1], format: "2019-12-11 17:45:22"
#>  $ title      : chr "Show HN: Calendar API for Developers"
#>  $ type       : chr "story"
#>  $ url        : chr "https://campaigndays.com"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[29]]
#> List of 8
#>  $ by         : chr "FPurchess"
#>  $ descendants: int 0
#>  $ id         : int 21773359
#>  $ score      : int 4
#>  $ time       : POSIXct[1:1], format: "2019-12-12 17:37:03"
#>  $ title      : chr "Show HN: Lots of Bots – find and compare mobile robots"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.lotsofbots.com/en/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[30]]
#> List of 9
#>  $ by         : chr "tollie93"
#>  $ descendants: int 6
#>  $ id         : int 21773019
#>  $ kids       : int [1:3] 21773317 21786443 21773063
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-12-12 17:06:14"
#>  $ title      : chr "Show HN: A free self-enforcing standard deal for freelancers and their clients"
#>  $ type       : chr "story"
#>  $ url        : chr "http://freelancerprotocol.com/?source=hn"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[31]]
#> List of 8
#>  $ by         : chr "freetonik"
#>  $ descendants: int 0
#>  $ id         : int 21776503
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-12-12 22:10:08"
#>  $ title      : chr "Show HN: Ziptalks – tech talks, summarized and illustrated"
#>  $ type       : chr "story"
#>  $ url        : chr "https://gumroad.com/ziptalks"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[32]]
#> List of 9
#>  $ by         : chr "atum47"
#>  $ descendants: int 1
#>  $ id         : int 21781003
#>  $ kids       : int 21784308
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-12-13 13:35:28"
#>  $ title      : chr "Show HN: Losing an argument online? Respond with fire"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/victorqribeiro/fireEffect"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[33]]
#> List of 8
#>  $ by         : chr "AkihiroSuda"
#>  $ descendants: int 0
#>  $ id         : int 21771628
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-12-12 14:29:06"
#>  $ title      : chr "Show HN: Instance-per-Pod Webhook: IaaS-Level Isolation for Kubernetes Pods"
#>  $ type       : chr "story"
#>  $ url        : chr "https://medium.com/nttlabs/instance-per-pod-bcbfb3ae2985"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[34]]
#> List of 8
#>  $ by         : chr "bilifuduo"
#>  $ descendants: int 0
#>  $ id         : int 21763751
#>  $ score      : int 4
#>  $ time       : POSIXct[1:1], format: "2019-12-11 17:07:16"
#>  $ title      : chr "Show HN: Find high-impact jobs tackling problems like climate change"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.splashwithdolphin.com/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[35]]
#> List of 8
#>  $ by         : chr "acrosync"
#>  $ descendants: int 0
#>  $ id         : int 21763397
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-12-11 16:37:56"
#>  $ title      : chr "Show HN: A new web GUI for Duplicacy – lock-free deduplication backup tool"
#>  $ type       : chr "story"
#>  $ url        : chr "https://duplicacy.com"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[36]]
#> List of 9
#>  $ by         : chr "danieljacksonno"
#>  $ descendants: int 3
#>  $ id         : int 21762643
#>  $ kids       : int [1:2] 21762652 21763072
#>  $ score      : int 8
#>  $ time       : POSIXct[1:1], format: "2019-12-11 15:30:30"
#>  $ title      : chr "Show HN: I taught myself coding to make this app that collects email receipts"
#>  $ type       : chr "story"
#>  $ url        : chr "http://receiptrunner.com/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[37]]
#> List of 9
#>  $ by         : chr "cmtRed"
#>  $ descendants: int 1
#>  $ id         : int 21786340
#>  $ kids       : int [1:2] 21788243 21786466
#>  $ score      : int 1
#>  $ time       : POSIXct[1:1], format: "2019-12-13 23:48:24"
#>  $ title      : chr "Show HN: Community.Red – Create a New Reddit in Seconds"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.community.red"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[38]]
#> List of 8
#>  $ by         : chr "smithgeek"
#>  $ descendants: int 0
#>  $ id         : int 21784761
#>  $ score      : int 1
#>  $ time       : POSIXct[1:1], format: "2019-12-13 20:51:26"
#>  $ title      : chr "Show HN: Realty Voice – Use Alexa when looking at homes to buy"
#>  $ type       : chr "story"
#>  $ url        : chr "http://www.realtyvoice.app"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[39]]
#> List of 8
#>  $ by         : chr "WolfOliver"
#>  $ descendants: int 0
#>  $ id         : int 21783757
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-12-13 18:58:01"
#>  $ title      : chr "Show HN: Distraction free writing of scientific papers or complex blog posts"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.monsterwriter.app/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[40]]
#> List of 8
#>  $ by         : chr "kbruneel"
#>  $ descendants: int 0
#>  $ id         : int 21783635
#>  $ score      : int 1
#>  $ time       : POSIXct[1:1], format: "2019-12-13 18:43:15"
#>  $ title      : chr "Show HN: Autonomous Racing: Proportional Controller"
#>  $ type       : chr "story"
#>  $ url        : chr "https://roboton.io/tutorial/proportional"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[41]]
#> List of 9
#>  $ by         : chr "szemy2"
#>  $ descendants: int 2
#>  $ id         : int 21783566
#>  $ kids       : int [1:2] 21783574 21783668
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-12-13 18:36:24"
#>  $ title      : chr "Show HN: Metascore ranking of Phones"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.techpulser.com/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[42]]
#> List of 8
#>  $ by         : chr "alexzender"
#>  $ descendants: int 0
#>  $ id         : int 21783508
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-12-13 18:30:11"
#>  $ title      : chr "Show HN: More Conversions with Web Push"
#>  $ type       : chr "story"
#>  $ url        : chr "http://pushok.io/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[43]]
#> List of 9
#>  $ by         : chr "dangwu"
#>  $ descendants: int 1
#>  $ id         : int 21783329
#>  $ kids       : int 21789318
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-12-13 18:09:01"
#>  $ title      : chr "Show HN: Octal, a Hacker News App for iOS"
#>  $ type       : chr "story"
#>  $ url        : chr "https://apps.apple.com/us/app/octal/id1308885491"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[44]]
#> List of 8
#>  $ by         : chr "kmlmhnn"
#>  $ descendants: int 0
#>  $ id         : int 21783068
#>  $ score      : int 1
#>  $ time       : POSIXct[1:1], format: "2019-12-13 17:38:50"
#>  $ title      : chr "Show HN: Gowiththeflow – Inverting the inversion of control in go"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/kmlmhnn/gowiththeflow"
#>  - attr(*, "class")= chr "hn_item"
#> 

# }