Hacker News latest show stories ids

Retrieves latest show stories ids using Hacker News API

Usage

get_latest_show_stories_ids()

Value

list of latest show Hacker News stories ids

Examples

# \donttest{
# get ids of latest show stories on Hacker News
show_stories_ids <- get_latest_show_stories_ids()
show_stories_ids
#>   [1] 46121860 46122139 46117937 46119932 46119790 46109843 46119615 46110897
#>   [9] 46119276 46119234 46093473 46109381 46053076 46090619 46117413 46117275
#>  [17] 46058566 46116399 46116100 46051449 46104092 46075882 46113369 46112714
#>  [25] 46100913 46112187 46080737 46111137 46111086 46098912 46062284 46105112
#>  [33] 46109742 46089764 46081443 46102753 46065997 46105672 46090284 46068847
#>  [41] 46069556 46105195 46048252 46104417 46069598 46104158 46097267 46103447
#>  [49] 46103373 46094552 46102695 46068138 46078571 46095474 46101032 46101008
#>  [57] 46100919 46100623 46100402 46100290 46099639 46099403 46099130 46098795
#>  [65] 46091442 46097293 46108404 46105103 46096369 46096192 46095915 46070676
#>  [73] 46087793 46108902 46095202 46087339 46094330 46092853 46075968 46091415
#>  [81] 46091036 46080141 46089344 46088823 46088732 46088179 46087405 46087224
#>  [89] 46087160 46086845 46057766 46086314 46086086 46065449 46057341 46083278
#>  [97] 46082962 46082916 46046004 46082045 46081666 46081520 46081403 46081007
#> [105] 46080958 46080138 46079163 46052530 46078010 46077764 46079963 46061453
#> [113] 46076356 46076303 46075307 46075018 46053385
# }