Hacker News updated profiles

Retrieves changed item ids and changed profile usernames using Hacker News API

Usage

get_updates()

Value

list of changed item ids and list of changed profile usernames

Examples

# \donttest{
# get ids/usernames of updated items/profiles
hn_updates <- get_updates()
hn_updates
#> $items
#>  [1] 43615024 43615149 43614059 43611074 43615136 43613460 43615150 43610736
#>  [9] 43612631 43615151 43615012 43615069 43609118 43615018 43613691 43609820
#> [17] 43614359 43614416 43611944 43610923 43506763 43610717 43615049 43612913
#> [25] 43615040 43612986 43609496 43614928 43615153 43613970 43562054 43615154
#> [33] 43614788 43613434 43615037 43609635 43615098 43614593 43611759 43613199
#> [41] 43614983 43601213 43610678 43577490 43609661 43613079 43605003 43610527
#> [49] 43613305 43615152 43614592 43580815 43613194 43612807 43603307 43613180
#> [57] 43609242 43495644 43614534 43614285 43605695 43571332 43584208 43614199
#> [65] 43606315 43547609 43600582 43558464 43610760 43612765 43595269 43614505
#> [73] 43603646
#> 
#> $profiles
#>  [1] "gnabgib"         "oyashirochama"   "diggan"          "justswim"       
#>  [5] "johnisgood"      "jmuguy"          "addicted"        "DiggyJohnson"   
#>  [9] "TaurenHunter"    "jasonlotito"     "rickdeckard"     "rs186"          
#> [13] "Mohanrraj"       "tpm"             "em3rgent0rdr"    "gamblor956"     
#> [17] "nickpell"        "Bhilai"          "sebg"            "QuantumSeed"    
#> [21] "mrcjkb"          "billybuckwheat"  "foobahify"       "ndr"            
#> [25] "opem"            "exotic_porpoise" "SupremumLimit"   "adisann"        
#> [29] "dylan604"        "xivzgrev"       
#> 
# }