Hacker News latest ask stories

Retrieves latest ask stories using Hacker News API

Usage

get_latest_ask_stories(max_items = NULL)

Arguments

max_items

Maximum number of items to retrieve. If max_items = NULL, returns all available

Value

list of latest ask HN stories ids

Examples

# \donttest{
# get the lastest ask story on Hacker News
latest_ask_story <- get_latest_ask_stories(max_items = 1)
latest_ask_story
#> [[1]]
#> List of 9
#>  $ by         : chr "salestoscale"
#>  $ descendants: int 34
#>  $ id         : int 43584547
#>  $ kids       : int [1:21] 43585926 43585437 43584867 43586188 43585903 43587161 43585333 43586129 43585925 43586859 ...
#>  $ score      : int 28
#>  $ text       : chr "I recently launched an independent sales agency specializing in early-stage startups.  My model is adapted from"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-04 16:15:48"
#>  $ title      : chr "Ask HN: As a Founder, is this business model compelling?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get 10 latest ask stories on Hacker News
latest_10_ask_stories <- get_latest_ask_stories(max_items = 10)
latest_10_ask_stories
#> [[1]]
#> List of 9
#>  $ by         : chr "salestoscale"
#>  $ descendants: int 34
#>  $ id         : int 43584547
#>  $ kids       : int [1:21] 43585926 43585437 43584867 43586188 43585903 43587161 43585333 43586129 43585925 43586859 ...
#>  $ score      : int 28
#>  $ text       : chr "I recently launched an independent sales agency specializing in early-stage startups.  My model is adapted from"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-04 16:15:48"
#>  $ title      : chr "Ask HN: As a Founder, is this business model compelling?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 9
#>  $ by         : chr "whoishiring"
#>  $ descendants: int 385
#>  $ id         : int 43547611
#>  $ kids       : int [1:272] 43586785 43586321 43564151 43584569 43572590 43566287 43580161 43548465 43572200 43552927 ...
#>  $ score      : int 253
#>  $ text       : chr "Please state the location and include REMOTE for remote work, REMOTE (US)\nor similar if the country is restric"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 15:01:11"
#>  $ title      : chr "Ask HN: Who is hiring? (April 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 9
#>  $ by         : chr "whoishiring"
#>  $ descendants: int 238
#>  $ id         : int 43547609
#>  $ kids       : int [1:247] 43587885 43587686 43587261 43587063 43586224 43587205 43585089 43584336 43584386 43586925 ...
#>  $ score      : int 66
#>  $ text       : chr "Share your information if you are looking for work. Please use this format:<p><pre><code>  Location:\n  Remote:"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 15:01:11"
#>  $ title      : chr "Ask HN: Who wants to be hired? (April 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 9
#>  $ by         : chr "dang"
#>  $ descendants: int 462
#>  $ id         : int 43558671
#>  $ kids       : int [1:122] 43560481 43559535 43559609 43565390 43561038 43560360 43559266 43561913 43564571 43559576 ...
#>  $ score      : int 1792
#>  $ text       : chr "Hi all,<p>Tom Howard is going public as HN moderator today. He has been doing HN moderation work for years alre"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 16:49:56"
#>  $ title      : chr "Tell HN: Announcing tomhow as a public moderator"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 9
#>  $ by         : chr "bavarianbob"
#>  $ descendants: int 36
#>  $ id         : int 43548589
#>  $ kids       : int [1:12] 43548817 43548695 43550246 43548950 43553438 43550191 43550012 43550236 43565416 43549804 ...
#>  $ score      : int 119
#>  $ text       : chr "EDIT: Back online?!<p>NPM discussion: <a href=\"https:&#x2F;&#x2F;github.com&#x2F;npm&#x2F;cli&#x2F;issues&#x2F"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 16:19:28"
#>  $ title      : chr "Tell HN: Camelgate NPM Outage (Cloudflare)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 9
#>  $ by         : chr "david927"
#>  $ descendants: int 972
#>  $ id         : int 43527452
#>  $ kids       : int [1:592] 43531555 43528885 43528339 43527869 43528428 43529032 43527768 43535835 43532927 43531691 ...
#>  $ score      : int 385
#>  $ text       : chr "What are you working on?  Any new ideas that you&#x27;re thinking about?"
#>  $ time       : POSIXct[1:1], format: "2025-03-30 20:43:15"
#>  $ title      : chr "Ask HN: What are you working on? (March 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 9
#>  $ by         : chr "Brysonbw"
#>  $ descendants: int 64
#>  $ id         : int 43559733
#>  $ kids       : int [1:43] 43560914 43563603 43560923 43560393 43560318 43569894 43560388 43561195 43560449 43560297 ...
#>  $ score      : int 79
#>  $ text       : chr "How would one go about being a &#x27;rogue&#x27; OSS contributor so to speak? Live off of donations, bounties, "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 18:24:51"
#>  $ title      : chr "Ask HN: How do you make a living contributing to and/or creating OSS projects?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "OrbitalShotput_"
#>  $ descendants: int 17
#>  $ id         : int 43564550
#>  $ kids       : int [1:8] 43564583 43565515 43568498 43566094 43572997 43565857 43566758 43576142
#>  $ score      : int 17
#>  $ text       : chr "This is coming up a lot these days in the news- but Customs and Border Patrol have increased the amount of sear"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 03:53:26"
#>  $ title      : chr "Ask HN: For the privacy minded, How do you prepare for gadget Border Searches?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 9
#>  $ by         : chr "r_singh"
#>  $ descendants: int 8
#>  $ id         : int 43572949
#>  $ kids       : int [1:8] 43582622 43573331 43578085 43573578 43574111 43578048 43576099 43574733
#>  $ score      : int 12
#>  $ text       : chr "I feel like talking about how to think is going to be a bigger topic of discussion going forward now that we ha"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 17:35:48"
#>  $ title      : chr "Ask HN: Best books written on “How to think”?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 9
#>  $ by         : chr "mywittyname"
#>  $ descendants: int 39
#>  $ id         : int 43539544
#>  $ kids       : int [1:21] 43541034 43582469 43540390 43541108 43540394 43549407 43541277 43542447 43542906 43554212 ...
#>  $ score      : int 59
#>  $ text       : chr "Currently, I use Perplexity or ChatGPT via the web prompt for small coding tasks, but sometimes I&#x27;ll use O"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-03-31 20:26:35"
#>  $ title      : chr "Ask HN: What's the best way to get started with LLM-assisted programing?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get all latest ask stories on Hacker News
latest_ask_stories <- get_latest_ask_stories()
latest_ask_stories
#> [[1]]
#> List of 9
#>  $ by         : chr "salestoscale"
#>  $ descendants: int 34
#>  $ id         : int 43584547
#>  $ kids       : int [1:21] 43585926 43585437 43584867 43586188 43585903 43587161 43585333 43586129 43585925 43586859 ...
#>  $ score      : int 28
#>  $ text       : chr "I recently launched an independent sales agency specializing in early-stage startups.  My model is adapted from"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-04 16:15:48"
#>  $ title      : chr "Ask HN: As a Founder, is this business model compelling?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 9
#>  $ by         : chr "whoishiring"
#>  $ descendants: int 385
#>  $ id         : int 43547611
#>  $ kids       : int [1:272] 43586785 43586321 43564151 43584569 43572590 43566287 43580161 43548465 43572200 43552927 ...
#>  $ score      : int 253
#>  $ text       : chr "Please state the location and include REMOTE for remote work, REMOTE (US)\nor similar if the country is restric"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 15:01:11"
#>  $ title      : chr "Ask HN: Who is hiring? (April 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 9
#>  $ by         : chr "whoishiring"
#>  $ descendants: int 238
#>  $ id         : int 43547609
#>  $ kids       : int [1:247] 43587885 43587686 43587261 43587063 43586224 43587205 43585089 43584336 43584386 43586925 ...
#>  $ score      : int 66
#>  $ text       : chr "Share your information if you are looking for work. Please use this format:<p><pre><code>  Location:\n  Remote:"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 15:01:11"
#>  $ title      : chr "Ask HN: Who wants to be hired? (April 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 9
#>  $ by         : chr "dang"
#>  $ descendants: int 462
#>  $ id         : int 43558671
#>  $ kids       : int [1:122] 43560481 43559535 43559609 43565390 43561038 43560360 43559266 43561913 43564571 43559576 ...
#>  $ score      : int 1792
#>  $ text       : chr "Hi all,<p>Tom Howard is going public as HN moderator today. He has been doing HN moderation work for years alre"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 16:49:56"
#>  $ title      : chr "Tell HN: Announcing tomhow as a public moderator"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 9
#>  $ by         : chr "bavarianbob"
#>  $ descendants: int 36
#>  $ id         : int 43548589
#>  $ kids       : int [1:12] 43548817 43548695 43550246 43548950 43553438 43550191 43550012 43550236 43565416 43549804 ...
#>  $ score      : int 119
#>  $ text       : chr "EDIT: Back online?!<p>NPM discussion: <a href=\"https:&#x2F;&#x2F;github.com&#x2F;npm&#x2F;cli&#x2F;issues&#x2F"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 16:19:28"
#>  $ title      : chr "Tell HN: Camelgate NPM Outage (Cloudflare)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 9
#>  $ by         : chr "david927"
#>  $ descendants: int 972
#>  $ id         : int 43527452
#>  $ kids       : int [1:592] 43531555 43528885 43528339 43527869 43528428 43529032 43527768 43535835 43532927 43531691 ...
#>  $ score      : int 385
#>  $ text       : chr "What are you working on?  Any new ideas that you&#x27;re thinking about?"
#>  $ time       : POSIXct[1:1], format: "2025-03-30 20:43:15"
#>  $ title      : chr "Ask HN: What are you working on? (March 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 9
#>  $ by         : chr "Brysonbw"
#>  $ descendants: int 64
#>  $ id         : int 43559733
#>  $ kids       : int [1:43] 43560914 43563603 43560923 43560393 43560318 43569894 43560388 43561195 43560449 43560297 ...
#>  $ score      : int 79
#>  $ text       : chr "How would one go about being a &#x27;rogue&#x27; OSS contributor so to speak? Live off of donations, bounties, "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 18:24:51"
#>  $ title      : chr "Ask HN: How do you make a living contributing to and/or creating OSS projects?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "OrbitalShotput_"
#>  $ descendants: int 17
#>  $ id         : int 43564550
#>  $ kids       : int [1:8] 43564583 43565515 43568498 43566094 43572997 43565857 43566758 43576142
#>  $ score      : int 17
#>  $ text       : chr "This is coming up a lot these days in the news- but Customs and Border Patrol have increased the amount of sear"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 03:53:26"
#>  $ title      : chr "Ask HN: For the privacy minded, How do you prepare for gadget Border Searches?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 9
#>  $ by         : chr "r_singh"
#>  $ descendants: int 8
#>  $ id         : int 43572949
#>  $ kids       : int [1:8] 43582622 43573331 43578085 43573578 43574111 43578048 43576099 43574733
#>  $ score      : int 12
#>  $ text       : chr "I feel like talking about how to think is going to be a bigger topic of discussion going forward now that we ha"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 17:35:48"
#>  $ title      : chr "Ask HN: Best books written on “How to think”?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 9
#>  $ by         : chr "mywittyname"
#>  $ descendants: int 39
#>  $ id         : int 43539544
#>  $ kids       : int [1:21] 43541034 43582469 43540390 43541108 43540394 43549407 43541277 43542447 43542906 43554212 ...
#>  $ score      : int 59
#>  $ text       : chr "Currently, I use Perplexity or ChatGPT via the web prompt for small coding tasks, but sometimes I&#x27;ll use O"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-03-31 20:26:35"
#>  $ title      : chr "Ask HN: What's the best way to get started with LLM-assisted programing?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[11]]
#> List of 9
#>  $ by         : chr "LouisLazaris"
#>  $ descendants: int 14
#>  $ id         : int 43550843
#>  $ kids       : int [1:7] 43553648 43550981 43573884 43557366 43557521 43554307 43572809
#>  $ score      : int 11
#>  $ text       : chr "I’ve been involved in web dev in different forms for 20 years, but I’ve never done anything with these types of"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 20:10:08"
#>  $ title      : chr "Ask HN: Are Squarespace and Wix sites worth it?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[12]]
#> List of 9
#>  $ by         : chr "spacebanana7"
#>  $ descendants: int 187
#>  $ id         : int 43547309
#>  $ kids       : int [1:41] 43547461 43547919 43547462 43548054 43547594 43547586 43547535 43547492 43547416 43547545 ...
#>  $ score      : int 188
#>  $ text       : chr "I appreciate developing ROCm into something competitive with CUDA would require a lot of work, both internally "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 14:37:54"
#>  $ title      : chr "Ask HN: Why hasn’t AMD made a viable CUDA alternative?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[13]]
#> List of 9
#>  $ by         : chr "jeyzolo"
#>  $ descendants: int 1
#>  $ id         : int 43561104
#>  $ kids       : int 43561286
#>  $ score      : int 2
#>  $ text       : chr "As an independent developer, I write code at home every day. I have been wondering if I am too out of the marke"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 20:26:01"
#>  $ title      : chr "For a solo funder, I think social is to connect on the internet"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[14]]
#> List of 9
#>  $ by         : chr "sonderotis"
#>  $ descendants: int 2
#>  $ id         : int 43564410
#>  $ kids       : int [1:2] 43564430 43575174
#>  $ score      : int 8
#>  $ text       : chr "I have a blog but it seems most people who are with my content hardly like it.\nI had a review recently that sa"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 03:33:51"
#>  $ title      : chr "Ask HN: How to write better tech blogs"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[15]]
#> List of 8
#>  $ by         : chr "mooreds"
#>  $ descendants: int 0
#>  $ id         : int 43570742
#>  $ score      : int 16
#>  $ text       : chr "A few days ago I posted Vets Who Code[0] on HN[1]. It got put into the second chance pool and made the front pa"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 14:59:23"
#>  $ title      : chr "Tell HN: Effects of the front page on a non-profit"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[16]]
#> List of 9
#>  $ by         : chr "unrealizex"
#>  $ descendants: int 2
#>  $ id         : int 43563174
#>  $ kids       : int [1:3] 43563339 43564553 43563324
#>  $ score      : int 12
#>  $ text       : chr "Hey HN&#x2F;Reddit,\nWe&#x27;re a team of AI researchers passionate about simplifying 3D modeling. We&#x27;ve b"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 23:48:43"
#>  $ title      : chr "We Built an AI Tool to Create High-Quality 3D Models from Regular Videos"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[17]]
#> List of 8
#>  $ by         : chr "anon115"
#>  $ descendants: int 2
#>  $ id         : int 43583827
#>  $ kids       : int [1:2] 43584495 43584540
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2025-04-04 15:27:18"
#>  $ title      : chr "Ask HN: Should you launch your stupid ideas too?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[18]]
#> List of 9
#>  $ by         : chr "frabia"
#>  $ descendants: int 13
#>  $ id         : int 43544254
#>  $ kids       : int [1:16] 43545403 43544777 43549448 43549431 43561205 43550446 43549108 43547393 43545297 43574677 ...
#>  $ score      : int 21
#>  $ text       : chr "Between newsletters, podcasts, Youtube videos, (tech) news, articles, and (of course) books, how does one keep "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 08:33:50"
#>  $ title      : chr "When so much knowledge is produced every day, how do you keep up?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[19]]
#> List of 9
#>  $ by         : chr "krrishd"
#>  $ descendants: int 25
#>  $ id         : int 43542213
#>  $ kids       : int [1:21] 43543288 43576842 43545293 43563006 43547072 43542498 43545930 43547738 43553957 43543213 ...
#>  $ score      : int 11
#>  $ text       : chr "Question says it all - I&#x27;m curious what the state of the art is for a community like HN (that, intuitively"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 02:26:43"
#>  $ title      : chr "Ask HN: Do you still self-host a blog? What's your publishing stack?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[20]]
#> List of 9
#>  $ by         : chr "warnke"
#>  $ descendants: int 1
#>  $ id         : int 43564752
#>  $ kids       : int [1:3] 43567121 43564884 43565971
#>  $ score      : int 10
#>  $ text       : chr "Heyah,<p>This is Lukasz. I am running Wisent, a representation engineering company. I created guardrails that a"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 04:28:44"
#>  $ title      : chr "Latent Space Guardrails That Reduce Hallucinations by 43 Percent Now Open Source"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[21]]
#> List of 9
#>  $ by         : chr "reimertz"
#>  $ descendants: int 20
#>  $ id         : int 43569524
#>  $ kids       : int [1:5] 43574273 43576367 43574335 43570261 43572111
#>  $ score      : int 13
#>  $ text       : chr "I am currently working remotely from EU against US and I wonder how the ongoing and escalating trade wars will "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 13:43:50"
#>  $ title      : chr "Ask HN: How will Trump tarrifs affect remote workers outside US?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[22]]
#> List of 9
#>  $ by         : chr "dezzadk"
#>  $ descendants: int 1
#>  $ id         : int 43560813
#>  $ kids       : int 43560965
#>  $ score      : int 4
#>  $ text       : chr "I was looking at https:&#x2F;&#x2F;github.com&#x2F;notifications when I accidentally pressed the mouse over the"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 19:54:40"
#>  $ title      : chr "Unfollow accidents on GitHub/notifications is too darn easy"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[23]]
#> List of 9
#>  $ by         : chr "mecHacker"
#>  $ descendants: int 26
#>  $ id         : int 43534185
#>  $ kids       : int [1:12] 43535625 43535820 43535809 43535456 43535851 43535623 43535673 43534999 43535640 43534977 ...
#>  $ score      : int 38
#>  $ text       : chr "Been wrestling with this for a while, figured some of you might have seen similar.\nThe Situation:\n* 15+ years"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-03-31 12:26:46"
#>  $ title      : chr "Ask HN: Finding Remote Enterprise SW Consulting Leads: Non-US Resident Edition"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[24]]
#> List of 8
#>  $ by         : chr "magundu"
#>  $ descendants: int 0
#>  $ id         : int 43559819
#>  $ score      : int 3
#>  $ text       : chr "I’m developing a system in Go that maintains a fixed pool of Docker containers (e.g., 10) running a specific im"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 18:31:05"
#>  $ title      : chr "Ask HN: Go libraries for managing Docker container pools and executing commands?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[25]]
#> List of 9
#>  $ by         : chr "tempcal"
#>  $ descendants: int 16
#>  $ id         : int 43533212
#>  $ kids       : int [1:11] 43534718 43586937 43537565 43554951 43538433 43533302 43533277 43534082 43533297 43541678 ...
#>  $ score      : int 13
#>  $ text       : chr "Employed. Unfairly compensated. Old. Love AI. How to capitalize on the new tech wave?"
#>  $ time       : POSIXct[1:1], format: "2025-03-31 10:17:04"
#>  $ title      : chr "Ask HN: How to make money in the new age of AI?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[26]]
#> List of 9
#>  $ by         : chr "wiraashm"
#>  $ descendants: int 3
#>  $ id         : int 43548999
#>  $ kids       : int [1:3] 43555377 43549911 43583973
#>  $ score      : int 4
#>  $ text       : chr "What do you guys think about a virtual whiteboard that has a LLM baked in as a thinking assistant providing ins"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 16:55:11"
#>  $ title      : chr "A Virtual Whiteboard That Has a Thinking Assistant Baked In"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[27]]
#> List of 9
#>  $ by         : chr "billconan"
#>  $ descendants: int 4
#>  $ id         : int 43548599
#>  $ kids       : int [1:2] 43571610 43550585
#>  $ score      : int 8
#>  $ text       : chr "I remember there was an old but popular post on HN front page about a failed startup project. The project is ab"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 16:19:56"
#>  $ title      : chr "Ask HN: Could you help me find an old post?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[28]]
#> List of 9
#>  $ by         : chr "lukejkwarren"
#>  $ descendants: int 5
#>  $ id         : int 43556197
#>  $ kids       : int [1:5] 43563103 43579337 43559985 43556218 43560266
#>  $ score      : int 3
#>  $ text       : chr "I recently built [PenZen](https:&#x2F;&#x2F;penzen.app). It scans websites for real vulnerabilities (not just S"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 13:01:36"
#>  $ title      : chr "Ask HN: What do you use to monitor website security (vulns, uptime, etc.)?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[29]]
#> List of 8
#>  $ by         : chr "dnpotter"
#>  $ descendants: int 0
#>  $ id         : int 43555518
#>  $ score      : int 2
#>  $ text       : chr "I&#x27;m toying with launching a secure data vault application. It&#x27;s designed to be a dropbox-like app but"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 11:16:33"
#>  $ title      : chr "Validating ZenBox – a self-hosted encrypted data vault"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[30]]
#> List of 9
#>  $ by         : chr "kinnth"
#>  $ descendants: int 2
#>  $ id         : int 43534873
#>  $ kids       : int [1:3] 43578204 43553988 43534901
#>  $ score      : int 2
#>  $ text       : chr "I have an 82 year old friend who I help and care for. She has about 10% of her sight left and she uses and need"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-03-31 13:33:57"
#>  $ title      : chr "Ask HN: An Android build/skin specifically created for Almost Blind People?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
# }