Hacker News latest ask stories

Retrieves latest ask stories using Hacker News API

Usage

get_latest_ask_stories(max_items = NULL)

Arguments

max_items

Maximum number of items to retrieve. If max_items = NULL, returns all available

Value

list of latest ask HN stories ids

Examples

# \donttest{
# get the lastest ask story on Hacker News
latest_ask_story <- get_latest_ask_stories(max_items = 1)
latest_ask_story
#> [[1]]
#> List of 8
#>  $ by         : chr "herberh2"
#>  $ descendants: int 0
#>  $ id         : int 43604299
#>  $ score      : int 2
#>  $ text       : chr "What are the best AI tools out there that can help take a product idea and refine it to get it to an MVP?<p>Not"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 19:41:48"
#>  $ title      : chr "What are the best AI tools to turn an idea into MVP?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get 10 latest ask stories on Hacker News
latest_10_ask_stories <- get_latest_ask_stories(max_items = 10)
latest_10_ask_stories
#> [[1]]
#> List of 8
#>  $ by         : chr "herberh2"
#>  $ descendants: int 0
#>  $ id         : int 43604299
#>  $ score      : int 2
#>  $ text       : chr "What are the best AI tools out there that can help take a product idea and refine it to get it to an MVP?<p>Not"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 19:41:48"
#>  $ title      : chr "What are the best AI tools to turn an idea into MVP?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 8
#>  $ by         : chr "muddi900"
#>  $ descendants: int 0
#>  $ id         : int 43603574
#>  $ score      : int 4
#>  $ text       : chr "The issue I have had with llms since the start has been the slop. Contrary to popular belief, the open web was "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 18:17:51"
#>  $ title      : chr "Ask HN: How to Use \"Deep Research\"?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 9
#>  $ by         : chr "whoishiring"
#>  $ descendants: int 406
#>  $ id         : int 43547611
#>  $ kids       : int [1:282] 43603775 43548465 43564151 43566287 43563428 43552927 43550621 43572590 43555467 43600107 ...
#>  $ score      : int 267
#>  $ text       : chr "Please state the location and include REMOTE for remote work, REMOTE (US)\nor similar if the country is restric"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 15:01:11"
#>  $ title      : chr "Ask HN: Who is hiring? (April 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 9
#>  $ by         : chr "whoishiring"
#>  $ descendants: int 263
#>  $ id         : int 43547609
#>  $ kids       : int [1:272] 43602281 43601354 43601288 43601532 43600395 43599506 43598817 43597860 43591361 43597400 ...
#>  $ score      : int 72
#>  $ text       : chr "Share your information if you are looking for work. Please use this format:<p><pre><code>  Location:\n  Remote:"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 15:01:11"
#>  $ title      : chr "Ask HN: Who wants to be hired? (April 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 9
#>  $ by         : chr "dang"
#>  $ descendants: int 464
#>  $ id         : int 43558671
#>  $ kids       : int [1:124] 43560481 43559535 43559609 43561038 43565390 43560360 43559266 43561913 43564571 43559244 ...
#>  $ score      : int 1805
#>  $ text       : chr "Hi all,<p>Tom Howard is going public as HN moderator today. He has been doing HN moderation work for years alre"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 16:49:56"
#>  $ title      : chr "Tell HN: Announcing tomhow as a public moderator"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 8
#>  $ by         : chr "pkkm"
#>  $ descendants: int 0
#>  $ id         : int 43601341
#>  $ score      : int 7
#>  $ text       : chr "There&#x27;s a lot of options for concurrency. In Python alone, you can use threads with shared memory, threads"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 13:37:14"
#>  $ title      : chr "Ask HN: Easiest and hardest concurrency models to use correctly?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 9
#>  $ by         : chr "david927"
#>  $ descendants: int 978
#>  $ id         : int 43527452
#>  $ kids       : int [1:597] 43531555 43528885 43528339 43527869 43528428 43529032 43527768 43535835 43532927 43528408 ...
#>  $ score      : int 387
#>  $ text       : chr "What are you working on?  Any new ideas that you&#x27;re thinking about?"
#>  $ time       : POSIXct[1:1], format: "2025-03-30 20:43:15"
#>  $ title      : chr "Ask HN: What are you working on? (March 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "bavarianbob"
#>  $ descendants: int 36
#>  $ id         : int 43548589
#>  $ kids       : int [1:12] 43548817 43548695 43550246 43548950 43553438 43550191 43550012 43550236 43549804 43565416 ...
#>  $ score      : int 119
#>  $ text       : chr "EDIT: Back online?!<p>NPM discussion: <a href=\"https:&#x2F;&#x2F;github.com&#x2F;npm&#x2F;cli&#x2F;issues&#x2F"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 16:19:28"
#>  $ title      : chr "Tell HN: Camelgate NPM Outage (Cloudflare)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 9
#>  $ by         : chr "OulaX"
#>  $ descendants: int 7
#>  $ id         : int 43591673
#>  $ kids       : int [1:5] 43595140 43595418 43592364 43591987 43591994
#>  $ score      : int 13
#>  $ text       : chr "I have a degree in Computer Science and currently work as a frontend web developer.<p>I live in a developing co"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-05 07:38:13"
#>  $ title      : chr "Ask HN: Looking to Break into Cybersecurity – Where Do I Start?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 8
#>  $ by         : chr "N3Xxus_6"
#>  $ descendants: int 0
#>  $ id         : int 43591349
#>  $ score      : int 9
#>  $ text       : chr "I’m experimenting with a novel approach that integrates symbolic logic directly into a transformer’s attention "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-05 06:27:49"
#>  $ title      : chr "Novel Logic-Enhanced LLM for Improved Symbolic Reasoning"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get all latest ask stories on Hacker News
latest_ask_stories <- get_latest_ask_stories()
latest_ask_stories
#> [[1]]
#> List of 8
#>  $ by         : chr "herberh2"
#>  $ descendants: int 0
#>  $ id         : int 43604299
#>  $ score      : int 2
#>  $ text       : chr "What are the best AI tools out there that can help take a product idea and refine it to get it to an MVP?<p>Not"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 19:41:48"
#>  $ title      : chr "What are the best AI tools to turn an idea into MVP?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 8
#>  $ by         : chr "muddi900"
#>  $ descendants: int 0
#>  $ id         : int 43603574
#>  $ score      : int 4
#>  $ text       : chr "The issue I have had with llms since the start has been the slop. Contrary to popular belief, the open web was "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 18:17:51"
#>  $ title      : chr "Ask HN: How to Use \"Deep Research\"?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 9
#>  $ by         : chr "whoishiring"
#>  $ descendants: int 406
#>  $ id         : int 43547611
#>  $ kids       : int [1:282] 43603775 43548465 43564151 43566287 43563428 43552927 43550621 43572590 43555467 43600107 ...
#>  $ score      : int 267
#>  $ text       : chr "Please state the location and include REMOTE for remote work, REMOTE (US)\nor similar if the country is restric"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 15:01:11"
#>  $ title      : chr "Ask HN: Who is hiring? (April 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 9
#>  $ by         : chr "whoishiring"
#>  $ descendants: int 263
#>  $ id         : int 43547609
#>  $ kids       : int [1:272] 43602281 43601354 43601288 43601532 43600395 43599506 43598817 43597860 43591361 43597400 ...
#>  $ score      : int 72
#>  $ text       : chr "Share your information if you are looking for work. Please use this format:<p><pre><code>  Location:\n  Remote:"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 15:01:11"
#>  $ title      : chr "Ask HN: Who wants to be hired? (April 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 9
#>  $ by         : chr "dang"
#>  $ descendants: int 464
#>  $ id         : int 43558671
#>  $ kids       : int [1:124] 43560481 43559535 43559609 43561038 43565390 43560360 43559266 43561913 43564571 43559244 ...
#>  $ score      : int 1805
#>  $ text       : chr "Hi all,<p>Tom Howard is going public as HN moderator today. He has been doing HN moderation work for years alre"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 16:49:56"
#>  $ title      : chr "Tell HN: Announcing tomhow as a public moderator"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 8
#>  $ by         : chr "pkkm"
#>  $ descendants: int 0
#>  $ id         : int 43601341
#>  $ score      : int 7
#>  $ text       : chr "There&#x27;s a lot of options for concurrency. In Python alone, you can use threads with shared memory, threads"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 13:37:14"
#>  $ title      : chr "Ask HN: Easiest and hardest concurrency models to use correctly?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 9
#>  $ by         : chr "david927"
#>  $ descendants: int 978
#>  $ id         : int 43527452
#>  $ kids       : int [1:597] 43531555 43528885 43528339 43527869 43528428 43529032 43527768 43535835 43532927 43528408 ...
#>  $ score      : int 387
#>  $ text       : chr "What are you working on?  Any new ideas that you&#x27;re thinking about?"
#>  $ time       : POSIXct[1:1], format: "2025-03-30 20:43:15"
#>  $ title      : chr "Ask HN: What are you working on? (March 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "bavarianbob"
#>  $ descendants: int 36
#>  $ id         : int 43548589
#>  $ kids       : int [1:12] 43548817 43548695 43550246 43548950 43553438 43550191 43550012 43550236 43549804 43565416 ...
#>  $ score      : int 119
#>  $ text       : chr "EDIT: Back online?!<p>NPM discussion: <a href=\"https:&#x2F;&#x2F;github.com&#x2F;npm&#x2F;cli&#x2F;issues&#x2F"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 16:19:28"
#>  $ title      : chr "Tell HN: Camelgate NPM Outage (Cloudflare)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 9
#>  $ by         : chr "OulaX"
#>  $ descendants: int 7
#>  $ id         : int 43591673
#>  $ kids       : int [1:5] 43595140 43595418 43592364 43591987 43591994
#>  $ score      : int 13
#>  $ text       : chr "I have a degree in Computer Science and currently work as a frontend web developer.<p>I live in a developing co"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-05 07:38:13"
#>  $ title      : chr "Ask HN: Looking to Break into Cybersecurity – Where Do I Start?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 8
#>  $ by         : chr "N3Xxus_6"
#>  $ descendants: int 0
#>  $ id         : int 43591349
#>  $ score      : int 9
#>  $ text       : chr "I’m experimenting with a novel approach that integrates symbolic logic directly into a transformer’s attention "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-05 06:27:49"
#>  $ title      : chr "Novel Logic-Enhanced LLM for Improved Symbolic Reasoning"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[11]]
#> List of 9
#>  $ by         : chr "franczesko"
#>  $ descendants: int 13
#>  $ id         : int 43593627
#>  $ kids       : int [1:14] 43594001 43594367 43599256 43594038 43593990 43594530 43594050 43595870 43602338 43593687 ...
#>  $ score      : int 12
#>  $ text       : chr "HN account for majority of my open tabs. With a very dynamic top submissions list, I struggle to keep up with a"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-05 14:02:42"
#>  $ title      : chr "Ask HN: How do you consume HN?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[12]]
#> List of 9
#>  $ by         : chr "hsiYwko"
#>  $ descendants: int 4
#>  $ id         : int 43596438
#>  $ kids       : int [1:4] 43600699 43599954 43597727 43599081
#>  $ score      : int 4
#>  $ text       : chr "Hi HN!<p>This past weekend, I built a tool called GitSummarize, it&#x27;s an open-source ai-powered documentati"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-05 20:09:39"
#>  $ title      : chr "Ask HN: As an eng leader, how useful is AI-generated internal documentation?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[13]]
#> List of 9
#>  $ by         : chr "Brysonbw"
#>  $ descendants: int 66
#>  $ id         : int 43559733
#>  $ kids       : int [1:45] 43560914 43563603 43560923 43560393 43560318 43560388 43569894 43561195 43560449 43560297 ...
#>  $ score      : int 82
#>  $ text       : chr "How would one go about being a &#x27;rogue&#x27; OSS contributor so to speak? Live off of donations, bounties, "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 18:24:51"
#>  $ title      : chr "Ask HN: How do you make a living contributing to and/or creating OSS projects?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[14]]
#> List of 8
#>  $ by         : chr "kevthecoder"
#>  $ descendants: int 0
#>  $ id         : int 43592201
#>  $ score      : int 12
#>  $ text       : chr "Simon used to maintain the BCPL compiler at Warwick University and helped many people start their careers in Co"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-05 09:46:27"
#>  $ title      : chr "Former BCPL compiler maintainer Simon Langley has died"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[15]]
#> List of 9
#>  $ by         : chr "milkoolong"
#>  $ descendants: int 10
#>  $ id         : int 43587758
#>  $ kids       : int [1:9] 43594964 43589847 43596187 43591706 43595050 43594035 43594250 43590556 43589113
#>  $ score      : int 6
#>  $ text       : chr "I heard reading well-written code can improve my skills. Instead of reading top-rated projects on GitHub, which"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-04 21:27:41"
#>  $ title      : chr "Ask HN: Which Opens Source Software have the sexiest code?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[16]]
#> List of 9
#>  $ by         : chr "OrbitalShotput_"
#>  $ descendants: int 19
#>  $ id         : int 43564550
#>  $ kids       : int [1:9] 43564583 43565515 43566758 43597207 43566094 43568498 43565857 43576142 43572997
#>  $ score      : int 18
#>  $ text       : chr "This is coming up a lot these days in the news- but Customs and Border Patrol have increased the amount of sear"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 03:53:26"
#>  $ title      : chr "Ask HN: For the privacy minded, How do you prepare for gadget Border Searches?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[17]]
#> List of 9
#>  $ by         : chr "r_singh"
#>  $ descendants: int 13
#>  $ id         : int 43572949
#>  $ kids       : int [1:10] 43600647 43582622 43573331 43597277 43576099 43578085 43578048 43573578 43574111 43574733
#>  $ score      : int 16
#>  $ text       : chr "I feel like talking about how to think is going to be a bigger topic of discussion going forward now that we ha"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 17:35:48"
#>  $ title      : chr "Ask HN: Best books written on “How to think”?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[18]]
#> List of 9
#>  $ by         : chr "spacebanana7"
#>  $ descendants: int 187
#>  $ id         : int 43547309
#>  $ kids       : int [1:41] 43547461 43547919 43547462 43548054 43547594 43547586 43547535 43547492 43547416 43547545 ...
#>  $ score      : int 190
#>  $ text       : chr "I appreciate developing ROCm into something competitive with CUDA would require a lot of work, both internally "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 14:37:54"
#>  $ title      : chr "Ask HN: Why hasn’t AMD made a viable CUDA alternative?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[19]]
#> List of 9
#>  $ by         : chr "amano-kenji"
#>  $ descendants: int 0
#>  $ id         : int 43598567
#>  $ kids       : int 43598699
#>  $ score      : int 4
#>  $ text       : chr "# The pros of lightweight threads and channels<p>Lightweight threads don&#x27;t have the cost of OS threads. Yo"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 02:49:26"
#>  $ title      : chr "Lightweight threads and channels are better than async/await syntax"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[20]]
#> List of 9
#>  $ by         : chr "krrishd"
#>  $ descendants: int 28
#>  $ id         : int 43542213
#>  $ kids       : int [1:23] 43599158 43543288 43545293 43576842 43588654 43577022 43547072 43563006 43542498 43545930 ...
#>  $ score      : int 12
#>  $ text       : chr "Question says it all - I&#x27;m curious what the state of the art is for a community like HN (that, intuitively"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 02:26:43"
#>  $ title      : chr "Ask HN: Do you still self-host a blog? What's your publishing stack?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[21]]
#> List of 9
#>  $ by         : chr "sonderotis"
#>  $ descendants: int 4
#>  $ id         : int 43564410
#>  $ kids       : int [1:3] 43564430 43575174 43594965
#>  $ score      : int 12
#>  $ text       : chr "I have a blog but it seems most people who are with my content hardly like it.\nI had a review recently that sa"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 03:33:51"
#>  $ title      : chr "Ask HN: How to write better tech blogs"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[22]]
#> List of 9
#>  $ by         : chr "mywittyname"
#>  $ descendants: int 42
#>  $ id         : int 43539544
#>  $ kids       : int [1:23] 43541034 43540390 43541108 43540394 43600850 43549407 43541277 43542447 43542906 43582469 ...
#>  $ score      : int 62
#>  $ text       : chr "Currently, I use Perplexity or ChatGPT via the web prompt for small coding tasks, but sometimes I&#x27;ll use O"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-03-31 20:26:35"
#>  $ title      : chr "Ask HN: What's the best way to get started with LLM-assisted programing?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[23]]
#> List of 9
#>  $ by         : chr "withinboredom"
#>  $ descendants: int 1
#>  $ id         : int 43595947
#>  $ kids       : int 43595997
#>  $ score      : int 5
#>  $ text       : chr "I&#x27;m either going crazy, or GitHub removed&#x2F;moved the raw diff ability. You used to be able to get a ra"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-05 19:04:29"
#>  $ title      : chr "Ask HN: Did GitHub remove the raw diff"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[24]]
#> List of 9
#>  $ by         : chr "frabia"
#>  $ descendants: int 16
#>  $ id         : int 43544254
#>  $ kids       : int [1:18] 43545403 43594656 43544777 43597507 43549448 43549431 43561205 43579488 43576159 43550446 ...
#>  $ score      : int 21
#>  $ text       : chr "Between newsletters, podcasts, Youtube videos, (tech) news, articles, and (of course) books, how does one keep "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 08:33:50"
#>  $ title      : chr "When so much knowledge is produced every day, how do you keep up?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[25]]
#> List of 9
#>  $ by         : chr "unrealizex"
#>  $ descendants: int 2
#>  $ id         : int 43563174
#>  $ kids       : int [1:3] 43563339 43564553 43563324
#>  $ score      : int 12
#>  $ text       : chr "Hey HN&#x2F;Reddit,\nWe&#x27;re a team of AI researchers passionate about simplifying 3D modeling. We&#x27;ve b"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 23:48:43"
#>  $ title      : chr "We Built an AI Tool to Create High-Quality 3D Models from Regular Videos"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[26]]
#> List of 9
#>  $ by         : chr "mecHacker"
#>  $ descendants: int 26
#>  $ id         : int 43534185
#>  $ kids       : int [1:12] 43535625 43535820 43535809 43535456 43535851 43535623 43535673 43534999 43535640 43534977 ...
#>  $ score      : int 40
#>  $ text       : chr "Been wrestling with this for a while, figured some of you might have seen similar.\nThe Situation:\n* 15+ years"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-03-31 12:26:46"
#>  $ title      : chr "Ask HN: Finding Remote Enterprise SW Consulting Leads: Non-US Resident Edition"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[27]]
#> List of 9
#>  $ by         : chr "funerr"
#>  $ descendants: int 7
#>  $ id         : int 43598347
#>  $ kids       : int [1:8] 43602742 43603969 43599348 43598710 43599974 43598655 43598361 43598707
#>  $ score      : int 4
#>  $ text       : chr "I&#x27;ve been using Cursor for the last couple of months but I&#x27;m feeling like I might be hitting the limi"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 01:55:14"
#>  $ title      : chr "Ask HN: Best AI IDE right now?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
# }