Hacker News latest ask stories

Retrieves latest ask stories using Hacker News API

Usage

get_latest_ask_stories(max_items = NULL)

Arguments

max_items

Maximum number of items to retrieve. If max_items = NULL, returns all available

Value

list of latest ask HN stories ids

Examples

# \donttest{
# get the lastest ask story on Hacker News
latest_ask_story <- get_latest_ask_stories(max_items = 1)
latest_ask_story
#> [[1]]
#> List of 8
#>  $ by         : chr "caspercrf"
#>  $ descendants: int 0
#>  $ id         : int 43613922
#>  $ score      : int 3
#>  $ text       : chr "I was recently laid off and have been considering a career change.  I&#x27;m 46 and have been in software testi"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-07 17:34:51"
#>  $ title      : chr "Ask HN: Is anyone here an Insurance Risk Engineer?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get 10 latest ask stories on Hacker News
latest_10_ask_stories <- get_latest_ask_stories(max_items = 10)
latest_10_ask_stories
#> [[1]]
#> List of 8
#>  $ by         : chr "caspercrf"
#>  $ descendants: int 0
#>  $ id         : int 43613922
#>  $ score      : int 3
#>  $ text       : chr "I was recently laid off and have been considering a career change.  I&#x27;m 46 and have been in software testi"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-07 17:34:51"
#>  $ title      : chr "Ask HN: Is anyone here an Insurance Risk Engineer?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 9
#>  $ by         : chr "whoishiring"
#>  $ descendants: int 421
#>  $ id         : int 43547611
#>  $ kids       : int [1:288] 43614891 43612652 43608769 43611812 43610715 43548465 43564151 43566287 43606672 43563428 ...
#>  $ score      : int 272
#>  $ text       : chr "Please state the location and include REMOTE for remote work, REMOTE (US)\nor similar if the country is restric"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 15:01:11"
#>  $ title      : chr "Ask HN: Who is hiring? (April 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 9
#>  $ by         : chr "LiamPowell"
#>  $ descendants: int 4
#>  $ id         : int 43614260
#>  $ kids       : int [1:5] 43614271 43614421 43614315 43614558 43614321
#>  $ score      : int 5
#>  $ text       : chr "Below is a link to a simple Gemini prompt where the LLM has been told it has access to a &quot;launch_missiles&"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-07 18:09:52"
#>  $ title      : chr "Could you convince a LLM to launch a nuclear strike?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 9
#>  $ by         : chr "whoishiring"
#>  $ descendants: int 276
#>  $ id         : int 43547609
#>  $ kids       : int [1:287] 43614843 43614425 43614500 43613870 43613943 43612305 43612670 43612673 43612173 43609947 ...
#>  $ score      : int 77
#>  $ text       : chr "Share your information if you are looking for work. Please use this format:<p><pre><code>  Location:\n  Remote:"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 15:01:11"
#>  $ title      : chr "Ask HN: Who wants to be hired? (April 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 8
#>  $ by         : chr "Haeuserschlucht"
#>  $ descendants: int 0
#>  $ id         : int 43611088
#>  $ score      : int 2
#>  $ text       : chr "All I can find are those parroted, meaningless definitions of koc, not when to use them, how much to invest and"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-07 13:21:55"
#>  $ title      : chr "Ask HN: How and when do I use knock-out certificates?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 9
#>  $ by         : chr "dang"
#>  $ descendants: int 463
#>  $ id         : int 43558671
#>  $ kids       : int [1:124] 43560481 43559535 43559609 43561038 43565390 43560360 43559266 43561913 43564571 43559244 ...
#>  $ score      : int 1810
#>  $ text       : chr "Hi all,<p>Tom Howard is going public as HN moderator today. He has been doing HN moderation work for years alre"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 16:49:56"
#>  $ title      : chr "Tell HN: Announcing tomhow as a public moderator"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 9
#>  $ by         : chr "nextn"
#>  $ descendants: int 8
#>  $ id         : int 43609719
#>  $ kids       : int [1:9] 43610019 43611441 43612171 43610410 43611355 43609862 43610054 43611556 43610972
#>  $ score      : int 4
#>  $ text       : chr "Then the user would get paid for getting interrupted.<p>The user could configure the browser&#x2F;email-client&"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-07 10:20:19"
#>  $ title      : chr "What if we made advertising pay the user?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "muddi900"
#>  $ descendants: int 14
#>  $ id         : int 43603574
#>  $ kids       : int [1:12] 43610955 43605144 43605680 43605685 43605032 43611751 43604562 43605228 43609590 43606245 ...
#>  $ score      : int 18
#>  $ text       : chr "The issue I have had with llms since the start has been the slop. Contrary to popular belief, the open web was "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 18:17:51"
#>  $ title      : chr "Ask HN: How to Use \"Deep Research\"?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 9
#>  $ by         : chr "herberh2"
#>  $ descendants: int 3
#>  $ id         : int 43604299
#>  $ kids       : int [1:3] 43608459 43604625 43609394
#>  $ score      : int 8
#>  $ text       : chr "What are the best AI tools out there that can help take a product idea and refine it to get it to an MVP?<p>Not"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 19:41:48"
#>  $ title      : chr "What are the best AI tools to turn an idea into MVP?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 9
#>  $ by         : chr "bavarianbob"
#>  $ descendants: int 36
#>  $ id         : int 43548589
#>  $ kids       : int [1:12] 43548817 43548695 43550246 43548950 43553438 43550191 43550012 43550236 43549804 43565416 ...
#>  $ score      : int 120
#>  $ text       : chr "EDIT: Back online?!<p>NPM discussion: <a href=\"https:&#x2F;&#x2F;github.com&#x2F;npm&#x2F;cli&#x2F;issues&#x2F"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 16:19:28"
#>  $ title      : chr "Tell HN: Camelgate NPM Outage (Cloudflare)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get all latest ask stories on Hacker News
latest_ask_stories <- get_latest_ask_stories()
latest_ask_stories
#> [[1]]
#> List of 8
#>  $ by         : chr "caspercrf"
#>  $ descendants: int 0
#>  $ id         : int 43613922
#>  $ score      : int 3
#>  $ text       : chr "I was recently laid off and have been considering a career change.  I&#x27;m 46 and have been in software testi"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-07 17:34:51"
#>  $ title      : chr "Ask HN: Is anyone here an Insurance Risk Engineer?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 9
#>  $ by         : chr "whoishiring"
#>  $ descendants: int 421
#>  $ id         : int 43547611
#>  $ kids       : int [1:288] 43614891 43612652 43608769 43611812 43610715 43548465 43564151 43566287 43606672 43563428 ...
#>  $ score      : int 272
#>  $ text       : chr "Please state the location and include REMOTE for remote work, REMOTE (US)\nor similar if the country is restric"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 15:01:11"
#>  $ title      : chr "Ask HN: Who is hiring? (April 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 9
#>  $ by         : chr "LiamPowell"
#>  $ descendants: int 4
#>  $ id         : int 43614260
#>  $ kids       : int [1:5] 43614271 43614421 43614315 43614558 43614321
#>  $ score      : int 5
#>  $ text       : chr "Below is a link to a simple Gemini prompt where the LLM has been told it has access to a &quot;launch_missiles&"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-07 18:09:52"
#>  $ title      : chr "Could you convince a LLM to launch a nuclear strike?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 9
#>  $ by         : chr "whoishiring"
#>  $ descendants: int 276
#>  $ id         : int 43547609
#>  $ kids       : int [1:287] 43614843 43614425 43614500 43613870 43613943 43612305 43612670 43612673 43612173 43609947 ...
#>  $ score      : int 77
#>  $ text       : chr "Share your information if you are looking for work. Please use this format:<p><pre><code>  Location:\n  Remote:"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 15:01:11"
#>  $ title      : chr "Ask HN: Who wants to be hired? (April 2025)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 8
#>  $ by         : chr "Haeuserschlucht"
#>  $ descendants: int 0
#>  $ id         : int 43611088
#>  $ score      : int 2
#>  $ text       : chr "All I can find are those parroted, meaningless definitions of koc, not when to use them, how much to invest and"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-07 13:21:55"
#>  $ title      : chr "Ask HN: How and when do I use knock-out certificates?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 9
#>  $ by         : chr "dang"
#>  $ descendants: int 463
#>  $ id         : int 43558671
#>  $ kids       : int [1:124] 43560481 43559535 43559609 43561038 43565390 43560360 43559266 43561913 43564571 43559244 ...
#>  $ score      : int 1810
#>  $ text       : chr "Hi all,<p>Tom Howard is going public as HN moderator today. He has been doing HN moderation work for years alre"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 16:49:56"
#>  $ title      : chr "Tell HN: Announcing tomhow as a public moderator"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 9
#>  $ by         : chr "nextn"
#>  $ descendants: int 8
#>  $ id         : int 43609719
#>  $ kids       : int [1:9] 43610019 43611441 43612171 43610410 43611355 43609862 43610054 43611556 43610972
#>  $ score      : int 4
#>  $ text       : chr "Then the user would get paid for getting interrupted.<p>The user could configure the browser&#x2F;email-client&"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-07 10:20:19"
#>  $ title      : chr "What if we made advertising pay the user?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "muddi900"
#>  $ descendants: int 14
#>  $ id         : int 43603574
#>  $ kids       : int [1:12] 43610955 43605144 43605680 43605685 43605032 43611751 43604562 43605228 43609590 43606245 ...
#>  $ score      : int 18
#>  $ text       : chr "The issue I have had with llms since the start has been the slop. Contrary to popular belief, the open web was "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 18:17:51"
#>  $ title      : chr "Ask HN: How to Use \"Deep Research\"?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 9
#>  $ by         : chr "herberh2"
#>  $ descendants: int 3
#>  $ id         : int 43604299
#>  $ kids       : int [1:3] 43608459 43604625 43609394
#>  $ score      : int 8
#>  $ text       : chr "What are the best AI tools out there that can help take a product idea and refine it to get it to an MVP?<p>Not"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 19:41:48"
#>  $ title      : chr "What are the best AI tools to turn an idea into MVP?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 9
#>  $ by         : chr "bavarianbob"
#>  $ descendants: int 36
#>  $ id         : int 43548589
#>  $ kids       : int [1:12] 43548817 43548695 43550246 43548950 43553438 43550191 43550012 43550236 43549804 43565416 ...
#>  $ score      : int 120
#>  $ text       : chr "EDIT: Back online?!<p>NPM discussion: <a href=\"https:&#x2F;&#x2F;github.com&#x2F;npm&#x2F;cli&#x2F;issues&#x2F"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 16:19:28"
#>  $ title      : chr "Tell HN: Camelgate NPM Outage (Cloudflare)"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[11]]
#> List of 9
#>  $ by         : chr "MITthrow123"
#>  $ descendants: int 86
#>  $ id         : int 43612448
#>  $ kids       : int [1:51] 43615135 43613413 43612547 43613606 43612947 43613389 43613487 43614795 43614967 43614917 ...
#>  $ score      : int 86
#>  $ text       : chr "I&#x27;m a senior at MIT studying Course 6 (EECS), and I&#x27;m graduating soon with no job lined up. I&#x27;ve"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-07 15:12:59"
#>  $ title      : chr "Ask HN: I'm an MIT senior and still unemployed – and so are most of my friends"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[12]]
#> List of 9
#>  $ by         : chr "colesantiago"
#>  $ descendants: int 2
#>  $ id         : int 43604753
#>  $ kids       : int [1:2] 43606907 43605152
#>  $ score      : int 5
#>  $ text       : chr "There was a time where there was frothy hype around LK-99, the superconductor that &#x27;worked&#x27; at room t"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 20:35:28"
#>  $ title      : chr "Ask HN: What happened after LK-99?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[13]]
#> List of 9
#>  $ by         : chr "Rooster61"
#>  $ descendants: int 3
#>  $ id         : int 43611739
#>  $ kids       : int [1:3] 43613269 43612168 43611744
#>  $ score      : int 6
#>  $ text       : chr "After doing a bit of research and trying to get my head around MCP, I am eager to take advantage of the idea. H"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-07 14:11:13"
#>  $ title      : chr "Ask HN: Is MCP mature enough to use safely?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[14]]
#> List of 9
#>  $ by         : chr "OulaX"
#>  $ descendants: int 8
#>  $ id         : int 43591673
#>  $ kids       : int [1:6] 43611565 43591987 43595140 43595418 43592364 43591994
#>  $ score      : int 14
#>  $ text       : chr "I have a degree in Computer Science and currently work as a frontend web developer.<p>I live in a developing co"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-05 07:38:13"
#>  $ title      : chr "Ask HN: Looking to Break into Cybersecurity – Where Do I Start?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[15]]
#> List of 9
#>  $ by         : chr "pkkm"
#>  $ descendants: int 3
#>  $ id         : int 43601341
#>  $ kids       : int [1:3] 43613117 43606887 43610119
#>  $ score      : int 9
#>  $ text       : chr "There&#x27;s a lot of options for concurrency. In Python alone, you can use threads with shared memory, threads"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-06 13:37:14"
#>  $ title      : chr "Ask HN: Easiest and hardest concurrency models to use correctly?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[16]]
#> List of 9
#>  $ by         : chr "Brysonbw"
#>  $ descendants: int 68
#>  $ id         : int 43559733
#>  $ kids       : int [1:46] 43560914 43563603 43611378 43560923 43560393 43560318 43560388 43569894 43561195 43560492 ...
#>  $ score      : int 84
#>  $ text       : chr "How would one go about being a &#x27;rogue&#x27; OSS contributor so to speak? Live off of donations, bounties, "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 18:24:51"
#>  $ title      : chr "Ask HN: How do you make a living contributing to and/or creating OSS projects?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[17]]
#> List of 8
#>  $ by         : chr "N3Xxus_6"
#>  $ descendants: int 0
#>  $ id         : int 43591349
#>  $ score      : int 9
#>  $ text       : chr "I’m experimenting with a novel approach that integrates symbolic logic directly into a transformer’s attention "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-05 06:27:49"
#>  $ title      : chr "Novel Logic-Enhanced LLM for Improved Symbolic Reasoning"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[18]]
#> List of 9
#>  $ by         : chr "hsiYwko"
#>  $ descendants: int 4
#>  $ id         : int 43596438
#>  $ kids       : int [1:4] 43600699 43599954 43597727 43599081
#>  $ score      : int 5
#>  $ text       : chr "Hi HN!<p>This past weekend, I built a tool called GitSummarize, it&#x27;s an open-source ai-powered documentati"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-05 20:09:39"
#>  $ title      : chr "Ask HN: As an eng leader, how useful is AI-generated internal documentation?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[19]]
#> List of 9
#>  $ by         : chr "franczesko"
#>  $ descendants: int 14
#>  $ id         : int 43593627
#>  $ kids       : int [1:15] 43607177 43594001 43594367 43594038 43599256 43593990 43594530 43595384 43594050 43595870 ...
#>  $ score      : int 13
#>  $ text       : chr "HN account for majority of my open tabs. With a very dynamic top submissions list, I struggle to keep up with a"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-05 14:02:42"
#>  $ title      : chr "Ask HN: How do you consume HN?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[20]]
#> List of 9
#>  $ by         : chr "OrbitalShotput_"
#>  $ descendants: int 20
#>  $ id         : int 43564550
#>  $ kids       : int [1:9] 43564583 43565515 43566758 43566094 43568498 43565857 43597207 43576142 43572997
#>  $ score      : int 20
#>  $ text       : chr "This is coming up a lot these days in the news- but Customs and Border Patrol have increased the amount of sear"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 03:53:26"
#>  $ title      : chr "Ask HN: For the privacy minded, How do you prepare for gadget Border Searches?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[21]]
#> List of 9
#>  $ by         : chr "wiley1454"
#>  $ descendants: int 5
#>  $ id         : int 43605775
#>  $ kids       : int [1:3] 43605801 43609017 43605804
#>  $ score      : int 4
#>  $ text       : chr "Do you think OpenAI paid off Studio Ghibli, or If the Studio is plotting a massive lawsuit, as they&#x27;ve been quiet."
#>  $ time       : POSIXct[1:1], format: "2025-04-06 23:15:11"
#>  $ title      : chr "Ask HN: Is Studio Ghibli plotting a lawsuit"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[22]]
#> List of 8
#>  $ by         : chr "kevthecoder"
#>  $ descendants: int 0
#>  $ id         : int 43592201
#>  $ score      : int 15
#>  $ text       : chr "Simon used to maintain the BCPL compiler at Warwick University and helped many people start their careers in Co"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-05 09:46:27"
#>  $ title      : chr "Former BCPL compiler maintainer Simon Langley has died"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[23]]
#> List of 9
#>  $ by         : chr "r_singh"
#>  $ descendants: int 14
#>  $ id         : int 43572949
#>  $ kids       : int [1:11] 43582622 43573331 43605512 43600647 43576099 43578085 43578048 43573578 43597277 43574111 ...
#>  $ score      : int 17
#>  $ text       : chr "I feel like talking about how to think is going to be a bigger topic of discussion going forward now that we ha"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 17:35:48"
#>  $ title      : chr "Ask HN: Best books written on “How to think”?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[24]]
#> List of 9
#>  $ by         : chr "spacebanana7"
#>  $ descendants: int 187
#>  $ id         : int 43547309
#>  $ kids       : int [1:41] 43547461 43547919 43547462 43548054 43547594 43547586 43547535 43547492 43547416 43547545 ...
#>  $ score      : int 193
#>  $ text       : chr "I appreciate developing ROCm into something competitive with CUDA would require a lot of work, both internally "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 14:37:54"
#>  $ title      : chr "Ask HN: Why hasn’t AMD made a viable CUDA alternative?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[25]]
#> List of 9
#>  $ by         : chr "jimsojim"
#>  $ descendants: int 43
#>  $ id         : int 43611474
#>  $ kids       : int [1:35] 43613992 43614857 43613261 43611681 43612541 43612025 43612543 43612522 43612794 43613486 ...
#>  $ score      : int 28
#>  $ text       : chr "Curious to know which books you&#x27;ve read that genuinely felt worth the time and attention you gave them. Ca"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-07 13:51:30"
#>  $ title      : chr "Ask HN: What books have been worth your time?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[26]]
#> List of 9
#>  $ by         : chr "sonderotis"
#>  $ descendants: int 6
#>  $ id         : int 43564410
#>  $ kids       : int [1:4] 43614964 43594965 43564430 43575174
#>  $ score      : int 13
#>  $ text       : chr "I have a blog but it seems most people who are with my content hardly like it.\nI had a review recently that sa"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-03 03:33:51"
#>  $ title      : chr "Ask HN: How to write better tech blogs"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[27]]
#> List of 9
#>  $ by         : chr "frabia"
#>  $ descendants: int 17
#>  $ id         : int 43544254
#>  $ kids       : int [1:19] 43610124 43545403 43544777 43594656 43549448 43549431 43597507 43561205 43550446 43579488 ...
#>  $ score      : int 21
#>  $ text       : chr "Between newsletters, podcasts, Youtube videos, (tech) news, articles, and (of course) books, how does one keep "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-01 08:33:50"
#>  $ title      : chr "When so much knowledge is produced every day, how do you keep up?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[28]]
#> List of 9
#>  $ by         : chr "unrealizex"
#>  $ descendants: int 2
#>  $ id         : int 43563174
#>  $ kids       : int [1:3] 43563339 43564553 43563324
#>  $ score      : int 13
#>  $ text       : chr "Hey HN&#x2F;Reddit,\nWe&#x27;re a team of AI researchers passionate about simplifying 3D modeling. We&#x27;ve b"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2025-04-02 23:48:43"
#>  $ title      : chr "We Built an AI Tool to Create High-Quality 3D Models from Regular Videos"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
# }