Hacker News latest ask stories

Retrieves latest ask stories using Hacker News API

get_latest_ask_stories(max_items = NULL)

Arguments

max_items
Maximum number of items to retrieve. If max_items = NULL, returns all available

Value

list of latest ask HN stories ids

Details

Parallel api requests can be enabled by running future::plan(future::multiprocess)

Examples

# get the lastest ask story on Hacker News
latest_ask_story <- get_latest_ask_stories(max_items = 1)
latest_ask_story

#> [[1]]
#> List of 9
#>  $ by         : chr "aosaigh"
#>  $ descendants: int 5
#>  $ id         : int 21669408
#>  $ kids       : int [1:4] 21670253 21670316 21669578 21670317
#>  $ score      : int 12
#>  $ text       : chr "I&#x27;m the only developer on a small web application that is targetting enterprise customers.<p>After meeting"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 11:33:25"
#>  $ title      : chr "Ask HN: Where to start with enterprise-level security and procedures?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 

# \donttest{
# get 10 latest ask stories on Hacker News
latest_10_ask_stories <- get_latest_ask_stories(max_items = 10)
latest_10_ask_stories

#> [[1]]
#> List of 9
#>  $ by         : chr "aosaigh"
#>  $ descendants: int 5
#>  $ id         : int 21669408
#>  $ kids       : int [1:4] 21670253 21670316 21669578 21670317
#>  $ score      : int 12
#>  $ text       : chr "I&#x27;m the only developer on a small web application that is targetting enterprise customers.<p>After meeting"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 11:33:25"
#>  $ title      : chr "Ask HN: Where to start with enterprise-level security and procedures?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 9
#>  $ by         : chr "ufarooqi"
#>  $ descendants: int 1
#>  $ id         : int 21669808
#>  $ kids       : int 21670082
#>  $ score      : int 2
#>  $ text       : chr "I was working on a product idea for last couple of weeks and thinking about ways to launch a MVP. After some re"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 14:12:03"
#>  $ title      : chr "Ask HN: Buffer/Hootsuite for a Product Launch?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 8
#>  $ by         : chr "slientscope21"
#>  $ descendants: int 0
#>  $ id         : int 21669771
#>  $ score      : int 2
#>  $ text       : chr "For anyone who started off programming quick fixes or on-top of systems. How did you learn the development, pac"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 13:59:48"
#>  $ title      : chr "Ask HN: Going from Programming to Full Developer"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 9
#>  $ by         : chr "NiloCK"
#>  $ descendants: int 230
#>  $ id         : int 21660754
#>  $ kids       : int [1:138] 21662896 21670256 21661257 21660857 21664213 21661905 21661148 21661256 21661406 21664614 ...
#>  $ score      : int 182
#>  $ text       : chr "I&#x27;m an educator with a CS &#x2F; programming background. There&#x27;s a possibility that I&#x27;ll be movi"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 00:04:47"
#>  $ title      : chr "Ask HN: What should be taught in high school?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 8
#>  $ by         : chr "federiconitidi"
#>  $ descendants: int 0
#>  $ id         : int 21669386
#>  $ score      : int 3
#>  $ text       : chr "I’m exploring with a friend the idea of creating a bioinfo services company focused on ad-hoc analysis of rna a"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 11:24:46"
#>  $ title      : chr "Ask HN: rna-seq/dna-seq/computational genomics service company"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 9
#>  $ by         : chr "cKQjwLJ6sn20"
#>  $ descendants: int 8
#>  $ id         : int 21665338
#>  $ kids       : int [1:8] 21668042 21667997 21668663 21668616 21668543 21668546 21668210 21668004
#>  $ score      : int 23
#>  $ text       : chr "I am a disabled software engineer at a mid stage startup who is planning to get an experimental surgery with th"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 17:47:38"
#>  $ title      : chr "Ask HN: Should I find new job before telling current job about surgery?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 7
#>  $ by         : chr "zefy"
#>  $ descendants: int 0
#>  $ id         : int 21669998
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-30 15:05:45"
#>  $ title      : chr "Ask HN: Is Anyone Hiring GCP Engineers?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "japhyr"
#>  $ descendants: int 280
#>  $ id         : int 21660118
#>  $ kids       : int [1:146] 21661331 21663210 21661382 21661440 21664325 21661430 21663180 21662581 21661976 21661691 ...
#>  $ score      : int 404
#>  $ text       : chr "I think this gets asked most years, and I always enjoy the threads that come out of it. I am thankful for a fam"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-28 21:55:06"
#>  $ title      : chr "Ask HN: What are you thankful for?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 9
#>  $ by         : chr "brianaluv"
#>  $ descendants: int 28
#>  $ id         : int 21667309
#>  $ kids       : int [1:10] 21668715 21668408 21668665 21668459 21668643 21668234 21668757 21667498 21668472 21667851
#>  $ score      : int 22
#>  $ text       : chr "I&#x27;ve been working(as a developer) on a software development project for quite a long time now.\nDuring thi"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 00:01:25"
#>  $ title      : chr "Ask HN: Dealing with newcomers' arrogance?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 9
#>  $ by         : chr "c89X"
#>  $ descendants: int 39
#>  $ id         : int 21656891
#>  $ kids       : int [1:21] 21660200 21661024 21657151 21660545 21660434 21658963 21660687 21665875 21660990 21660911 ...
#>  $ score      : int 105
#>  $ text       : chr "We run a small SaaS where users are able to create accounts, submit billing information and upload&#x2F;call ML"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-28 13:48:40"
#>  $ title      : chr "Ask HN: How do you deal with atomicity in microservice environments?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get all latest ask stories on Hacker News
latest_ask_stories <- get_latest_ask_stories()
latest_ask_stories

#> [[1]]
#> List of 9
#>  $ by         : chr "aosaigh"
#>  $ descendants: int 5
#>  $ id         : int 21669408
#>  $ kids       : int [1:4] 21670253 21670316 21669578 21670317
#>  $ score      : int 12
#>  $ text       : chr "I&#x27;m the only developer on a small web application that is targetting enterprise customers.<p>After meeting"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 11:33:25"
#>  $ title      : chr "Ask HN: Where to start with enterprise-level security and procedures?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 9
#>  $ by         : chr "ufarooqi"
#>  $ descendants: int 1
#>  $ id         : int 21669808
#>  $ kids       : int 21670082
#>  $ score      : int 2
#>  $ text       : chr "I was working on a product idea for last couple of weeks and thinking about ways to launch a MVP. After some re"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 14:12:03"
#>  $ title      : chr "Ask HN: Buffer/Hootsuite for a Product Launch?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 8
#>  $ by         : chr "slientscope21"
#>  $ descendants: int 0
#>  $ id         : int 21669771
#>  $ score      : int 2
#>  $ text       : chr "For anyone who started off programming quick fixes or on-top of systems. How did you learn the development, pac"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 13:59:48"
#>  $ title      : chr "Ask HN: Going from Programming to Full Developer"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 9
#>  $ by         : chr "NiloCK"
#>  $ descendants: int 230
#>  $ id         : int 21660754
#>  $ kids       : int [1:138] 21662896 21670256 21661257 21660857 21664213 21661905 21661148 21661256 21661406 21664614 ...
#>  $ score      : int 182
#>  $ text       : chr "I&#x27;m an educator with a CS &#x2F; programming background. There&#x27;s a possibility that I&#x27;ll be movi"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 00:04:47"
#>  $ title      : chr "Ask HN: What should be taught in high school?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 8
#>  $ by         : chr "federiconitidi"
#>  $ descendants: int 0
#>  $ id         : int 21669386
#>  $ score      : int 3
#>  $ text       : chr "I’m exploring with a friend the idea of creating a bioinfo services company focused on ad-hoc analysis of rna a"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 11:24:46"
#>  $ title      : chr "Ask HN: rna-seq/dna-seq/computational genomics service company"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 9
#>  $ by         : chr "cKQjwLJ6sn20"
#>  $ descendants: int 8
#>  $ id         : int 21665338
#>  $ kids       : int [1:8] 21668042 21667997 21668663 21668616 21668543 21668546 21668210 21668004
#>  $ score      : int 23
#>  $ text       : chr "I am a disabled software engineer at a mid stage startup who is planning to get an experimental surgery with th"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 17:47:38"
#>  $ title      : chr "Ask HN: Should I find new job before telling current job about surgery?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 7
#>  $ by         : chr "zefy"
#>  $ descendants: int 0
#>  $ id         : int 21669998
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-30 15:05:45"
#>  $ title      : chr "Ask HN: Is Anyone Hiring GCP Engineers?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "japhyr"
#>  $ descendants: int 280
#>  $ id         : int 21660118
#>  $ kids       : int [1:146] 21661331 21663210 21661382 21661440 21664325 21661430 21663180 21662581 21661976 21661691 ...
#>  $ score      : int 404
#>  $ text       : chr "I think this gets asked most years, and I always enjoy the threads that come out of it. I am thankful for a fam"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-28 21:55:06"
#>  $ title      : chr "Ask HN: What are you thankful for?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 9
#>  $ by         : chr "brianaluv"
#>  $ descendants: int 28
#>  $ id         : int 21667309
#>  $ kids       : int [1:10] 21668715 21668408 21668665 21668459 21668643 21668234 21668757 21667498 21668472 21667851
#>  $ score      : int 22
#>  $ text       : chr "I&#x27;ve been working(as a developer) on a software development project for quite a long time now.\nDuring thi"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 00:01:25"
#>  $ title      : chr "Ask HN: Dealing with newcomers' arrogance?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 9
#>  $ by         : chr "c89X"
#>  $ descendants: int 39
#>  $ id         : int 21656891
#>  $ kids       : int [1:21] 21660200 21661024 21657151 21660545 21660434 21658963 21660687 21665875 21660990 21660911 ...
#>  $ score      : int 105
#>  $ text       : chr "We run a small SaaS where users are able to create accounts, submit billing information and upload&#x2F;call ML"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-28 13:48:40"
#>  $ title      : chr "Ask HN: How do you deal with atomicity in microservice environments?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[11]]
#> List of 9
#>  $ by         : chr "burning_out_101"
#>  $ descendants: int 147
#>  $ id         : int 21661054
#>  $ kids       : int [1:90] 21663125 21662706 21663018 21662981 21662847 21662519 21661729 21664540 21663260 21663887 ...
#>  $ score      : int 263
#>  $ text       : chr "Burning out as an early stage start-up employee. I’ve been employed there a few years, worked miracles, but get"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 01:23:38"
#>  $ title      : chr "Ask HN: Burning Out"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[12]]
#> List of 9
#>  $ by         : chr "ufarooqi"
#>  $ descendants: int 2
#>  $ id         : int 21667131
#>  $ kids       : int 21667152
#>  $ score      : int 10
#>  $ text       : chr "I have seen founders using product hunt, hacker news, reddit and twitter etc. Are there other platforms like this?"
#>  $ time       : POSIXct[1:1], format: "2019-11-29 23:10:47"
#>  $ title      : chr "Ask HN: Which platforms do founders mostly use to launch products?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[13]]
#> List of 9
#>  $ by         : chr "abahlo"
#>  $ descendants: int 11
#>  $ id         : int 21665285
#>  $ kids       : int [1:8] 21666455 21667101 21666144 21667887 21665735 21665546 21667803 21665523
#>  $ score      : int 17
#>  $ text       : chr "Using a single page application like React has become the defacto standard for most web applications.\nI realiz"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 17:36:30"
#>  $ title      : chr "Ask HN: Why use an SPA?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[14]]
#> List of 9
#>  $ by         : chr "trango"
#>  $ descendants: int 4
#>  $ id         : int 21663038
#>  $ kids       : int [1:4] 21666023 21663655 21667660 21663112
#>  $ score      : int 13
#>  $ text       : chr "2019 saw more than double the data breaches in 2018. Are data breaches increasing that rapidly or are we just g"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 09:51:52"
#>  $ title      : chr "Data breaches up 100% in 2019 over 2018"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[15]]
#> List of 9
#>  $ by         : chr "rwol"
#>  $ descendants: int 18
#>  $ id         : int 21660859
#>  $ kids       : int [1:13] 21661472 21663464 21665916 21664732 21666868 21662274 21663381 21662872 21661179 21663581 ...
#>  $ score      : int 41
#>  $ text       : chr "Doing post-thanksgiving shopping and thought the HN folk might have some cool gift ideas and deal-hunting strat"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 00:31:36"
#>  $ title      : chr "Ask HN: Any (tech) black friday deals you'd recommend?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[16]]
#> List of 9
#>  $ by         : chr "sellingwebsite"
#>  $ descendants: int 10
#>  $ id         : int 21663651
#>  $ kids       : int [1:7] 21667553 21665272 21664854 21668388 21666932 21666049 21668217
#>  $ score      : int 13
#>  $ text       : chr "The title says it all"
#>  $ time       : POSIXct[1:1], format: "2019-11-29 12:13:13"
#>  $ title      : chr "Ask HN: What is the biggest bullshit you have encountered at your workplace?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[17]]
#> List of 7
#>  $ by         : chr "pro_sine"
#>  $ descendants: int 0
#>  $ id         : int 21668336
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-11-30 04:26:11"
#>  $ title      : chr "Ask HN: Best way to prioritize happiness while focusing on the important things?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[18]]
#> List of 9
#>  $ by         : chr "throwaway1m"
#>  $ descendants: int 397
#>  $ id         : int 21645117
#>  $ kids       : int [1:107] 21645516 21646369 21645205 21645233 21645882 21646213 21645284 21645321 21647310 21645596 ...
#>  $ score      : int 656
#>  $ text       : chr "I&#x27;ve spent the last decade in two FANG companies, and I&#x27;ve got to a high level where my total compens"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-27 04:56:51"
#>  $ title      : chr "Ask HN: Quitting Big Tech, what is it like?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[19]]
#> List of 9
#>  $ by         : chr "Fazel94"
#>  $ descendants: int 3
#>  $ id         : int 21666949
#>  $ kids       : int [1:2] 21667659 21668819
#>  $ score      : int 4
#>  $ text       : chr "I was studying the Embrace, Extend and Extinguish strategy(https:&#x2F;&#x2F;en.wikipedia.org&#x2F;wiki&#x2F;Em"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 22:40:52"
#>  $ title      : chr "Ask HN: Is ActiveDirectory an Embrace, Extend, Extinguish of LDAP?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[20]]
#> List of 9
#>  $ by         : chr "hegemeister"
#>  $ descendants: int 1
#>  $ id         : int 21666920
#>  $ kids       : int 21667102
#>  $ score      : int 4
#>  $ text       : chr "How do you handle international customers, when it comes to billing and taxes?<p>How can I setup a payment syst"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 22:34:56"
#>  $ title      : chr "Ask HN: SaaS in Germany – How do you handle international customers?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[21]]
#> List of 9
#>  $ by         : chr "Coxa"
#>  $ descendants: int 1
#>  $ id         : int 21666826
#>  $ kids       : int 21666976
#>  $ score      : int 6
#>  $ text       : chr "I was wondering how people stay up to date with the latest CVEs that might affect their company&#x27;s operations."
#>  $ time       : POSIXct[1:1], format: "2019-11-29 22:13:43"
#>  $ title      : chr "Ask HN: How do you stay up to date with the latest CVEs?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[22]]
#> List of 9
#>  $ by         : chr "varrock"
#>  $ descendants: int 62
#>  $ id         : int 21659537
#>  $ kids       : int [1:39] 21660316 21659673 21659815 21659848 21659585 21659840 21659578 21659630 21659765 21659886 ...
#>  $ score      : int 82
#>  $ text       : chr "Are there any resources you recommend? Or do personal projects give you enough problems? And how do you map sol"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-28 20:09:52"
#>  $ title      : chr "Ask HN: What has made you a better problem solver in software engineering?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[23]]
#> List of 9
#>  $ by         : chr "wil421"
#>  $ descendants: int 38
#>  $ id         : int 21648193
#>  $ kids       : int [1:13] 21649882 21648785 21648252 21648752 21649678 21649770 21648980 21655983 21648875 21648306 ...
#>  $ score      : int 39
#>  $ text       : chr "I was curious how much self-employed people are paying for healthcare for single and family plans. I was commen"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-27 15:43:37"
#>  $ title      : chr "How much are your US insurance premiums for self-employed people?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[24]]
#> List of 9
#>  $ by         : chr "dmortin"
#>  $ descendants: int 7
#>  $ id         : int 21664023
#>  $ kids       : int [1:2] 21664554 21664177
#>  $ score      : int 4
#>  $ text       : chr "Editors like Emacs, VI, etc. are great, but IDEs can do much for you if they know the language more intimately "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 13:50:24"
#>  $ title      : chr "Ask HN: Why are there no powerful free refactoring libraries for text editors?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[25]]
#> List of 9
#>  $ by         : chr "PaulMontreal"
#>  $ descendants: int 20
#>  $ id         : int 21653679
#>  $ kids       : int [1:8] 21658694 21660559 21656463 21660325 21656320 21658118 21657465 21656171
#>  $ score      : int 12
#>  $ text       : chr "I&#x27;m creating a series of articles that reverse engineer your favorite websites. With a focus on fundamenta"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-28 01:36:01"
#>  $ title      : chr "Which website can I reverse engineer for you?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[26]]
#> List of 9
#>  $ by         : chr "monkeydust"
#>  $ descendants: int 2
#>  $ id         : int 21663218
#>  $ kids       : int [1:2] 21664825 21664011
#>  $ score      : int 7
#>  $ text       : chr "We are exploring how to aggregate significant amount of data from different clients to use in a machine learnin"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 10:29:34"
#>  $ title      : chr "Ask HN: De-identifying data without destroying quality"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[27]]
#> List of 8
#>  $ by         : chr "iameuropean"
#>  $ descendants: int 0
#>  $ id         : int 21662851
#>  $ score      : int 2
#>  $ text       : chr "Currently working for a multinational insurer, seeking to build a globally scalable saas platform with a lean i"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 08:57:43"
#>  $ title      : chr "Ask HN: Independent Software Studio"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[28]]
#> List of 9
#>  $ by         : chr "dhruvkar"
#>  $ descendants: int 1
#>  $ id         : int 21658715
#>  $ kids       : int 21659307
#>  $ score      : int 8
#>  $ text       : chr "After the sale of .org to Ethos Capital, where should non-profits migrate to?"
#>  $ time       : POSIXct[1:1], format: "2019-11-28 17:57:34"
#>  $ title      : chr "Ask HN: What is the next safe haven for non-profits (after .org sale)?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[29]]
#> List of 9
#>  $ by         : chr "gamesbrainiac"
#>  $ descendants: int 5
#>  $ id         : int 21658199
#>  $ kids       : int [1:4] 21665640 21661307 21660367 21658556
#>  $ score      : int 7
#>  $ text       : chr "Hi there!<p>I am thinking of transitioning from Mac OS to Windows, because I have seen good things about WSL2. "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-28 17:00:27"
#>  $ title      : chr "Ask HN: Anyone Switch from Mac OS to WSL2? How Has the Experience Been?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[30]]
#> List of 9
#>  $ by         : chr "alexcnwy"
#>  $ descendants: int 4
#>  $ id         : int 21656719
#>  $ kids       : int [1:4] 21658152 21658462 21659281 21658131
#>  $ score      : int 7
#>  $ text       : chr "I&#x27;ve been a digital nomad for over 3 years and I check out NomadList every now and then but never joined a"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-28 13:22:54"
#>  $ title      : chr "Ask HN: Any paying NomadList users here?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[31]]
#> List of 9
#>  $ by         : chr "alphagrep12345"
#>  $ descendants: int 22
#>  $ id         : int 21653723
#>  $ kids       : int [1:14] 21654406 21653937 21655224 21656715 21654027 21655322 21669326 21660034 21655067 21657240 ...
#>  $ score      : int 17
#>  $ text       : chr "Most great companies attract people because they have a great mission which is marketed well. Say you have a gr"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-28 01:45:53"
#>  $ title      : chr "Ask HN: How do you convince people to work for you as a non-exciting startup?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[32]]
#> List of 9
#>  $ by         : chr "jacquesm"
#>  $ descendants: int 1
#>  $ id         : int 21653498
#>  $ kids       : int 21658652
#>  $ score      : int 12
#>  $ text       : chr "Hi there HN, you&#x27;re my &#x27;means of last resort&#x27;, or so I hope. I&#x27;m working on a music project"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-28 01:09:39"
#>  $ title      : chr "Ask HN: Does anybody have this annotated music dataset?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[33]]
#> List of 8
#>  $ by         : chr "ent101"
#>  $ descendants: int 1
#>  $ id         : int 21661581
#>  $ kids       : int 21664602
#>  $ score      : int 9
#>  $ time       : POSIXct[1:1], format: "2019-11-29 03:37:54"
#>  $ title      : chr "Ask HN: What are some of the best startup landing pages?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[34]]
#> List of 8
#>  $ by         : chr "jason_slack"
#>  $ descendants: int 0
#>  $ id         : int 21658539
#>  $ score      : int 4
#>  $ text       : chr "I hope everyone celebrating Thanksgiving has a wonderful day.<p>I’m thankful for a lot of things this year. Thi"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-28 17:36:30"
#>  $ title      : chr "Tell HN: Happy Thanksgiving"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[35]]
#> List of 9
#>  $ by         : chr "UrbanPiper"
#>  $ descendants: int 2
#>  $ id         : int 21658477
#>  $ kids       : int [1:2] 21658494 21666376
#>  $ score      : int 5
#>  $ text       : chr "Looking for recommendations for developer related things from HN crowd here."
#>  $ time       : POSIXct[1:1], format: "2019-11-28 17:29:46"
#>  $ title      : chr "Ask HN: Black Friday/Cyber Monday Recommendations for Developers"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[36]]
#> List of 8
#>  $ by         : chr "efojs"
#>  $ descendants: int 0
#>  $ id         : int 21652715
#>  $ score      : int 7
#>  $ text       : chr "Found 875 duplicating timestamps for the past few months<p>Currently (learn to) analysing HN posts (stories to "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-27 23:16:36"
#>  $ title      : chr "For the past few months 875 times we posted simultaneously"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[37]]
#> List of 9
#>  $ by         : chr "hotz"
#>  $ descendants: int 2
#>  $ id         : int 21646208
#>  $ kids       : int [1:2] 21648959 21646770
#>  $ score      : int 7
#>  $ text       : chr "I&#x27;m not sure if it&#x27;s because I&#x27;m a biker&#x2F;motorcyclist but the new Tesla CyberTruck really b"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-27 09:26:20"
#>  $ title      : chr "Why decrease visibility in modern vehicles?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[38]]
#> List of 9
#>  $ by         : chr "afpx"
#>  $ descendants: int 24
#>  $ id         : int 21658012
#>  $ kids       : int [1:15] 21658063 21658062 21658848 21659961 21658238 21662085 21658230 21660226 21664274 21660683 ...
#>  $ score      : int 10
#>  $ text       : chr "My skills are current. I’ve been a practicing full stack developer &#x2F; data engineer for many years. I’ll ad"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-28 16:37:32"
#>  $ title      : chr "Late 40s, no job offers. Could a PhD help?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[39]]
#> List of 8
#>  $ by         : chr "eshlomo"
#>  $ descendants: int 2
#>  $ id         : int 21665678
#>  $ kids       : int 21665861
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-11-29 18:33:41"
#>  $ title      : chr "Ask HN: Git merge vs. rebase, when and how?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[40]]
#> List of 9
#>  $ by         : chr "non-entity"
#>  $ descendants: int 17
#>  $ id         : int 21647502
#>  $ kids       : int [1:14] 21667575 21660247 21660405 21648492 21659660 21652380 21648789 21647624 21647911 21649868 ...
#>  $ score      : int 11
#>  $ text       : chr "So recently I&#x27;ve been trying to learn a lot of new things, and I&#x27;m beginning to realize just how expe"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-27 14:11:10"
#>  $ title      : chr "Ask HN: What is your experience with e-readers? Are they worth it?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[41]]
#> List of 9
#>  $ by         : chr "HerrmannM"
#>  $ descendants: int 8
#>  $ id         : int 21645883
#>  $ kids       : int [1:5] 21647900 21646032 21646264 21646017 21647873
#>  $ score      : int 12
#>  $ text       : chr "When I learned C++ 10 years ago, we had a textbook basically teaching C++ through design patterns. I remember t"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-27 08:00:03"
#>  $ title      : chr "Ask HN: Are design patterns still a thing?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[42]]
#> List of 9
#>  $ by         : chr "crazygringo"
#>  $ descendants: int 4
#>  $ id         : int 21651189
#>  $ kids       : int [1:4] 21651212 21651439 21652752 21651403
#>  $ score      : int 6
#>  $ text       : chr "This morning there was a #1 ranked story about Tim Berners Lee on HN that had gathered over 110 points and mayb"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-27 20:36:02"
#>  $ title      : chr "Ask HN: How to prevent abusive flagging of posts on HN?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[43]]
#> List of 9
#>  $ by         : chr "novice1234"
#>  $ descendants: int 6
#>  $ id         : int 21649562
#>  $ kids       : int [1:5] 21659818 21649668 21651384 21654045 21654882
#>  $ score      : int 9
#>  $ text       : chr "I was going through job listings on stackoverflow. I saw a lot of job listing for remote positions.<p>I find it"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-27 17:56:57"
#>  $ title      : chr "Ask HN: Compensation for Remote Developers"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[44]]
#> List of 9
#>  $ by         : chr "juan4thewin"
#>  $ descendants: int 4
#>  $ id         : int 21648695
#>  $ kids       : int [1:4] 21649434 21661598 21655129 21652407
#>  $ score      : int 4
#>  $ text       : chr "I have 75+ tabs open on my phone with my favorite HackerNews articles.<p>I want to move this to an extremely mi"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-27 16:34:55"
#>  $ title      : chr "Ask HN: How to Showcase Favorite Hacker News URLs?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[45]]
#> List of 9
#>  $ by         : chr "cfv"
#>  $ descendants: int 2
#>  $ id         : int 21647518
#>  $ kids       : int 21648058
#>  $ score      : int 5
#>  $ text       : chr "I made this benchmark  here: https:&#x2F;&#x2F;jsperf.com&#x2F;ways-to-get-a-random-string<p>And I lack the for"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-27 14:13:36"
#>  $ title      : chr "Ask HN: Why is Array.from so slow?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[46]]
#> List of 9
#>  $ by         : chr "ASlave2Gravity"
#>  $ descendants: int 5
#>  $ id         : int 21646566
#>  $ kids       : int [1:3] 21646827 21646609 21650268
#>  $ score      : int 5
#>  $ text       : chr "I often find threads with a few interesting and promising comments, which I&#x27;m assuming will have many repl"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-27 10:51:58"
#>  $ title      : chr "Ask HN: Email updates on comment replies for non-OP?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[47]]
#> List of 9
#>  $ by         : chr "tkainrad"
#>  $ descendants: int 8
#>  $ id         : int 21646385
#>  $ kids       : int [1:4] 21646857 21647782 21647285 21648413
#>  $ score      : int 5
#>  $ text       : chr "I started to blog ~6 months ago. By now, I have some posts that people actually read, I have been on the front "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-27 10:07:20"
#>  $ title      : chr "Ask HN: How do you release new versions of your blog posts?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[48]]
#> List of 9
#>  $ by         : chr "brundolf"
#>  $ descendants: int 10
#>  $ id         : int 21665846
#>  $ kids       : int [1:4] 21666380 21666199 21666347 21665988
#>  $ score      : int 6
#>  $ text       : chr "My father is a climate denier, but one that&#x27;s willing to hear out an argument. He thinks observed climate "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 19:06:21"
#>  $ title      : chr "Ask HN: Concise sources of evidence making the case for climate change?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[49]]
#> List of 8
#>  $ by         : chr "vishnu_ks"
#>  $ descendants: int 1
#>  $ id         : int 21650655
#>  $ kids       : int 21653746
#>  $ score      : int 8
#>  $ time       : POSIXct[1:1], format: "2019-11-27 19:42:12"
#>  $ title      : chr "Ask HN: What are some profitable businesses that came out of Show HN?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[50]]
#> List of 8
#>  $ by         : chr "stbullard"
#>  $ descendants: int 7
#>  $ id         : int 21660078
#>  $ kids       : int [1:6] 21662416 21660943 21667168 21665675 21660939 21661096
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-11-28 21:47:41"
#>  $ title      : chr "Ask HN: What technologies are you thankful for?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[51]]
#> List of 8
#>  $ by         : chr "lackita"
#>  $ descendants: int 11
#>  $ id         : int 21647898
#>  $ kids       : int [1:6] 21649828 21668391 21648343 21669086 21652013 21652397
#>  $ score      : int 10
#>  $ time       : POSIXct[1:1], format: "2019-11-27 15:05:57"
#>  $ title      : chr "Ask HN: Are you still using a feed reader?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[52]]
#> List of 8
#>  $ by         : chr "stirrat"
#>  $ descendants: int 4
#>  $ id         : int 21652494
#>  $ kids       : int [1:3] 21652777 21652594 21652583
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-11-27 22:50:22"
#>  $ title      : chr "Ask HN: Is social media a broken idea or simply inefficient for our needs?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[53]]
#> List of 8
#>  $ by         : chr "andrewstuart"
#>  $ descendants: int 1
#>  $ id         : int 21645542
#>  $ kids       : int 21646046
#>  $ score      : int 7
#>  $ time       : POSIXct[1:1], format: "2019-11-27 06:35:23"
#>  $ title      : chr "Ask HN: Where is the web animation community on the Internet?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[54]]
#> List of 9
#>  $ by         : chr "haysstanford"
#>  $ descendants: int 1
#>  $ id         : int 21667780
#>  $ kids       : int 21667802
#>  $ score      : int 1
#>  $ text       : chr "Anyone have any advice for finding companies that were founded in 1977 in the Bay Area?<p>Trying to do this for"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 01:53:30"
#>  $ title      : chr "How to find a list of companies by date of incorporation?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[55]]
#> List of 9
#>  $ by         : chr "throwaway40324"
#>  $ descendants: int 10
#>  $ id         : int 21667406
#>  $ kids       : int [1:4] 21667591 21667467 21667622 21668510
#>  $ score      : int 3
#>  $ text       : chr "I believe in the &#x27;magic&#x27; of Christmas and Santa, but need to explain in the most positive way to my c"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 00:27:58"
#>  $ title      : chr "Ask HN: How and at what age did you tell your children the truth about Santa?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[56]]
#> List of 8
#>  $ by         : chr "reilly3000"
#>  $ descendants: int 0
#>  $ id         : int 21666831
#>  $ score      : int 1
#>  $ text       : chr "I&#x27;m not big into Black Friday &#x2F; Cyber Monday, but I don&#x27;t want to pass up a deal on something I "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 22:14:18"
#>  $ title      : chr "Ask HN: Any legitimately compelling deals for hackers?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[57]]
#> List of 8
#>  $ by         : chr "vuxel"
#>  $ descendants: int 0
#>  $ id         : int 21666441
#>  $ score      : int 1
#>  $ text       : chr "Hi. I am redoing my Resume. I feel that while it has a lot of technical details, and I&#x27;m doing my best to "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 20:54:30"
#>  $ title      : chr "Ask HN: Resume Writing Tips?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[58]]
#> List of 9
#>  $ by         : chr "mehtest"
#>  $ descendants: int 1
#>  $ id         : int 21665908
#>  $ kids       : int 21666040
#>  $ score      : int 2
#>  $ text       : chr "I&#x27;ve tried searching for good articles that cover Topic Models and LDA, but on Google I&#x27;d only find m"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 19:18:46"
#>  $ title      : chr "Ask HN: Good articles that explain topic models and LDA?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[59]]
#> List of 9
#>  $ by         : chr "miketat"
#>  $ descendants: int 0
#>  $ id         : int 21665827
#>  $ kids       : int 21667894
#>  $ score      : int 1
#>  $ text       : chr " I have 4 years old product.  \nIt is JavaScript Grid library for Enterprise.  \nhttps:&#x2F;&#x2F;fancygrid.co"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 19:02:45"
#>  $ title      : chr "JavaScript Grid Component: Look for Investors"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[60]]
#> List of 9
#>  $ by         : chr "kylebenzle"
#>  $ descendants: int 1
#>  $ id         : int 21665820
#>  $ kids       : int 21665940
#>  $ score      : int 3
#>  $ text       : chr "For fun my side project was writing a SciFi novel. I&#x27;ve gotten a dozen or so people to read and and respon"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 19:01:36"
#>  $ title      : chr "I wrote a SciFi novel for fun which everyone that has read it loves it"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[61]]
#> List of 8
#>  $ by         : chr "meagher"
#>  $ descendants: int 0
#>  $ id         : int 21665805
#>  $ score      : int 1
#>  $ text       : chr "Based on PG’s essay http:&#x2F;&#x2F;paulgraham.com&#x2F;top.html<p>Most people only have one top idea in their"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 18:58:13"
#>  $ title      : chr "Ask HN: What’s the top idea in your mind? Is it the wrong one?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[62]]
#> List of 8
#>  $ by         : chr "unearthed"
#>  $ descendants: int 0
#>  $ id         : int 21665224
#>  $ score      : int 2
#>  $ text       : chr "Whats the easiest way that I can make some income but without starting a 9-5 salaried job.<p>Really I&#x27;m so"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 17:25:28"
#>  $ title      : chr "Ask HN: Make money programming w/o full-time job"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[63]]
#> List of 8
#>  $ by         : chr "throwawayt856"
#>  $ descendants: int 1
#>  $ id         : int 21665008
#>  $ kids       : int 21665130
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-11-29 16:48:38"
#>  $ title      : chr "Ask HN: What are your favorite tech-specific books and blogs?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[64]]
#> List of 8
#>  $ by         : chr "throwaway_cfo"
#>  $ descendants: int 0
#>  $ id         : int 21664767
#>  $ score      : int 2
#>  $ text       : chr "Hi all,<p>TL;DR: 35-yo guy seeking advice for jumping in the &quot;safest&quot; (ie with little revenue loss) p"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 16:04:24"
#>  $ title      : chr "Ask HN: Jumping from Corporate Finance to Tech"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[65]]
#> List of 8
#>  $ by         : chr "throwawayt856"
#>  $ descendants: int 1
#>  $ id         : int 21664744
#>  $ kids       : int 21664950
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-29 16:01:17"
#>  $ title      : chr "Ask HN: What are your favorite blogs/books, and why?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[66]]
#> List of 8
#>  $ by         : chr "liamg"
#>  $ descendants: int 0
#>  $ id         : int 21664684
#>  $ score      : int 4
#>  $ text       : chr "All secure variables have disappeared from build output and are not defined in the build process.<p>No status p"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 15:50:49"
#>  $ title      : chr "Travis secure variables are broken for all builds"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[67]]
#> List of 8
#>  $ by         : chr "666lumberjack"
#>  $ descendants: int 0
#>  $ id         : int 21664569
#>  $ score      : int 1
#>  $ text       : chr "In the case that you can only optimise for one of these, do you think it is more valuable to have a role where "| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 15:28:21"
#>  $ title      : chr "Ask HN: Technically Interesting vs. Interesting Product Area?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[68]]
#> List of 7
#>  $ by         : chr "mlejva"
#>  $ descendants: int 0
#>  $ id         : int 21664514
#>  $ score      : int 1
#>  $ time       : POSIXct[1:1], format: "2019-11-29 15:18:44"
#>  $ title      : chr "Ask HN: Solo founders, how do you keep the momentum going?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[69]]
#> List of 9
#>  $ by         : chr "tictoc"
#>  $ descendants: int 1
#>  $ id         : int 21664431
#>  $ kids       : int 21664779
#>  $ score      : int 2
#>  $ text       : chr "Say I am a mid to senior developer. I&#x27;ve heard people making ridiculous amounts of money. Where do I get t"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-29 15:05:33"
#>  $ title      : chr "Ask HN: What can I reasonably expect to make at a FAANG?"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 

# }