Hacker News latest job stories

Retrieves latest job stories using Hacker News API

Usage

get_latest_job_stories(max_items = NULL)

Arguments

max_items

Maximum number of items to retrieve. If max_items = NULL, returns all available

Value

list of latest job stories

Examples

# \donttest{
# get the latest job story on Hacker News
latest_job_story <- get_latest_job_stories(max_items = 1)
latest_job_story
#> [[1]]
#> List of 7
#>  $ by   : chr "annamonaco"
#>  $ id   : int 43619346
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-08 07:51:50"
#>  $ title: chr "Paradigm (YC W24) Hiring Founding Engineers in SF"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/paradigm/jobs/nFNWweP-founding-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get 10 latest job stories on Hacker News
latest_10_job_stories <- get_latest_job_stories(max_items = 10)
latest_10_job_stories
#> [[1]]
#> List of 7
#>  $ by   : chr "annamonaco"
#>  $ id   : int 43619346
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-08 07:51:50"
#>  $ title: chr "Paradigm (YC W24) Hiring Founding Engineers in SF"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/paradigm/jobs/nFNWweP-founding-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 7
#>  $ by   : chr "lakabimanil"
#>  $ id   : int 43602940
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-06 17:01:00"
#>  $ title: chr "Foundry (YC F24) Is Hiring"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/foundry/jobs/WvDDlqc-founding-fullstack-engineer-building-the-future-of-browser-agents"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 7
#>  $ by   : chr "nedwin"
#>  $ id   : int 43594871
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-05 17:00:21"
#>  $ title: chr "Great Question (YC W21) Is Hiring Applied AI Engineers"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/great-question/jobs/AtPa8pe-ai-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 7
#>  $ by   : chr "cpeterson42"
#>  $ id   : int 43585103
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-04 17:00:40"
#>  $ title: chr "Thunder Compute (YC S24) is hiring a GPU systems engineer"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/thunder-compute/jobs/fRSS8JQ-systems-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 7
#>  $ by   : chr "ankitdce"
#>  $ id   : int 43575433
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-03 21:08:31"
#>  $ title: chr "Aviator (YC S21) is hiring engineers to build AI DevEx platform"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/aviator/jobs"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 7
#>  $ by   : chr "yuhongsun"
#>  $ id   : int 43572518
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-03 17:01:01"
#>  $ title: chr "Onyx (YC W24) Is Hiring"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/onyx/jobs/CUHpbpE-founding-devrel-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 7
#>  $ by   : chr "stewfortier"
#>  $ id   : int 43568380
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-03 12:00:16"
#>  $ title: chr "Type (YC W23) is hiring a senior software engineer"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/type/jobs/m3GcN1t-senior-software-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 7
#>  $ by   : chr "Old_Thrashbarg"
#>  $ id   : int 43561524
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-02 21:01:48"
#>  $ title: chr "UpCodes (YC S17) Is Hiring SWEs and CSMs to make construction more productive"
#>  $ type : chr "job"
#>  $ url  : chr "https://up.codes/careers?utm_source=HN"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 7
#>  $ by   : chr "sidmanchkanti21"
#>  $ id   : int 43558803
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-02 17:00:43"
#>  $ title: chr "Pulse (YC S24) Is Hiring"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/pulse-3/jobs/6o4mkAj-machine-learning-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 7
#>  $ by   : chr "joshwget"
#>  $ id   : int 43551276
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-01 21:00:08"
#>  $ title: chr "Hightouch (YC S19) Is Hiring Front End Engineers"
#>  $ type : chr "job"
#>  $ url  : chr "https://job-boards.greenhouse.io/hightouch/jobs/5437380004"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get all latest job stories on Hacker News
latest_job_stories <- get_latest_job_stories()
latest_job_stories
#> [[1]]
#> List of 7
#>  $ by   : chr "annamonaco"
#>  $ id   : int 43619346
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-08 07:51:50"
#>  $ title: chr "Paradigm (YC W24) Hiring Founding Engineers in SF"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/paradigm/jobs/nFNWweP-founding-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 7
#>  $ by   : chr "lakabimanil"
#>  $ id   : int 43602940
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-06 17:01:00"
#>  $ title: chr "Foundry (YC F24) Is Hiring"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/foundry/jobs/WvDDlqc-founding-fullstack-engineer-building-the-future-of-browser-agents"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 7
#>  $ by   : chr "nedwin"
#>  $ id   : int 43594871
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-05 17:00:21"
#>  $ title: chr "Great Question (YC W21) Is Hiring Applied AI Engineers"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/great-question/jobs/AtPa8pe-ai-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 7
#>  $ by   : chr "cpeterson42"
#>  $ id   : int 43585103
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-04 17:00:40"
#>  $ title: chr "Thunder Compute (YC S24) is hiring a GPU systems engineer"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/thunder-compute/jobs/fRSS8JQ-systems-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 7
#>  $ by   : chr "ankitdce"
#>  $ id   : int 43575433
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-03 21:08:31"
#>  $ title: chr "Aviator (YC S21) is hiring engineers to build AI DevEx platform"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/aviator/jobs"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 7
#>  $ by   : chr "yuhongsun"
#>  $ id   : int 43572518
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-03 17:01:01"
#>  $ title: chr "Onyx (YC W24) Is Hiring"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/onyx/jobs/CUHpbpE-founding-devrel-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 7
#>  $ by   : chr "stewfortier"
#>  $ id   : int 43568380
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-03 12:00:16"
#>  $ title: chr "Type (YC W23) is hiring a senior software engineer"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/type/jobs/m3GcN1t-senior-software-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 7
#>  $ by   : chr "Old_Thrashbarg"
#>  $ id   : int 43561524
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-02 21:01:48"
#>  $ title: chr "UpCodes (YC S17) Is Hiring SWEs and CSMs to make construction more productive"
#>  $ type : chr "job"
#>  $ url  : chr "https://up.codes/careers?utm_source=HN"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 7
#>  $ by   : chr "sidmanchkanti21"
#>  $ id   : int 43558803
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-02 17:00:43"
#>  $ title: chr "Pulse (YC S24) Is Hiring"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/pulse-3/jobs/6o4mkAj-machine-learning-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 7
#>  $ by   : chr "joshwget"
#>  $ id   : int 43551276
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-01 21:00:08"
#>  $ title: chr "Hightouch (YC S19) Is Hiring Front End Engineers"
#>  $ type : chr "job"
#>  $ url  : chr "https://job-boards.greenhouse.io/hightouch/jobs/5437380004"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[11]]
#> List of 7
#>  $ by   : chr "kride"
#>  $ id   : int 43549076
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-01 17:01:07"
#>  $ title: chr "Meru Health (YC S18) Is Hiring"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/meru-health/jobs/ZjKcsyk-director-growth-marketing"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[12]]
#> List of 7
#>  $ by   : chr "kbyatnal"
#>  $ id   : int 43545725
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-01 12:01:40"
#>  $ title: chr "Extend (YC W23) is hiring engineers to build LLM document processing"
#>  $ type : chr "job"
#>  $ url  : chr "https://jobs.ashbyhq.com/extend/9d4d8974-bd9b-432d-84ec-8268e5a8ed37"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[13]]
#> List of 7
#>  $ by   : chr "ea016"
#>  $ id   : int 43543681
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-04-01 07:01:30"
#>  $ title: chr "Photoroom (YC S20) Is Hiring Rust, API, and Front End Engineers (Europe, Remote)"
#>  $ type : chr "job"
#>  $ url  : chr "https://jobs.ashbyhq.com/photoroom?departmentId=5a691019-9344-462b-9f4b-4efb68086e05&utm_source=yc"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[14]]
#> List of 7
#>  $ by   : chr "jasoncsullivan"
#>  $ id   : int 43537194
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-31 17:00:06"
#>  $ title: chr "Distro (YC S24) Is Hiring a Lead Solutions Engineer"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/distro/jobs/hJQCfVH-lead-solutions-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[15]]
#> List of 7
#>  $ by   : chr "youssefarizk"
#>  $ id   : int 43532009
#>  $ score: int 1
#>  $ text : chr "We are looking for a Senior Product Designer to join us on our mission to make audio content creation seamlessl"| __truncated__
#>  $ time : POSIXct[1:1], format: "2025-03-31 07:00:19"
#>  $ title: chr "Wondercraft (YC S22) Is Hiring"
#>  $ type : chr "job"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[16]]
#> List of 7
#>  $ by   : chr "richard_pepper"
#>  $ id   : int 43525636
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-30 17:01:47"
#>  $ title: chr "Spice Data (YC S19) is hiring a software engineer"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/spice-data/jobs/TijA35R-software-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[17]]
#> List of 7
#>  $ by   : chr "tk90"
#>  $ id   : int 43518695
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-29 21:00:35"
#>  $ title: chr "Spark AI (YC W24) is hiring a full-stack engineer in San Francisco"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/spark/jobs/kDeJlPK-software-engineer-full-stack"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[18]]
#> List of 7
#>  $ by   : chr "festinalente"
#>  $ id   : int 43509782
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-28 21:01:49"
#>  $ title: chr "Finley (YC W21) Is Hiring a Technical Implementations Specialist"
#>  $ type : chr "job"
#>  $ url  : chr "https://ats.rippling.com/finley-technologies/jobs"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[19]]
#> List of 7
#>  $ by   : chr "samuelklam"
#>  $ id   : int 43504299
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-28 12:02:32"
#>  $ title: chr "PermitFlow (YC W22) Is Hiring Senior/Staff Engineers in NYC"
#>  $ type : chr "job"
#>  $ url  : chr "https://jobs.ashbyhq.com/permitflow?departmentId=d33195eb-8978-4439-abc6-5a8a072de808"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[20]]
#> List of 7
#>  $ by   : chr "amrutha_"
#>  $ id   : int 43502357
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-28 07:00:01"
#>  $ title: chr "Software Eng Intern – Structured Labs (YC S23) Is Hiring"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/structured-labs/jobs/MWU8Ws3-software-engineer-intern"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[21]]
#> List of 7
#>  $ by   : chr "hsial"
#>  $ id   : int 43500179
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-28 01:00:54"
#>  $ title: chr "Helpcare AI (YC F24) Is Hiring"
#>  $ type : chr "job"
#>  $ url  : chr "https://docs.google.com/forms/d/e/1FAIpQLScpzOyP_mk3muEpbKrnW8UTZB_yP5SJwjbeT8_6A6fhdvpJCg/viewform?usp=preview"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[22]]
#> List of 7
#>  $ by   : chr "landseer"
#>  $ id   : int 43498109
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-27 21:00:29"
#>  $ title: chr "Revyl (YC F24) Is Hiring a Front End Engineer Intern"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/revyl/jobs/5rbIJLP-frontend-engineer-intern"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[23]]
#> List of 7
#>  $ by   : chr "amasad"
#>  $ id   : int 43495571
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-27 17:01:12"
#>  $ title: chr "Replit (YC W18) Is Hiring Engineers to Automate Coding"
#>  $ type : chr "job"
#>  $ url  : chr "https://replit.com/careers"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[24]]
#> List of 7
#>  $ by   : chr "nspatel99"
#>  $ id   : int 43492680
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-27 12:00:10"
#>  $ title: chr "Craniometrix (YC W22) Is Hiring"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/craniometrix/jobs/ugwcSrU-chief-of-staff"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[25]]
#> List of 7
#>  $ by   : chr "awalias"
#>  $ id   : int 43491060
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-27 07:02:22"
#>  $ title: chr "Supabase (YC S20) Is Hiring a Senior Auth Engineer (Remote)"
#>  $ type : chr "job"
#>  $ url  : chr "https://jobs.ashbyhq.com/supabase/a2ede2c7-3792-4bfd-951d-686267b2b9a8"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[26]]
#> List of 7
#>  $ by   : chr "adchurch"
#>  $ id   : int 43484283
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-26 17:00:00"
#>  $ title: chr "Weave (YC W25) is hiring a founding product engineer"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/weave-3/jobs/E0LFOgu-founding-product-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[27]]
#> List of 7
#>  $ by   : chr "davidbuniat"
#>  $ id   : int 43473478
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-25 17:00:36"
#>  $ title: chr "Activeloop (YC S18) Is Hiring Senior Python Back End and AI Search Engineers"
#>  $ type : chr "job"
#>  $ url  : chr "https://careers.activeloop.ai/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[28]]
#> List of 7
#>  $ by   : chr "clemo_ra"
#>  $ id   : int 43459100
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-24 09:56:12"
#>  $ title: chr "Langfuse (YC W23) Is Hiring in Berlin, Germany"
#>  $ type : chr "job"
#>  $ url  : chr "https://langfuse.com/careers"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[29]]
#> List of 7
#>  $ by   : chr "greenfish6"
#>  $ id   : int 43454142
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-23 17:00:09"
#>  $ title: chr "Modern Realty (YC S24) Is Hiring"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.workatastartup.com/jobs/66546"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[30]]
#> List of 7
#>  $ by   : chr "calderwoodra"
#>  $ id   : int 43452413
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-23 12:15:51"
#>  $ title: chr "Hadrius (YC W23) Is Hiring Engineers with Ambitious Career Goals"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/hadrius/jobs/CObSGR0-founding-software-engineer-tech-lead"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[31]]
#> List of 7
#>  $ by   : chr "Nicoding27"
#>  $ id   : int 43450060
#>  $ score: int 1
#>  $ time : POSIXct[1:1], format: "2025-03-23 01:00:12"
#>  $ title: chr "Domu Technology Inc. (YC S24) Is Hiring a Vibe Coder"
#>  $ type : chr "job"
#>  $ url  : chr "https://www.ycombinator.com/companies/domu-technology-inc/jobs/hwWsGdU-vibe-coder-ai-engineer"
#>  - attr(*, "class")= chr "hn_item"
#> 
# }