Hacker News latest show stories

Retrieves latest show stories using Hacker News API

get_latest_show_stories(max_items = NULL)

Arguments

max_items
Maximum number of items to retrieve. If max_items = NULL, returns all available

Value

list of latest show HN stories

Details

Parallel api requests can be enabled by running future::plan(future::multiprocess)

Examples

# get the latest show story on Hacker News
latest_show_story <- get_latest_show_stories(max_items = 1)
latest_show_story

#> [[1]]
#> List of 9
#>  $ by         : chr "ggerganov"
#>  $ descendants: int 1
#>  $ id         : int 21670239
#>  $ kids       : int 21670247
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-30 16:02:14"
#>  $ title      : chr "Show HN: Dot-to-ASCII – Graphviz to ASCII Converter"
#>  $ type       : chr "story"
#>  $ url        : chr "https://dot-to-ascii.ggerganov.com"
#>  - attr(*, "class")= chr "hn_item"
#> 

# \donttest{
# get 10 latest show stories on Hacker News
latest_10_show_stories <- get_latest_show_stories(max_items = 10)
latest_10_show_stories

#> [[1]]
#> List of 9
#>  $ by         : chr "ggerganov"
#>  $ descendants: int 1
#>  $ id         : int 21670239
#>  $ kids       : int 21670247
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-30 16:02:14"
#>  $ title      : chr "Show HN: Dot-to-ASCII – Graphviz to ASCII Converter"
#>  $ type       : chr "story"
#>  $ url        : chr "https://dot-to-ascii.ggerganov.com"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 9
#>  $ by         : chr "acasajus"
#>  $ descendants: int 21
#>  $ id         : int 21666983
#>  $ kids       : int [1:8] 21668874 21669710 21668317 21668142 21668178 21668399 21667806 21669033
#>  $ score      : int 75
#>  $ time       : POSIXct[1:1], format: "2019-11-29 22:46:12"
#>  $ title      : chr "Show HN: Keycat – A self-hosted end-to-end encrypted password manager"
#>  $ type       : chr "story"
#>  $ url        : chr "https://key.cat/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 8
#>  $ by         : chr "ekoutanov"
#>  $ descendants: int 0
#>  $ id         : int 21669382
#>  $ score      : int 4
#>  $ time       : POSIXct[1:1], format: "2019-11-30 11:21:46"
#>  $ title      : chr "Show HN: Broker-less simulation and mocking for Kafka"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/obsidiandynamics/jackdaw"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 8
#>  $ by         : chr "alexellisuk"
#>  $ descendants: int 0
#>  $ id         : int 21669285
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-30 10:38:49"
#>  $ title      : chr "Show HN: Inletsctl kfwd for Kubernetes port-forwarding"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/inlets/inletsctl/pull/11"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 9
#>  $ by         : chr "sloev"
#>  $ descendants: int 8
#>  $ id         : int 21660627
#>  $ kids       : int [1:3] 21660647 21661147 21666984
#>  $ score      : int 31
#>  $ time       : POSIXct[1:1], format: "2019-11-28 23:37:17"
#>  $ title      : chr "Show HN: A live-drawn portrait of the common face"
#>  $ type       : chr "story"
#>  $ url        : chr "https://vrangen.net/fuzzy"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 9
#>  $ by         : chr "steven_pack"
#>  $ descendants: int 76
#>  $ id         : int 21654705
#>  $ kids       : int [1:32] 21656548 21658361 21656763 21657376 21656963 21655165 21657042 21667372 21656852 21667485 ...
#>  $ score      : int 105
#>  $ time       : POSIXct[1:1], format: "2019-11-28 06:00:50"
#>  $ title      : chr "Show HN: Notimeforbooks.com – Read a book, one page at a time, in your inbox"
#>  $ type       : chr "story"
#>  $ url        : chr "https://notimeforbooks.com"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 9
#>  $ by         : chr "modnine"
#>  $ descendants: int 1
#>  $ id         : int 21668770
#>  $ kids       : int 21668772
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-30 06:46:03"
#>  $ title      : chr "Show HN: Running OpenFaaS Locally with Kind"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/zpratt/local-openfaas-workbench"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "mardix"
#>  $ descendants: int 40
#>  $ id         : int 21660996
#>  $ kids       : int [1:14] 21663287 21664637 21664458 21664727 21663441 21662879 21666391 21664790 21664033 21665887 ...
#>  $ score      : int 63
#>  $ time       : POSIXct[1:1], format: "2019-11-29 01:06:21"
#>  $ title      : chr "Show HN: Assembly: A Pythonic Object-Oriented Web Framework Built on Flask"
#>  $ type       : chr "story"
#>  $ url        : chr "https://mardix.github.io/assembly/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 9
#>  $ by         : chr "johndamaia"
#>  $ descendants: int 67
#>  $ id         : int 21656864
#>  $ kids       : int [1:20] 21656872 21658414 21660050 21657952 21660794 21658581 21660830 21659179 21658640 21660953 ...
#>  $ score      : int 138
#>  $ time       : POSIXct[1:1], format: "2019-11-28 13:45:13"
#>  $ title      : chr "Show HN: Building websites from Sketch using deep learning – public launch"
#>  $ type       : chr "story"
#>  $ url        : chr "https://zecoda.com/product"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 9
#>  $ by         : chr "willbrowning"
#>  $ descendants: int 171
#>  $ id         : int 21648096
#>  $ kids       : int [1:45] 21650347 21649092 21653118 21648961 21649028 21649969 21649355 21654407 21648618 21649860 ...
#>  $ score      : int 383
#>  $ time       : POSIXct[1:1], format: "2019-11-27 15:33:20"
#>  $ title      : chr "Show HN: I made an open-source anonymous email forwarding service"
#>  $ type       : chr "story"
#>  $ url        : chr "https://anonaddy.com/"
#>  - attr(*, "class")= chr "hn_item"
#> 

# get all latest show stories on Hacker News
latest_show_stories <- get_latest_show_stories()
latest_show_stories

#> [[1]]
#> List of 9
#>  $ by         : chr "ggerganov"
#>  $ descendants: int 1
#>  $ id         : int 21670239
#>  $ kids       : int 21670247
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-30 16:02:14"
#>  $ title      : chr "Show HN: Dot-to-ASCII – Graphviz to ASCII Converter"
#>  $ type       : chr "story"
#>  $ url        : chr "https://dot-to-ascii.ggerganov.com"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[2]]
#> List of 9
#>  $ by         : chr "acasajus"
#>  $ descendants: int 21
#>  $ id         : int 21666983
#>  $ kids       : int [1:8] 21668874 21669710 21668317 21668142 21668178 21668399 21667806 21669033
#>  $ score      : int 75
#>  $ time       : POSIXct[1:1], format: "2019-11-29 22:46:12"
#>  $ title      : chr "Show HN: Keycat – A self-hosted end-to-end encrypted password manager"
#>  $ type       : chr "story"
#>  $ url        : chr "https://key.cat/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[3]]
#> List of 8
#>  $ by         : chr "ekoutanov"
#>  $ descendants: int 0
#>  $ id         : int 21669382
#>  $ score      : int 4
#>  $ time       : POSIXct[1:1], format: "2019-11-30 11:21:46"
#>  $ title      : chr "Show HN: Broker-less simulation and mocking for Kafka"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/obsidiandynamics/jackdaw"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[4]]
#> List of 8
#>  $ by         : chr "alexellisuk"
#>  $ descendants: int 0
#>  $ id         : int 21669285
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-30 10:38:49"
#>  $ title      : chr "Show HN: Inletsctl kfwd for Kubernetes port-forwarding"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/inlets/inletsctl/pull/11"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[5]]
#> List of 9
#>  $ by         : chr "sloev"
#>  $ descendants: int 8
#>  $ id         : int 21660627
#>  $ kids       : int [1:3] 21660647 21661147 21666984
#>  $ score      : int 31
#>  $ time       : POSIXct[1:1], format: "2019-11-28 23:37:17"
#>  $ title      : chr "Show HN: A live-drawn portrait of the common face"
#>  $ type       : chr "story"
#>  $ url        : chr "https://vrangen.net/fuzzy"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[6]]
#> List of 9
#>  $ by         : chr "steven_pack"
#>  $ descendants: int 76
#>  $ id         : int 21654705
#>  $ kids       : int [1:32] 21656548 21658361 21656763 21657376 21656963 21655165 21657042 21667372 21656852 21667485 ...
#>  $ score      : int 105
#>  $ time       : POSIXct[1:1], format: "2019-11-28 06:00:50"
#>  $ title      : chr "Show HN: Notimeforbooks.com – Read a book, one page at a time, in your inbox"
#>  $ type       : chr "story"
#>  $ url        : chr "https://notimeforbooks.com"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[7]]
#> List of 9
#>  $ by         : chr "modnine"
#>  $ descendants: int 1
#>  $ id         : int 21668770
#>  $ kids       : int 21668772
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-30 06:46:03"
#>  $ title      : chr "Show HN: Running OpenFaaS Locally with Kind"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/zpratt/local-openfaas-workbench"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[8]]
#> List of 9
#>  $ by         : chr "mardix"
#>  $ descendants: int 40
#>  $ id         : int 21660996
#>  $ kids       : int [1:14] 21663287 21664637 21664458 21664727 21663441 21662879 21666391 21664790 21664033 21665887 ...
#>  $ score      : int 63
#>  $ time       : POSIXct[1:1], format: "2019-11-29 01:06:21"
#>  $ title      : chr "Show HN: Assembly: A Pythonic Object-Oriented Web Framework Built on Flask"
#>  $ type       : chr "story"
#>  $ url        : chr "https://mardix.github.io/assembly/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[9]]
#> List of 9
#>  $ by         : chr "johndamaia"
#>  $ descendants: int 67
#>  $ id         : int 21656864
#>  $ kids       : int [1:20] 21656872 21658414 21660050 21657952 21660794 21658581 21660830 21659179 21658640 21660953 ...
#>  $ score      : int 138
#>  $ time       : POSIXct[1:1], format: "2019-11-28 13:45:13"
#>  $ title      : chr "Show HN: Building websites from Sketch using deep learning – public launch"
#>  $ type       : chr "story"
#>  $ url        : chr "https://zecoda.com/product"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[10]]
#> List of 9
#>  $ by         : chr "willbrowning"
#>  $ descendants: int 171
#>  $ id         : int 21648096
#>  $ kids       : int [1:45] 21650347 21649092 21653118 21648961 21649028 21649969 21649355 21654407 21648618 21649860 ...
#>  $ score      : int 383
#>  $ time       : POSIXct[1:1], format: "2019-11-27 15:33:20"
#>  $ title      : chr "Show HN: I made an open-source anonymous email forwarding service"
#>  $ type       : chr "story"
#>  $ url        : chr "https://anonaddy.com/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[11]]
#> List of 9
#>  $ by         : chr "nymphmechanic"
#>  $ descendants: int 2
#>  $ id         : int 21668152
#>  $ kids       : int [1:2] 21670309 21668902
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-11-30 03:31:41"
#>  $ title      : chr "Show HN: Nymph, a C like programming language"
#>  $ type       : chr "story"
#>  $ url        : chr "https://gitHub.com/maelswarm/nymph"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[12]]
#> List of 9
#>  $ by         : chr "vforissi"
#>  $ descendants: int 1
#>  $ id         : int 21667532
#>  $ kids       : int 21667827
#>  $ score      : int 7
#>  $ text       : chr "Made this app because I was tired of not knowing 50 words&#x2F;book when I read in English.<p>You can try it on"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-30 00:58:09"
#>  $ title      : chr "Show HN: Flashcard app to learn the vocabulary of English books/series/movies"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[13]]
#> List of 9
#>  $ by         : chr "dsteinman"
#>  $ descendants: int 35
#>  $ id         : int 21651310
#>  $ kids       : int [1:9] 21651674 21652006 21651657 21651925 21655068 21653080 21652970 21652201 21651654
#>  $ score      : int 70
#>  $ time       : POSIXct[1:1], format: "2019-11-27 20:47:20"
#>  $ title      : chr "Show HN: Borg Voice Generator"
#>  $ type       : chr "story"
#>  $ url        : chr "https://jaxcore.github.io/jaxcore-say/borg-example/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[14]]
#> List of 9
#>  $ by         : chr "joshi4"
#>  $ descendants: int 17
#>  $ id         : int 21658878
#>  $ kids       : int [1:10] 21660372 21660179 21660850 21659715 21660110 21665667 21660039 21660587 21666999 21660444
#>  $ score      : int 67
#>  $ text       : chr "Hi HN,<p>I created Flowshare (<a href=\"https:&#x2F;&#x2F;flowshare.io&#x2F;\" rel=\"nofollow\">https:&#x2F;&#x"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-28 18:23:55"
#>  $ title      : chr "Show HN: Flowshare: Create and discover how-to guides for everyday software"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[15]]
#> List of 9
#>  $ by         : chr "makaronich"
#>  $ descendants: int 2
#>  $ id         : int 21647243
#>  $ kids       : int [1:2] 21647639 21648769
#>  $ score      : int 50
#>  $ time       : POSIXct[1:1], format: "2019-11-27 13:19:34"
#>  $ title      : chr "Show HN: SQL Habit – learn SQL with Real Examples from Product and Marketing"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.sqlhabit.com/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[16]]
#> List of 9
#>  $ by         : chr "bluestreak"
#>  $ descendants: int 5
#>  $ id         : int 21666790
#>  $ kids       : int [1:4] 21666800 21669920 21669831 21666883
#>  $ score      : int 16
#>  $ time       : POSIXct[1:1], format: "2019-11-29 22:03:32"
#>  $ title      : chr "Show HN: QuestDB – fast time series database, zero-GC Java"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.questdb.io/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[17]]
#> List of 9
#>  $ by         : chr "rcarmo"
#>  $ descendants: int 3
#>  $ id         : int 21657316
#>  $ kids       : int 21657409
#>  $ score      : int 17
#>  $ time       : POSIXct[1:1], format: "2019-11-28 14:52:12"
#>  $ title      : chr "Show HN: Piku, a Micro-PaaS for ARM and x86"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/piku/piku"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[18]]
#> List of 9
#>  $ by         : chr "whitezebra"
#>  $ descendants: int 1
#>  $ id         : int 21666252
#>  $ kids       : int 21669446
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-11-29 20:28:08"
#>  $ title      : chr "Show HN: Workplace search engine for all of your internal documents"
#>  $ type       : chr "story"
#>  $ url        : chr "https://evertrove.co"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[19]]
#> List of 9
#>  $ by         : chr "nextdns"
#>  $ descendants: int 43
#>  $ id         : int 21648662
#>  $ kids       : int [1:10] 21652611 21651758 21649862 21650919 21653468 21649681 21652704 21652694 21654552 21650592
#>  $ score      : int 123
#>  $ time       : POSIXct[1:1], format: "2019-11-27 16:31:26"
#>  $ title      : chr "Show HN: TrackingTheTrackers – Is a website disguising its 3rd-party trackers?"
#>  $ type       : chr "story"
#>  $ url        : chr "https://trackingthetrackers.com/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[20]]
#> List of 8
#>  $ by         : chr "alexellisuk"
#>  $ descendants: int 0
#>  $ id         : int 21657711
#>  $ score      : int 26
#>  $ time       : POSIXct[1:1], format: "2019-11-28 15:51:41"
#>  $ title      : chr "Show HN: Openfaas-Cloud: Managed OpenFaaS for Teams"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/openfaas/openfaas-cloud"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[21]]
#> List of 8
#>  $ by         : chr "__app_dev__"
#>  $ descendants: int 0
#>  $ id         : int 21651998
#>  $ score      : int 16
#>  $ time       : POSIXct[1:1], format: "2019-11-27 21:52:26"
#>  $ title      : chr "Show HN: SVG Hello World Animation – Sun, Earth, Moon Rotating"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.dataformsjs.com/examples/hello-world/en-js.htm"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[22]]
#> List of 9
#>  $ by         : chr "lutun"
#>  $ descendants: int 17
#>  $ id         : int 21644570
#>  $ kids       : int [1:6] 21652905 21652602 21665185 21645474 21651852 21645264
#>  $ score      : int 75
#>  $ time       : POSIXct[1:1], format: "2019-11-27 02:40:14"
#>  $ title      : chr "Show HN: Gorse – A recommender system back end based on collaborative filtering"
#>  $ type       : chr "story"
#>  $ url        : chr "https://gorse.io/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[23]]
#> List of 8
#>  $ by         : chr "olegnn"
#>  $ descendants: int 0
#>  $ id         : int 21664904
#>  $ score      : int 4
#>  $ time       : POSIXct[1:1], format: "2019-11-29 16:28:07"
#>  $ title      : chr "Show HN: `join!` – Rust macros with combinators to join everything"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/olegnn/join"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[24]]
#> List of 8
#>  $ by         : chr "ImpressiveWebs"
#>  $ descendants: int 0
#>  $ id         : int 21666995
#>  $ score      : int 7
#>  $ time       : POSIXct[1:1], format: "2019-11-29 22:47:59"
#>  $ title      : chr "Show HN: I've compiled all my JavaScript/DOM tips/articles into 3 E-Books"
#>  $ type       : chr "story"
#>  $ url        : chr "https://leanpub.com/b/javascriptdombundle"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[25]]
#> List of 9
#>  $ by         : chr "alyssaxuu"
#>  $ descendants: int 6
#>  $ id         : int 21649066
#>  $ kids       : int [1:4] 21649096 21653256 21655461 21652784
#>  $ score      : int 45
#>  $ time       : POSIXct[1:1], format: "2019-11-27 17:10:30"
#>  $ title      : chr "Show HN: Flowy – A minimal JavaScript library to create flowcharts"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/alyssaxuu/flowy"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[26]]
#> List of 9
#>  $ by         : chr "janesconference"
#>  $ descendants: int 2
#>  $ id         : int 21663359
#>  $ kids       : int [1:2] 21663361 21668488
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-11-29 11:01:55"
#>  $ title      : chr "Show HN: Dovetail, Spectral Subtraction Mixing in the Browser"
#>  $ type       : chr "story"
#>  $ url        : chr "https://dsp.audio/dovetail/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[27]]
#> List of 9
#>  $ by         : chr "sukeesh"
#>  $ descendants: int 1
#>  $ id         : int 21647096
#>  $ kids       : int 21655585
#>  $ score      : int 10
#>  $ time       : POSIXct[1:1], format: "2019-11-27 12:49:01"
#>  $ title      : chr "Show HN: Kubernetes Cronjob/Job Notifier to Slack"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/sukeesh/k8s-job-notify"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[28]]
#> List of 9
#>  $ by         : chr "jph"
#>  $ descendants: int 7
#>  $ id         : int 21653441
#>  $ kids       : int [1:2] 21653673 21653461
#>  $ score      : int 11
#>  $ time       : POSIXct[1:1], format: "2019-11-28 01:01:54"
#>  $ title      : chr "Show HN: Decision record – how to improve teamwork decisions by using templates"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/joelparkerhenderson/decision_record"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[29]]
#> List of 8
#>  $ by         : chr "lukaszmtw"
#>  $ descendants: int 0
#>  $ id         : int 21658741
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-11-28 18:00:56"
#>  $ title      : chr "Show HN: Epic Black Friday – Best Black Friday Deals for Webdevs, No Bullshit"
#>  $ type       : chr "story"
#>  $ url        : chr "https://epicblackfriday.co"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[30]]
#> List of 8
#>  $ by         : chr "verumnoslib"
#>  $ descendants: int 0
#>  $ id         : int 21646910
#>  $ score      : int 8
#>  $ time       : POSIXct[1:1], format: "2019-11-27 12:02:26"
#>  $ title      : chr "Show HN: Wreeto – a fast, cluttered-free knowledge organisation toolset"
#>  $ type       : chr "story"
#>  $ url        : chr "https://wreeto.com"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[31]]
#> List of 9
#>  $ by         : chr "bramk"
#>  $ descendants: int 1
#>  $ id         : int 21647207
#>  $ kids       : int 21647219
#>  $ score      : int 10
#>  $ time       : POSIXct[1:1], format: "2019-11-27 13:12:37"
#>  $ title      : chr "Show HN: Road to Scale – 300+ content items for every stage of your startup"
#>  $ type       : chr "story"
#>  $ url        : chr "https://roadtoscale.com/showhn"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[32]]
#> List of 9
#>  $ by         : chr "yelabbassi"
#>  $ descendants: int 3
#>  $ id         : int 21652173
#>  $ kids       : int [1:2] 21670221 21653279
#>  $ score      : int 9
#>  $ time       : POSIXct[1:1], format: "2019-11-27 22:10:26"
#>  $ title      : chr "Show HN: A modern app to watch public domain movies"
#>  $ type       : chr "story"
#>  $ url        : chr "https://cinedantan.com"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[33]]
#> List of 9
#>  $ by         : chr "kehers"
#>  $ descendants: int 2
#>  $ id         : int 21659271
#>  $ kids       : int 21660499
#>  $ score      : int 4
#>  $ time       : POSIXct[1:1], format: "2019-11-28 19:24:08"
#>  $ title      : chr "Show HN: Analytics, reporting and BI for transactional emails and campaigns"
#>  $ type       : chr "story"
#>  $ url        : chr "https://mailintel.io"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[34]]
#> List of 9
#>  $ by         : chr "yaj54"
#>  $ descendants: int 1
#>  $ id         : int 21658499
#>  $ kids       : int 21658509
#>  $ score      : int 4
#>  $ time       : POSIXct[1:1], format: "2019-11-28 17:33:11"
#>  $ title      : chr "Show HN: HN Theater – The Best Talks and Videos of Hacker News"
#>  $ type       : chr "story"
#>  $ url        : chr "https://yahnd.com/theater/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[35]]
#> List of 8
#>  $ by         : chr "alexellisuk"
#>  $ descendants: int 0
#>  $ id         : int 21658365
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-11-28 17:17:54"
#>  $ title      : chr "Show HN: inlets 2.6.3 tunnels with free Heroku support"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/inlets/inlets/releases/tag/2.6.3"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[36]]
#> List of 8
#>  $ by         : chr "naikas82"
#>  $ descendants: int 0
#>  $ id         : int 21652303
#>  $ score      : int 7
#>  $ time       : POSIXct[1:1], format: "2019-11-27 22:30:24"
#>  $ title      : chr "Show HN: Dupre – Streamlit UI to Identify Duplicate and Near Duplicate Images"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/amolnaik/dupre"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[37]]
#> List of 8
#>  $ by         : chr "alexellisuk"
#>  $ descendants: int 0
#>  $ id         : int 21657487
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-28 15:15:28"
#>  $ title      : chr "Show HN: Fast HTTP (L7) and TCP (L4) Tunnels Written in Go"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/inlets/inletsctl"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[38]]
#> List of 8
#>  $ by         : chr "jakon89"
#>  $ descendants: int 0
#>  $ id         : int 21657448
#>  $ score      : int 4
#>  $ time       : POSIXct[1:1], format: "2019-11-28 15:10:25"
#>  $ title      : chr "Show HN: Go + JVMTI"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/jakon89/golang-jmvti"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[39]]
#> List of 8
#>  $ by         : chr "alexellisuk"
#>  $ descendants: int 0
#>  $ id         : int 21657397
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-11-28 15:02:32"
#>  $ title      : chr "Show HN: Microservice for Kubernetes and Helm 3 with Express.js"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/alexellis/helm3-expressjs-tutorial"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[40]]
#> List of 9
#>  $ by         : chr "nrjames"
#>  $ descendants: int 5
#>  $ id         : int 21648104
#>  $ kids       : int [1:4] 21665064 21648124 21654148 21654142
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-11-27 15:34:08"
#>  $ title      : chr "Show HN: The Jumping Jouster, a Godot game I made in my spare time"
#>  $ type       : chr "story"
#>  $ url        : chr "https://clay.itch.io/the-jumping-jouster"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[41]]
#> List of 8
#>  $ by         : chr "pheelicks"
#>  $ descendants: int 0
#>  $ id         : int 21655500
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-11-28 09:12:42"
#>  $ title      : chr "Show HN: Beautiful 3D mapping engine using WebGL and Open Data"
#>  $ type       : chr "story"
#>  $ url        : chr "http://www.procedural.eu"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[42]]
#> List of 8
#>  $ by         : chr "podcatr"
#>  $ descendants: int 0
#>  $ id         : int 21654408
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-28 04:34:58"
#>  $ title      : chr "Show HN: Podcatr – a podcast directory powered by machine learning"
#>  $ type       : chr "story"
#>  $ url        : chr "https://podcatr.com"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[43]]
#> List of 9
#>  $ by         : chr "csdvrx"
#>  $ descendants: int 6
#>  $ id         : int 21644566
#>  $ kids       : int 21644585
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-11-27 02:39:29"
#>  $ title      : chr "Show HN: Sixel-tmux, display graphics because it does not eat escape sequences"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/csdvrx/sixel-tmux"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[44]]
#> List of 9
#>  $ by         : chr "amozoss"
#>  $ descendants: int 2
#>  $ id         : int 21649950
#>  $ kids       : int [1:2] 21650542 21650187
#>  $ score      : int 8
#>  $ time       : POSIXct[1:1], format: "2019-11-27 18:33:07"
#>  $ title      : chr "Show HN: Pushback.io – Interactive push notifications with curl"
#>  $ type       : chr "story"
#>  $ url        : chr "https://pushback.io/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[45]]
#> List of 9
#>  $ by         : chr "Bjarnee"
#>  $ descendants: int 2
#>  $ id         : int 21649225
#>  $ kids       : int 21655832
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-11-27 17:25:31"
#>  $ title      : chr "Show HN: Crowdsourced List of Black Friday and Cyber Monday SaaS and Tech Deals"
#>  $ type       : chr "story"
#>  $ url        : chr "https://vumiu.com/bf2019/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[46]]
#> List of 9
#>  $ by         : chr "Ravikiran"
#>  $ descendants: int 6
#>  $ id         : int 21649011
#>  $ kids       : int [1:4] 21669142 21649030 21651348 21649644
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-11-27 17:06:26"
#>  $ title      : chr "Show HN: An app to find how much time you waste in meetings"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.timewise.how/?ref=hackernews"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[47]]
#> List of 8
#>  $ by         : chr "mohammedri"
#>  $ descendants: int 0
#>  $ id         : int 21659487
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-11-28 20:00:14"
#>  $ title      : chr "Show HN: Building Systems to Monitor Data and Model Health in Production Systems"
#>  $ type       : chr "story"
#>  $ url        : chr "https://youtu.be/WZNEsc7ynxI"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[48]]
#> List of 9
#>  $ by         : chr "minimaxir"
#>  $ descendants: int 1
#>  $ id         : int 21648688
#>  $ kids       : int 21648696
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-11-27 16:34:07"
#>  $ title      : chr "Show HN: Python package to quickly generate a word cloud of Twitter tweets"
#>  $ type       : chr "story"
#>  $ url        : chr "https://github.com/minimaxir/twcloud"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[49]]
#> List of 8
#>  $ by         : chr "soygul"
#>  $ descendants: int 0
#>  $ id         : int 21648135
#>  $ score      : int 3
#>  $ text       : chr "I&#x27;ve just published a guide on fundamental engineering principles. It discusses common expectations from a"| __truncated__
#>  $ time       : POSIXct[1:1], format: "2019-11-27 15:36:49"
#>  $ title      : chr "Show HN: Fundamental Engineering Principles Guide"
#>  $ type       : chr "story"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[50]]
#> List of 8
#>  $ by         : chr "lostintangent"
#>  $ descendants: int 0
#>  $ id         : int 21662248
#>  $ score      : int 4
#>  $ time       : POSIXct[1:1], format: "2019-11-29 06:21:31"
#>  $ title      : chr "Show HN: Live Share Spaces – A Virtual Team-Room Within VS Code"
#>  $ type       : chr "story"
#>  $ url        : chr "https://marketplace.visualstudio.com/items?itemName=vsls-contrib.spaces"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[51]]
#> List of 8
#>  $ by         : chr "akanshgulati"
#>  $ descendants: int 0
#>  $ id         : int 21645271
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-27 05:32:26"
#>  $ title      : chr "Show HN: Integration of Lighthouse CI with Static Website Generators"
#>  $ type       : chr "story"
#>  $ url        : chr "https://blog.akansh.com/integrate-lighthouse-ci-with-static-site-generators/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[52]]
#> List of 9
#>  $ by         : chr "mlsarecmg"
#>  $ descendants: int 1
#>  $ id         : int 21656356
#>  $ kids       : int [1:2] 21656359 21663572
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-11-28 12:24:42"
#>  $ title      : chr "Show HN: What switching on React concurrent mode means for apps"
#>  $ type       : chr "story"
#>  $ url        : chr "https://twitter.com/0xca0a/status/1199997552466288641"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[53]]
#> List of 8
#>  $ by         : chr "theknight"
#>  $ descendants: int 0
#>  $ id         : int 21650027
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-11-27 18:40:06"
#>  $ title      : chr "Show HN: UV Radiation Exposure in the US. Check Your County's Exposure in Wh/M²"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.apolloyard.com/data-story/us-county-level-uv-exposure-wh-m2"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[54]]
#> List of 9
#>  $ by         : chr "nandreev"
#>  $ descendants: int 1
#>  $ id         : int 21666499
#>  $ kids       : int 21666517
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-11-29 21:03:08"
#>  $ title      : chr "Show HN: United States Wage/Job Growth Projections (Source: BLS.gov)"
#>  $ type       : chr "story"
#>  $ url        : chr "https://joboutlook.netlify.com/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[55]]
#> List of 9
#>  $ by         : chr "bithavoc"
#>  $ descendants: int 4
#>  $ id         : int 21666488
#>  $ kids       : int [1:3] 21666514 21666647 21666537
#>  $ score      : int 5
#>  $ time       : POSIXct[1:1], format: "2019-11-29 21:01:15"
#>  $ title      : chr "Show HN: I made a privacy-first, minimalistic browser for Hacker News"
#>  $ type       : chr "story"
#>  $ url        : chr "https://apps.apple.com/us/app/top-now/id1486916333?ls=1"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[56]]
#> List of 8
#>  $ by         : chr "freekman"
#>  $ descendants: int 0
#>  $ id         : int 21666244
#>  $ score      : int 1
#>  $ time       : POSIXct[1:1], format: "2019-11-29 20:26:15"
#>  $ title      : chr "Show HN: Travel app without a new tech-a dead end?"
#>  $ type       : chr "story"
#>  $ url        : chr "https://freekinapp.com/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[57]]
#> List of 8
#>  $ by         : chr "jranand"
#>  $ descendants: int 0
#>  $ id         : int 21665936
#>  $ score      : int 2
#>  $ time       : POSIXct[1:1], format: "2019-11-29 19:24:08"
#>  $ title      : chr "Show HN: Design, Develop and Publish Static Web Pages to Netlify"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.gridbox.io"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[58]]
#> List of 8
#>  $ by         : chr "wsieroci"
#>  $ descendants: int 0
#>  $ id         : int 21665597
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-11-29 18:24:28"
#>  $ title      : chr "Show HN: I've built Plumm - Bring all communication channels in one place"
#>  $ type       : chr "story"
#>  $ url        : chr "https://www.plummapp.com"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[59]]
#> List of 9
#>  $ by         : chr "justicz"
#>  $ descendants: int 1
#>  $ id         : int 21665462
#>  $ kids       : int 21665465
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-11-29 18:05:07"
#>  $ title      : chr "Show HN: A site I built to donate 10k mosquito nets to charity"
#>  $ type       : chr "story"
#>  $ url        : chr "https://giveanet.org/?r=M4GWBOWLMIHY4"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[60]]
#> List of 9
#>  $ by         : chr "trkulja"
#>  $ descendants: int 1
#>  $ id         : int 21664958
#>  $ kids       : int 21664965
#>  $ score      : int 3
#>  $ time       : POSIXct[1:1], format: "2019-11-29 16:39:01"
#>  $ title      : chr "Show HN: Eclidus – A touchscreen Rubik’s cube-like game"
#>  $ type       : chr "story"
#>  $ url        : chr "http://www.eclidus.com/"
#>  - attr(*, "class")= chr "hn_item"
#> 
#> [[61]]
#> List of 9
#>  $ by         : chr "adamcharnock"
#>  $ descendants: int 3
#>  $ id         : int 21664548
#>  $ kids       : int [1:2] 21664605 21664865
#>  $ score      : int 6
#>  $ time       : POSIXct[1:1], format: "2019-11-29 15:24:50"
#>  $ title      : chr "Show HN: Lightbus – Python/Redis Message Bus for Eventing and RPCs"
#>  $ type       : chr "story"
#>  $ url        : chr "https://lightbus.org/"
#>  - attr(*, "class")= chr "hn_item"
#> 

# }