Hacker News latest show stories ids

Retrieves latest show stories ids using Hacker News API

Usage

get_latest_show_stories_ids()

Value

list of latest show Hacker News stories ids

Examples

# \donttest{
# get ids of latest show stories on Hacker News
show_stories_ids <- get_latest_show_stories_ids()
show_stories_ids
#>   [1] 43601301 43572134 43593789 43603441 43602484 43602243 43601495 43588420
#>   [9] 43583478 43572733 43590998 43596722 43569724 43555820 43573539 43597411
#>  [17] 43591060 43578190 43558517 43568771 43555409 43570207 43573299 43569961
#>  [25] 43543241 43549618 43580465 43597299 43545641 43600411 43596862 43529774
#>  [33] 43596662 43574807 43543235 43595442 43595661 43515622 43560916 43595168
#>  [41] 43547835 43569190 43593774 43593470 43593463 43570749 43551767 43592770
#>  [49] 43554769 43582881 43551915 43592298 43549983 43580157 43567701 43503833
#>  [57] 43553031 43585007 43591213 43591013 43535889 43590589 43535449 43576487
#>  [65] 43508061 43580162 43589049 43514383 43561971 43588591 43597987 43520953
#>  [73] 43597738 43587140 43526957 43532967 43585750 43517588 43557290 43580598
#>  [81] 43506068 43569488 43523251 43555239 43549072 43576986 43562678 43506951
#>  [89] 43560960 43575770 43575612 43574845 43529539 43557519 43573484 43532994
#>  [97] 43556135 43571908 43547838 43571867 43587780 43554922 43569813 43537746
#> [105] 43568958 43511742 43566776 43565699 43564240 43549273 43548137 43563587
#> [113] 43562993 43562963 43560115 43521652 43559501 43558873 43558617 43528679
#> [121] 43557761 43557186 43557024 43534240 43556358 43541058 43540427 43555762
#> [129] 43555739 43539969 43555218 43514976 43514103 43533827 43521521 43533687
#> [137] 43519802 43552254 43551845 43530283 43550772 43549881 43549186
# }