Hacker News updated profiles

Retrieves changed item ids and changed profile usernames using Hacker News API

Usage

get_updates()

Value

list of changed item ids and list of changed profile usernames

Examples

# \donttest{
# get ids/usernames of updated items/profiles
hn_updates <- get_updates()
hn_updates
#> $items
#>  [1] 43613762 43609916 43616729 43618407 43612731 43615800 43615823 43609496
#>  [9] 43612018 43618042 43612526 43618897 43618443 43618146 43618445 43614942
#> [17] 43614965 43618709 43615736 43614074 43618841 43618503 43615747 31980526
#> [25] 43618649 43609242 43606713 43577490 43617352 43606315 43555996 43616604
#> [33] 43606027 43592116 43617190 43575833 43600582 31903076
#> 
#> $profiles
#>  [1] "ffsm8"       "gwbas1c"     "autoexec"    "munchler"    "pjmlp"      
#>  [6] "chasing"     "PaulHoule"   "Aurornis"    "motorest"    "rkagerer"   
#> [11] "jaredklewis"
#> 
# }