Hacker News updated profiles

Retrieves changed item ids and changed profile usernames using Hacker News API

Usage

get_updates()

Value

list of changed item ids and list of changed profile usernames

Examples

# \donttest{
# get ids/usernames of updated items/profiles
hn_updates <- get_updates()
hn_updates
#> $items
#>  [1] 43603772 43604254 43604164 43604241 43604429 43604123 43604443 43603084
#>  [9] 43604259 43604165 43603708 43604082 43603487 43603733 43577205 43604325
#> [17] 43600959 43603648 43603871 43604444 43604187 43603003 43604445 43604214
#> [25] 43598127 43597720 43594652 43600685 43604341 43599959 43604446 43602638
#> [33] 43604402 43604147 43603928 43604423 43601855 43603686 43596430 43604447
#> [41] 43604328 43603872 43603786 43604228 43604027 43604121 43603960 43603073
#> [49] 43604448 43603622 43603771 43580930 43604449 43601558 43600409 43602688
#> [57] 43603453 43603307 43603111 43604440 43592201 43593789 43597425 43603566
#> [65] 43602777 43575884 43585742 43570676 43603091 43596864 43549893 43599637
#> [73] 43596419 43601421 43596543 43599864 43595269 43603333 43574026 43603999
#> [81] 43601213
#> 
#> $profiles
#>  [1] "FirmwareBurner" "mikevm"         "tanjtanjtanj"   "jchw"          
#>  [5] "amazingamazing" "melvynmarrero"  "pclmulqdq"      "gjsman-1000"   
#>  [9] "slavik81"       "gerdesj"        "dymk"           "tough"         
#> [13] "phyzix5761"     "hinkley"        "wildrhythms"    "fnordpiglet"   
#> [17] "shalmanese"     "marxisttemp"    "pjmlp"          "immibis"       
#> [21] "mdaniel"        "khaledh"        "Henchman21"     "stavros"       
#> [25] "marris"         "999900000999"   "JumpCrisscross" "qingcharles"   
#> [29] "01msw007"       "hayst4ck"       "Shacklz"        "rcpt"          
#> 
# }