Hacker News updated profiles

Retrieves changed item ids and changed profile usernames using Hacker News API

Usage

get_updates()

Value

list of changed item ids and list of changed profile usernames

Examples

# \donttest{
# get ids/usernames of updated items/profiles
hn_updates <- get_updates()
hn_updates
#> $items
#>  [1] 43616177 43624362 43625680 43624557 43625797 43625817 43626188 43621149
#>  [9] 43625854 43626206 43625043 43625757 43625485 43626200 43626207 43625695
#> [17] 43625475 43623687 43624991 43626077 43622720 43624493 43625701 43625394
#> [25] 43623753 43625981 43622126 43623951 43626209 43625824 43625657 43625365
#> [33] 43625905 43626203 43623792 43625280 43596485 43617186 43624766 43626210
#> [41] 43626198 43624109 43626211 43617319 43626212 43620204 43626119 43615912
#> [49] 43601624 43620141 43622703 43624888 43624220 43606027 43620859 43625474
#> [57] 43619768 43625800 43620309 43577490 43622719 43620007 43626208 43619548
#> [65] 43588337 43624671 43595269 43613305 43624293 43596563 43620452 43590569
#> 
#> $profiles
#>  [1] "statskier"      "tmpz22"         "seneca"         "contagiousflow"
#>  [5] "ilrwbwrkhv"     "temp0826"       "dartharva"      "oofbey"        
#>  [9] "davidkuennen"   "da_chicken"     "aborsy"         "hj353"         
#> [13] "marvinborner"   "jasonjmcghee"   "simplesort"     "jacobmarble"   
#> [17] "FirmwareBurner" "neuralRiot"     "jjvortex"       "hn_acker"      
#> [21] "EliasWatson"    "cjbillington"   "antics"         "kirtakat"      
#> [25] "npilk"          "cpeterso"       "wongarsu"       "thewisenerd"   
#> 
# }