Hacker News updated profiles

Retrieves changed item ids and changed profile usernames using Hacker News API

Usage

get_updates()

Value

list of changed item ids and list of changed profile usernames

Examples

# \donttest{
# get ids/usernames of updated items/profiles
hn_updates <- get_updates()
hn_updates
#> $items
#>  [1] 43587862 43587898 43587858 43587854 43587877 43587947 43587928 43587855
#>  [9] 43583732 43587948 43587903 43586300 43584963 43587901 43585351 43587653
#> [17] 43587940 43578606 43587907 43587949 43587760 43586691 43583238 43587950
#> [25] 43579421 43587702 43587768 43587951 43587816 43587896 43587919 43587196
#> [33] 43587927 43585890 43583553 43587942 43580743 43584860 43587952 43587582
#> [41] 43582073 43587870 43587776 43587749 43587158 43586111 43585190 43587642
#> [49] 43575842 43587831 43586882  9249768 43583426 43587662 43585278 43579417
#> [57] 43582086 43586271 43587953 43587649 43583190 43587611 43587479 43533715
#> [65] 43538853 43587385 43587628 43582292 43584303 43584663 43581584 43555696
#> [73] 43539595 43580449 43571851 43580103 43563580 43555220 43527452 43575884
#> [81] 43587380 43538073 43578310 43585565 43570533 43586654 43585801 43568655
#> [89] 43586727  9248328 43573755 43586554 43559605 43587946
#> 
#> $profiles
#>  [1] "ascorbic"      "sofixa"        "jaennaet"      "fortran77"    
#>  [5] "option"        "tlogan"        "tuhins"        "FartyMcFarter"
#>  [9] "mmooss"        "alabastervlog" "ingve"         "fuzzfactor"   
#> [13] "latexr"        "mystified5016" "yieldcrv"      "Brajeshwar"   
#> [17] "humanpotato"   "skydhash"      "chubot"        "heymijo"      
#> [21] "hinkley"       "klipklop"      "bryanlarsen"  
#> 
# }